% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{add_age_group}
\alias{add_age_group}
\title{Create age group categories}
\usage{
add_age_group(age.months)
}
\arguments{
\item{age.months}{column containing age in months}
}
\value{
a column containing age group categories
}
\description{
Create age group categories
}
