% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{add_prov_npafp_table}
\alias{add_prov_npafp_table}
\title{Add province name back to those without district names}
\usage{
add_prov_npafp_table(npafp.output)
}
\arguments{
\item{npafp.output}{tibble output after running f.npafp.rate.01 at the district
level}
}
\value{
tibble of output with province names for those without didtrict names
}
\description{
Add province name back to those without district names
}
