% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{add_rolling_date_info}
\alias{add_rolling_date_info}
\title{Adds rolling date information}
\usage{
add_rolling_date_info(lab_data, start_date, end_date)
}
\arguments{
\item{lab_data}{\code{tibble} Lab dataset.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}
}
\value{
\code{tibble} Lab data with rolling date period information added.
}
\description{
Adds rolling date information
}
\keyword{internal}
