% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.table.functions.R
\name{add_rolling_years}
\alias{add_rolling_years}
\title{Label rolling year periods}
\usage{
add_rolling_years(df, start_date, date_col, period = months(12, FALSE))
}
\arguments{
\item{df}{\code{tibble} A dataset containing at least one date column.}

\item{start_date}{\code{str} Start date of Year 1. All years are classified in
reference to this date.}

\item{date_col}{\code{str} The name of the date column.}

\item{period}{\code{period} A \code{\link[lubridate:period]{lubridate::period()}} object. Defaults to
\code{months(12, FALSE)}.}
}
\value{
\code{tibble} A tibble with rolling year information.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

The function labels and categorizes dates based on the rolling period specified.
The start year will always be Year 1 and the rolling period is defined by
the start date and the number of periods to account for in a given rolling year.
For example, if the start date is defined as Jan 1, 2021 and we would like to
calculate a 12-month rolling period, the end date would be Dec 31, 2021.
}
\examples{
\dontrun {
raw_data <- get_all_polio_data()
afp_data <- add_rolling_years(raw_data$afp, "2021-01-01", "dateonset")
}
}
