% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{add_zero_dose_col}
\alias{add_zero_dose_col}
\title{Add columns in the AFP data for number of doses and dose category}
\usage{
add_zero_dose_col(afp.data)
}
\arguments{
\item{afp.data}{AFP linelist (specifically, afp.all.2)}
}
\value{
tibble with AFP data with columns for number of doses and dose category
}
\description{
Add columns in the AFP data for number of doses and dose category
}
