% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.table.functions.R
\name{adjust_rolling_years}
\alias{adjust_rolling_years}
\title{Adjusts the end final year of the rolling period}
\usage{
adjust_rolling_years(data, end_date, date_col)
}
\arguments{
\item{data}{\code{tibble} Data with rolling period columns.}

\item{end_date}{\code{str} The specified end date.}

\item{date_col}{\code{str} Column used when filtering by the end date.}
}
\value{
\code{tibble} Tibble with adjusted rolling period for the final year.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

In some instances, the final 12-month rolling period year goes beyond the
stated end date of the analysis. The function adjusts the final year rolling
period so that it is only up to the end date.
}
\details{
As a consequence of the adjustment, the function will also ensure that the
final data output also only contains records that are only up to the end date.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data(attach.spatial.data = F)
afp_data <- raw_data$afp
afp_data <- add_rolling_years(afp_data, start_date = "2022-01-01")
afp_data <- adjust_rolling_years(afp_data, "2025-02-25")
}
}
