% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{build_detection_map}
\alias{build_detection_map}
\title{Build detection map}
\usage{
build_detection_map(
  m_base_region,
  m_base_prov,
  data_p,
  m_data_prov,
  new_detect,
  virus_type,
  surv_options,
  start_date,
  date_3a,
  download_date,
  emg_cols,
  country,
  labels,
  clean_maps,
  data_r,
  .owner
)
}
\arguments{
\item{m_base_region}{\code{sf} Map base for region.}

\item{m_base_prov}{\code{sf} Map base for provinces.}

\item{data_p}{\code{tibble} Data for map creation.}

\item{m_data_prov}{\code{tibble} Map data for provinces.}

\item{new_detect}{\code{logical} Whether to include new detections.}

\item{virus_type}{\code{str} or \code{list} Virus type to display.}

\item{surv_options}{\code{str} or \code{list} Surveillance options.}

\item{start_date}{\code{date} Start date.}

\item{date_3a}{\code{date} ???}

\item{download_date}{\code{date} Date global polio data was downloaded.}

\item{emg_cols}{\code{list} Emergence colors.}

\item{country}{\code{str} Country or countries of interest.}

\item{labels}{\code{logical} Whether to label provinces.}

\item{clean_maps}{\code{ggplot} Clean maps.}

\item{data_r}{\code{tibble} Reported detections.}

\item{.owner}{\code{str} Entity that produced the map.}
}
\value{
\code{ggplot} Map of recent detections.
}
\description{
This is the key function that builds the detection map. This is internally
referred to as \code{g1} within \code{\link[=generate_adhoc_map]{generate_adhoc_map()}}.
}
\keyword{internal}
