% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.parquet.R
\name{build_parquet_raw_data}
\alias{build_parquet_raw_data}
\title{Recreate raw data from local parquet folder}
\usage{
build_parquet_raw_data(
  path = "GID/PEB/SIR/Data/analytic",
  from_edav = TRUE,
  dataset = "all",
  container = get_azure_storage_connection()
)
}
\arguments{
\item{path}{\code{str} Absolute path to the parquet folder.}

\item{from_edav}{\code{bool} Build using local files or files in EDAV? Defaults to TRUE.}

\item{container}{\code{azcontainer} An instance of an Azure container to connect.
to. Pass \code{\link[=get_azure_storage_connection]{get_azure_storage_connection()}} using defaults if not accessing
using a service principal.}
}
\value{
\code{list} A list containing connections to the folders associated with
individual datasets in the original output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}.
}
\description{
Recreates an output of \code{\link[=get_all_polio_data]{get_all_polio_data()}} from a folder housing
data in parquet format.
}
\details{
For tibbles with Shapes, pass to \code{\link[=from_wkb_to_sf]{from_wkb_to_sf()}} first before creating maps.
}
\examples{
\dontrun{
# Building raw_data locally
parquet_path <- "C:/Users/ABC1/Desktop/parquet_folder"
raw_data <- build_parquet_raw_data(parquet_path)

# Build raw_data from EDAV
raw_data <- build_parquet_raw_data()
}
}
