% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.parquet.R
\name{build_parquet_raw_data_edav}
\alias{build_parquet_raw_data_edav}
\title{Build raw_data using EDAV files}
\usage{
build_parquet_raw_data_edav(path = NULL, container = NULL, ...)
}
\arguments{
\item{path}{\code{str} Path to EDAV folder containing parquet files. This must
be the absolute file path from the Blob endpoint of the container.}

\item{container}{\code{azcontainer} An instance of an Azure container to connect
to. Pass \code{\link[=get_azure_storage_connection]{get_azure_storage_connection()}} using defaults if not accessing
using a service principal.}
}
\value{
\code{list} A list containing connections to the folders associated with
individual datasets in the original output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}.
}
\description{
Build raw_data using EDAV files
}
\keyword{internal}
