% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.parquet.R
\name{build_parquet_raw_data_local}
\alias{build_parquet_raw_data_local}
\title{Build raw_data using local parquet files}
\usage{
build_parquet_raw_data_local(path = NULL)
}
\arguments{
\item{path}{\code{str} A path to the parquet directory}
}
\value{
\code{list} A list containing connections to the folders associated with
individual datasets in the original output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}.
}
\description{
Build raw_data using local parquet files
}
\keyword{internal}
