% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{cache_raw_data}
\alias{cache_raw_data}
\title{Cache the raw data}
\usage{
cache_raw_data(raw_data, analytic_folder_path, use_edav)
}
\arguments{
\item{raw_data}{\code{list} Processed list of all polio data.}

\item{analytic_folder_path}{\code{str} Path to analytic folder.}

\item{use_edav}{\code{logical} Whether to use EDAV or not.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Cache the raw data
}
\keyword{internal}
