% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpd.dal.R
\name{calc_hr_range}
\alias{calc_hr_range}
\title{Return an array of numbers summarized in human readable format}
\usage{
calc_hr_range(numbers)
}
\arguments{
\item{numbers}{\code{int} An array of numbers.}
}
\value{
A string in human readable format.
}
\description{
Given an array of numbers this returns those numbers in human readable format.
}
\keyword{internal}
