% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{check_afp_geographies}
\alias{check_afp_geographies}
\title{Check whether the AFP geography matches those of the population dataset}
\usage{
check_afp_geographies(afp_data, pop_data, spatial_scale, fix_afp = FALSE)
}
\arguments{
\item{afp_data}{\code{tibble} AFP dataset}

\item{pop_data}{\code{tibble} Population dataset}

\item{spatial_scale}{\code{str} Any of the following: \verb{"ctry", "prov", "dist}.}

\item{fix_afp}{\code{logical} Whether to update the results to show corrected GUIDs
based on the population dataset.}
}
\value{
\code{tibble} Tibble with a column used for checking accuracy.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

In rare cases, the GUIDs assigned for a case may be incorrect. For example,
it may have a GUID that is incorrect for a specific year. This function checks
each AFP record for such instances.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data(attach.spatial.data = FALSE)
check_afp <- check_afp_geographies(raw_data$afp, raw_data$ctry.pop, "ctry")
}
}
