% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{check_cache}
\alias{check_cache}
\title{Checks the cache file and see if it should load parameters from it}
\usage{
check_cache(
  param_path,
  start_date,
  end_date,
  country_name = Sys.getenv("DR_COUNTRY")
)
}
\arguments{
\item{param_path}{path to parameters.RData}

\item{start_date}{start date of the desk review}

\item{end_date}{end date of the desk review}

\item{country_name}{name of the country}
}
\value{
boolean whether to use cache or not
}
\description{
Checks the cache file and see if it should load parameters from it
}
