% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{check_data_folder}
\alias{check_data_folder}
\title{Checks for required subfolders in the data folder}
\usage{
check_data_folder(
  data_folder,
  polis_folder,
  core_ready_folder,
  use_edav,
  cache
)
}
\arguments{
\item{data_folder}{\code{str} Path to the data folder.}

\item{polis_folder}{\code{str} POLIS folder with preprocessed data.}

\item{core_ready_folder}{\code{str} Name of the core ready folder. Need to be specified if preprocessing specific regions, which have their own core ready folder.}

\item{use_edav}{\code{logical} Whether to use EDAV or not.}

\item{cache}{\code{logical} Whether to cache the preprocessed data to data/polis subfolder.}
}
\value{
\code{list} List of paths to the specific subfolders.
}
\description{
Checks for required subfolders in the data folder
}
\keyword{internal}
