% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.stool.ad.02.R
\name{check_missing_afp_var}
\alias{check_missing_afp_var}
\title{Helper function to check missing variables in the AFP linelist based on spatial scale}
\usage{
check_missing_afp_var(afp_data, spatial_scale)
}
\arguments{
\item{afp_data}{AFP dataset}

\item{spatial_scale}{"ctry", "prov", or "dist"}
}
\description{
Helper function to check missing variables in the AFP linelist based on spatial scale
}
