% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{check_missing_geo}
\alias{check_missing_geo}
\title{Check for missing geographic data from afp data}
\usage{
check_missing_geo(afp.data, spatial.scale)
}
\arguments{
\item{afp.data}{tibble containing AFP data}

\item{spatial.scale}{what geographic level to check for}
}
\value{
tibble containing records with missing geographic data
}
\description{
Check for missing geographic data from afp data
}
