% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.pop.functions.R
\name{check_missing_pop}
\alias{check_missing_pop}
\title{Checks missing population at each geographic level}
\usage{
check_missing_pop(pop.data, spatial.scale)
}
\arguments{
\item{pop.data}{population data}

\item{spatial.scale}{"ctry", "prov", or "dist"}
}
\value{
tibble with missing population data
}
\description{
Checks missing population at each geographic level
}
