% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.stool.ad.02.R
\name{check_missing_pop_var}
\alias{check_missing_pop_var}
\title{Helper function to check missing variables in the pop file based on spatial scale}
\usage{
check_missing_pop_var(pop_data, spatial_scale)
}
\arguments{
\item{pop_data}{population dataset}

\item{spatial_scale}{"ctry", "prov", or "dist"}
}
\description{
Helper function to check missing variables in the pop file based on spatial scale
}
