% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{check_missing_rows}
\alias{check_missing_rows}
\title{Check the number of cases missing age and months}
\usage{
check_missing_rows(df, .col_name, .group_by)
}
\arguments{
\item{df}{\code{tibble} Dataset to check.}

\item{.col_name}{\code{str} Name of the target column.}

\item{.group_by}{\code{str} or \code{list} A string or a list of strings to group the
check by.}
}
\value{
\code{tibble} A summary of the number of rows missing for the target
variable.
}
\description{
A general function that checks the number of \code{NA} rows for a particular
column.
}
\examples{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
missing <- check_missing_rows(raw.data$afp, "age.months", c("place.admin.0", "yronset"))
}
