% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.pop.functions.R
\name{check_pop_rollout}
\alias{check_pop_rollout}
\title{Check the population data in country data from a roll up in province and
district population data}
\usage{
check_pop_rollout(ctry.data)
}
\arguments{
\item{ctry.data}{country data to check for}
}
\value{
tibble containing differences in population data
}
\description{
Check the population data in country data from a roll up in province and
district population data
}
\keyword{internal}
