% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{check_spatial_data_for_processing}
\alias{check_spatial_data_for_processing}
\title{Create the spatial data for processing}
\usage{
check_spatial_data_for_processing(data_folder, use_edav)
}
\arguments{
\item{data_folder}{\code{str} Path to the data folder.}

\item{use_edav}{\code{logical} Use EDAV or not.}
}
\value{
\code{list} Contains spatial datasets.
}
\description{
Create the spatial data for processing
}
\keyword{internal}
