% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.stool.ad.02.R
\name{check_spatial_scale}
\alias{check_spatial_scale}
\title{Helper function to check if spatial scale of data matches spatial scale chosen}
\usage{
check_spatial_scale(admin_data, spatial_scale)
}
\arguments{
\item{admin_data}{population file dataset}

\item{spatial_scale}{"ctry", "prov", or "dist"}
}
\description{
Helper function to check if spatial scale of data matches spatial scale chosen
}
