% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data}
\alias{clean_lab_data}
\title{Clean the lab data that's been loaded}
\usage{
clean_lab_data(lab.data, ctry.data, start_date, end_date)
}
\arguments{
\item{lab.data}{raw lab data}

\item{ctry.data}{country data}
}
\value{
a tibble
}
\description{
Clean the lab data that's been loaded
}
