% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data_regional}
\alias{clean_lab_data_regional}
\title{Clean lab data from the regional offices.}
\usage{
clean_lab_data_regional(
  lab_data,
  start_date,
  end_date,
  afp_data = NULL,
  ctry_name = NULL,
  lab_locs_path = NULL,
  use_edav = TRUE
)
}
\arguments{
\item{lab_data}{\code{tibble} Lab dataset.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}

\item{afp_data}{\code{tibble} AFP linelist. Either \code{ctry.data$afp.all.2} or \code{raw.data$afp}.}

\item{ctry_name}{\code{str} or \code{list} Name or a list of countries. Defaults to \code{NULL}.}

\item{lab_locs_path}{\code{str} Location of testing lab locations. Default is \code{NULL}. Will download from EDAV, if necessary.}

\item{use_edav}{\code{logical} Whether to obtain data from EDAV. Defaults to \code{TRUE}.}
}
\value{
\code{tibble} Cleaned lab data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
Regional data have different columns compared to the WHO lab data. The cleaning
of the regional data is adapted from the lab data cleaning code from the GPSAP
indicator script.

Cleans regional lab data. This was previously used in \code{\link[=clean_lab_data]{clean_lab_data()}}, but can be used on its own.
}
\examples{
\dontrun{
lab_path <- "C:/Users/XRG9/lab_data_region.csv"
ctry.data <- init_dr("algeria", lab_data_path = lab_path)
ctry.data$lab_data <- clean_lab_data_regional(
  lab.data, "2021-01-01", "2023-12-31",
  ctry.data$afp.all.2, "algeria"
)

# Not using the desk review pipeline
raw.data <- get_all_polio_data()
ctry.data <- extract_country_data("algeria", raw.data)
ctry.data$lab_data <- read_csv(lab_path)
ctry.data$lab_data <- clean_lab_data_regional(ctry.data, "2021-01-01", "2023-12-31")
}
}
