% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data_who}
\alias{clean_lab_data_who}
\title{Clean polio lab data from WHO}
\usage{
clean_lab_data_who(
  lab_data,
  start_date,
  end_date,
  afp_data = NULL,
  ctry_name = NULL
)
}
\arguments{
\item{lab_data}{\code{tibble} Lab dataset.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}

\item{afp_data}{\code{tibble} AFP linelist. Either \code{ctry.data$afp.all.2} or \code{raw.data$afp}.}

\item{ctry_name}{\code{str} or \code{list} Name or a list of countries. Defaults to \code{NULL}.}
}
\value{
\code{tibble} Cleaned lab data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
}
\details{
this function is no longer referenced within the sirfunctions package.

Cleans the lab data from WHO. This was previously used in \code{\link[=clean_lab_data]{clean_lab_data()}}, but can be used on its own.
}
\examples{
\dontrun{
lab_path <- "C:/Users/XRG9/lab_data_who.csv"
ctry.data <- init_dr("algeria", lab_data_path = lab_path)
ctry.data$lab_data <- clean_lab_data_who(ctry.data, "2021-01-01", "2023-12-31")

# Not using the desk review pipeline
raw.data <- get_all_polio_data()
ctry.data <- extract_country_data("algeria", raw.data)
ctry.data$lab_data <- read_csv(lab_path)
ctry.data$lab_data <- clean_lab_data_who(ctry.data, "2021-01-01", "2023-12-31")
}
}
