% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_local_polio_data_cache.R
\name{clear_local_polio_data_cache}
\alias{clear_local_polio_data_cache}
\title{Clear local polio data cache}
\usage{
clear_local_polio_data_cache()
}
\value{
\code{NULL}, invisibly.
}
\description{
The local polio cache is created when \code{get_all_polio_data()} is called with
\code{local_caching = TRUE}. The function clears the cache by deleting all the files within it.
}
\examples{
\dontrun{
clear_local_polio_data_cache()
}
}
