% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{col_to_datecol}
\alias{col_to_datecol}
\title{Convert character date columns to date columns in the AFP linelist}
\usage{
col_to_datecol(afp.data)
}
\arguments{
\item{afp.data}{tibble: AFP data}
}
\value{
AFP linelist with converted date columns and additional timeliness cols
}
\description{
Convert character date columns to date columns in the AFP linelist
}
\keyword{internal}
