% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{compress_png}
\alias{compress_png}
\title{Compress .png files using pngquant}
\usage{
compress_png(img, pngquant_path = NULL, suffix = "")
}
\arguments{
\item{img}{file path to the png file}

\item{pngquant_path}{file path to pngquant.exe}

\item{suffix}{add a suffix to the compressed image}
}
\description{
Compress .png files using pngquant
}
