% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{copy_dr_functions}
\alias{copy_dr_functions}
\title{Get functions used for the desk review from Github}
\usage{
copy_dr_functions(branch = "main", output_folder = Sys.getenv("DR_FUNC_PATH"))
}
\arguments{
\item{branch}{which branch to use}

\item{output_folder}{where the function scripts should be stored}
}
\description{
Get functions used for the desk review from Github
}
\keyword{internal}
