% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{copy_dr_template_code}
\alias{copy_dr_template_code}
\title{Downloads the desk review template code}
\usage{
copy_dr_template_code(output_path = Sys.getenv("DR_PATH"))
}
\arguments{
\item{output_path}{where to download the desk review template code}
}
\description{
Downloads the desk review template code
}
\keyword{internal}
