% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desk.review.functions.R
\name{copy_new_data}
\alias{copy_new_data}
\title{Pull new data and write to the specified file path}
\usage{
copy_new_data(data_size, country_name, dr_data_path)
}
\arguments{
\item{data_size}{\code{str} size of polio data to pull.
"small" (>= 2019), "medium" (>= 2016), large" (full)}

\item{country_name}{\code{str} name of the country to pull data from}

\item{dr_data_path}{\code{str} path to save the data set to`}
}
\value{
\code{list} large list containing polio data
}
\description{
Pull new data and write to the specified file path
}
