% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_afp_export}
\alias{create_afp_export}
\title{Export the AFP linelist}
\usage{
create_afp_export(
  stool.data,
  country = Sys.getenv("DR_COUNTRY"),
  excel_output_path = Sys.getenv("DR_TABLE_PATH")
)
}
\arguments{
\item{stool.data}{\code{tibble} AFP data with final adequacy columns.
This is the output of \code{\link[=generate_stool_data]{generate_stool_data()}}.}

\item{country}{\code{str} Name of the country.}

\item{excel_output_path}{\code{str} Output path of the Excel file.}
}
\value{
None.
}
\description{
Export the AFP linelist with \code{adequacy.final2} column. The \code{adequacy.final2}
column describes the status of a stool sample, such as if a
stool sample is adequate or inadequate. Specifically, it is
created from \code{\link[=generate_stool_data]{generate_stool_data()}} which takes parameters on how to deal with missing or
inadequate stool samples.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
stool.data <- generate_stool_data(
  ctry.data$afp.all.2, "good", "inadequate",
  "2021-01-01", "2023-12-31"
)
create_afp_export(stool.data)
}

}
