% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cg_expansion_map.R
\name{create_cg_super_regions}
\alias{create_cg_super_regions}
\title{Function to create super regions from subsets of consequential geographies}
\usage{
create_cg_super_regions(cg, ctry, prov, dist)
}
\arguments{
\item{cg}{\code{tibble} table containing data about existing CGs, the dataset
is expected to contain the following headers: \code{type}, \code{label}, \code{ctry}, \code{prov},
\code{dist}, \code{adm_level}. You can download an example dataset using:
sirfunctions_io("read", file_loc = "Data/misc/consequential_geographies.rds")}

\item{ctry}{\code{sp} country level spatial objects}

\item{prov}{\code{sp} province level spatial objects}

\item{dist}{\code{sp} district level spatial objects}
}
\value{
\code{sp} single spatial objects with all consequential geography
super regions
}
\description{
Function to create super regions from subsets of consequential geographies
}
\examples{
\dontrun{
create_cg_super_regions(cg = sirfunctions_io("read", file_loc = "Data/misc/consequential_geographies.rds"),
ctry = global.ctry, prov = global.prov, dist = global.dist)
}
}
