% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_emergence_group_gif.R
\name{create_emergence_group_gif}
\alias{create_emergence_group_gif}
\title{Generate Emergence Group Movement Gifs}
\usage{
create_emergence_group_gif(
  emergence_group,
  pos,
  dist,
  ctry,
  include_env = T,
  cumulative = T,
  out_gif
)
}
\arguments{
\item{emergence_group}{str: designation of the emergence group to review}

\item{pos}{tibble: positives data set}

\item{dist}{sf: shapefile of all districts}

\item{ctry}{sf: shapefile of all countries}

\item{include_env}{boolean: To include environmental detections in analysis}

\item{cumulative}{boolean: To display cases as cumulative}

\item{out_gif}{str: location where gif should be saved}
}
\value{
GIF written out to a location
}
\description{
Generate the figures and stitch together a GIF to evaluate emergence group
movement over time, generally aggregated as cumulative per month
}
