% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{create_metadata}
\alias{create_metadata}
\title{Creates a meta data file}
\usage{
create_metadata(path)
}
\arguments{
\item{path}{location where the .txt file is located/should be created}
}
\value{
this function does not return anything
}
\description{
Creates a meta data file
}
