% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_npafp_export}
\alias{create_npafp_export}
\title{Export NPAFP data}
\usage{
create_npafp_export(
  ctry.case.ind,
  prov.case.ind,
  dis.case.ind,
  excel_output_path = Sys.getenv("DR_TABLE_PATH")
)
}
\arguments{
\item{ctry.case.ind}{country NPAFP indicator}

\item{prov.case.ind}{province NPAFP indicator}

\item{dis.case.ind}{district NPAFP indicator}

\item{excel_output_path}{output path}
}
\value{
does not return anything
}
\description{
Export NPAFP data
}
