% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{create_polis_data_folder}
\alias{create_polis_data_folder}
\title{Creates polis folder to the data folder}
\usage{
create_polis_data_folder(data_folder, polis_folder, use_edav)
}
\arguments{
\item{data_folder}{\code{str} Path to the data folder.}

\item{polis_folder}{\code{str} Path to the core ready folder}

\item{use_edav}{\code{bool} Are file paths on EDAV?}
}
\description{
Copies the updated POLIS data into the data folder. This function is used
inside get_all_polio_data().
}
\keyword{internal}
