% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_pop_check_export}
\alias{create_pop_check_export}
\title{File for checking population roll-ups}
\usage{
create_pop_check_export(
  ctry.data,
  ctry.data.error.log,
  country,
  excel_output_path
)
}
\arguments{
\item{ctry.data}{RDS file containing polio data for a country}

\item{ctry.data.error.log}{error log of a country}

\item{country}{name of the country}

\item{excel_output_path}{output path}
}
\value{
does not return anything
}
\description{
File for checking population roll-ups
}
