% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_pot_comp_clust_export}
\alias{create_pot_comp_clust_export}
\title{Import potentially compatible and compatible summaries}
\usage{
create_pot_comp_clust_export(pot.c.clust, country, excel_output_path)
}
\arguments{
\item{pot.c.clust}{potentially compatible cluster summary}

\item{country}{name of the country}

\item{excel_output_path}{output path}
}
\value{
does not return anything
}
\description{
Import potentially compatible and compatible summaries
}
