% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.parquet.R
\name{create_raw_data_parquet}
\alias{create_raw_data_parquet}
\title{Convert raw data into a parquet hierarchal folder}
\usage{
create_raw_data_parquet(raw_data, path)
}
\arguments{
\item{raw_data}{\code{list} A \code{raw_data} object.}

\item{path}{\code{str} Path to export the parquet folder to.}
}
\value{
None.
}
\description{
The function takes a \code{raw_data} object (output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}) and
saves it into a parquet directory
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
create_raw_data_parquet(raw_data, "C:/Users/ABC1/Desktop/raw_data_parquet")
}
}
