% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{create_raw_data_tags}
\alias{create_raw_data_tags}
\title{Create timestamps for raw data and spatial data}
\usage{
create_raw_data_tags(data_folders_paths, use_edav)
}
\arguments{
\item{data_folders_paths}{\code{list} Output of \code{\link[=check_data_folder]{check_data_folder()}}.}

\item{use_edav}{\code{logical} Whether to use EDAV or not.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Create timestamps for raw data and spatial data
}
\keyword{internal}
