% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_stool_adequacy_export}
\alias{create_stool_adequacy_export}
\title{Export stool adequacy data}
\usage{
create_stool_adequacy_export(
  cstool,
  pstool,
  dstool,
  excel_output_path = Sys.getenv("DR_TABLE_PATH")
)
}
\arguments{
\item{cstool}{\code{tibble} Stool adequacy at country level.}

\item{pstool}{\code{tibble} Stool adequacy at province level.}

\item{dstool}{\code{tibble} Stool adequacy at district level.}

\item{excel_output_path}{\code{str} Output path.}
}
\value{
None.
}
\description{
The function combines the stool adequacy summary tables from \code{\link[=f.stool.ad.01]{f.stool.ad.01()}}
and exports to an Excel file, with each geographic level on its own tab.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
cstool <- f.stool.ad.01(
  ctry.data$afp.all.2, ctry.data$ctry.pop,
  "2021-01-01", "2023-01-01", "ctry"
)
pstool <- f.stool.ad.01(
  ctry.data$afp.all.2, ctry.data$prov.pop,
  "2021-01-01", "2023-01-01", "prov"
)
dstool <- f.stool.ad.01(
  ctry.data$afp.all.2, ctry.data$dist.pop,
  "2021-01-01", "2023-01-01", "dist"
)
create_stool_adequacy_export(cstool, pstool, dstool)
}

}
