% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.table.functions.R
\name{export_kpi_table}
\alias{export_kpi_table}
\title{Export KPI tables}
\usage{
export_kpi_table(
  c1 = NULL,
  c2 = NULL,
  c3 = NULL,
  c4 = NULL,
  output_path = Sys.getenv("KPI_TABLES"),
  drop_label_cols = TRUE
)
}
\arguments{
\item{c1}{\code{tibble} Output of \code{\link[=generate_c1_table]{generate_c1_table()}}. Defaults to \code{NULL}.}

\item{c2}{\code{tibble} Output of \code{\link[=generate_c2_table]{generate_c2_table()}}. Defaults to \code{NULL}.}

\item{c3}{\code{tibble} Output of \code{\link[=generate_c3_table]{generate_c3_table()}}. Defaults to \code{NULL}.}

\item{c4}{\code{tibble} Output of \code{\link[=generate_c4_table]{generate_c4_table()}}. Defaults to \code{NULL}.}

\item{output_path}{\code{str} Path to output the table to. Defaults to the path
initiated after running \code{\link[=init_kpi]{init_kpi()}}.}

\item{drop_label_cols}{\code{bool} Keep or discard label columns. Defaults to
\code{TRUE}.}
}
\value{
None.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Performs formatting and export of the C1-C4 KPI tables.
}
\examples{
\dontrun{
init_kpi()
c1 <- generate_c1_table(raw_data, "2021-01-01", "2023-12-31")
c2 <- generate_c2_table(raw_data$afp, raw_data$ctry.pop, "2021-01-01", "2023-12-31")
c3 <- generate_c3_table(raw_data$es, "2021-01-01", "2023-12-31")
c4 <- generate_c4_table(lab_data, raw_data$afp, "2021-01-01", "2024-12-31")
}
}
