% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_country_data.R
\name{extract_country_data_single}
\alias{extract_country_data_single}
\title{Extract country specific information from raw polio data}
\usage{
extract_country_data_single(.country, .raw.data = raw.data)
}
\arguments{
\item{.country}{\code{str} A string or a vector of strings containing country name(s) of interest. Case insensitive.}

\item{.raw.data}{\code{list} Output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}.}
}
\value{
Named \code{list} with country specific datasets.
}
\description{
Filters country specific data from the CDC generated \code{raw.data} object from \code{\link[=get_all_polio_data]{get_all_polio_data()}}.
}
\keyword{internal}
