% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.ev.rate.01.R
\name{f.ev.rate.01}
\alias{f.ev.rate.01}
\title{Calculate EV detection rate function}
\usage{
f.ev.rate.01(es.data, start.date, end.date)
}
\arguments{
\item{es.data}{tibble: ES data which includes site name (site.name),
country (ADM0_NAME),
date of collection (collect.date), and a binary ev detection variable (ev.detect)
that indicates absence/presence (0, 1) of enterovius in an ES sample}

\item{start.date}{chr: "YYYY-MM-DD"}

\item{end.date}{chr: "YYYY-MM-DD"}
}
\value{
tibble
}
\description{
Function to calcualte the EV detection rate in sites from POLIS
}
