% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.expand.bbox.R
\name{f.expand.bbox}
\alias{f.expand.bbox}
\title{Expand bbox
Sourced from https://rdrr.io/github/Chrisjb/basemapR/src/R/expand_bbox.R
A function to take a bounding box (generated using st_bbox) and expand it by x meters in the X direction and y meters in the Y direction.}
\usage{
f.expand.bbox(bbox, X, Y, X2 = X, Y2 = Y, crs_out = 4326)
}
\arguments{
\item{bbox}{a bounding box generated by sf::st_bbox() or named numeric vector with xmin, ymin, xmax, ymax.}

\item{X}{the distance in metres that we want to expand the bounding box by in the X direction}

\item{Y}{the distance in metres that we want to expand the bounding box by in the Y direction}

\item{X2}{if specified, the meters in the Easterly direction and metresX becomes metres in the Westerly direction.}

\item{Y2}{if specified, the metres to the South. metresY becomes metres to the North.}

\item{crs_out}{what epsg coordinate system to return the bbox in. Defaults to lat/lng (4326).}
}
\value{
a named numeric vector of bbox extents. This can be converted into an sf object using sf::st_as_sfc(). Coordinates are in lat/lng (4326).
}
\description{
A function to take a bounding box (generated using st_bbox) and expand it by x meters in the X direction and y meters in the Y direction.
}
