% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.npafp.rate.01.R
\name{f.npafp.rate.01}
\alias{f.npafp.rate.01}
\title{Calculate non-polio AFP rate}
\usage{
f.npafp.rate.01(
  afp.data,
  pop.data,
  start.date,
  end.date,
  spatial.scale,
  pending = T,
  rolling = F,
  sp_continuity_validation = T
)
}
\arguments{
\item{afp.data}{\code{tibble} AFP data which includes GUID at a given spatial scale
formatted as \verb{adm(0,1,2)guid}, onset date as \code{date} and \code{cdc.classification.all2} which includes
\verb{"NPAFP", "PENDING", "LAB PENDING"}.}

\item{pop.data}{\code{tibble} Under 15 population data by a given spatial scale including
\code{year}, \verb{adm(0,1,2)guid}, \code{u15pop}, and \code{ctry/prov/dist} as appropriate.}

\item{start.date}{\code{chr} Start date with the format "YYYY-MM-DD".}

\item{end.date}{\code{chr} Start date with the format "YYYY-MM-DD".}

\item{spatial.scale}{\code{chr} Spatial scale for analysis.
\itemize{
\item \code{"prov"} Province level
\item \code{"dist"} District level
\item \code{"ctry"} Country level
}}

\item{pending}{\code{bool} Should cases classified as \code{PENDING} or \verb{LAB PENDING} be included in calculations? Default \code{TRUE}.}

\item{rolling}{\code{bool} Should the analysis be performed on a rolling bases? Default \code{FALSE}.}

\item{sp_continuity_validation}{Should we filter places that are not present
for the entirety of the analysis dates? Default \code{TRUE}.}
}
\value{
\code{tibble} A table containing NPAFP rates as well as additional information relevant to each location analyzed.
}
\description{
Calculate the NPAFP rate from POLIS data. Can either pass \code{raw.data} to calculate NPAFP rates
on the global dataset, or a \code{ctry.data} dataset.
}
