% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_stool_ad_function.R
\name{f.stool.ad.01}
\alias{f.stool.ad.01}
\title{Calculate Stool Adequacy Function}
\usage{
f.stool.ad.01(ctry.data, start.date, end.date, prov.dist, missing, bad.data)
}
\arguments{
\item{ctry.data}{list: country specific data object}

\item{start.date}{chr:}

\item{end.date}{chr:}

\item{prov.dist}{chr: "prov" or "dist" or "ctry"}

\item{missing}{chr: "good" or "bad" or "missing"}

\item{bad.data}{chr: "remove" or "inadequate" or "adequate"}
}
\value{
tibble
}
\description{
creates an adequacy variable - 'missing' parameter defines how
missing data is treated - "Good" classifies missing data as good quality (POLIS method)
"Bad" classifies all missing as bad quality, "missing" excludes missing from
the calculations
}
