% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.timely.01.R
\name{f.timely.01}
\alias{f.timely.01}
\title{Calculate timeliness of stool collection using only epi variables}
\usage{
f.timely.01(
  afp.data,
  admin.data,
  start.date,
  end.date,
  spatial.scale,
  intervals.manual = F
)
}
\arguments{
\item{afp.data}{tibble: AFP data which includes GUID at a given spatial scale
formated as "adm{0,1,2}guid, onset date as "date"}

\item{admin.data}{tibble: Full list of country administrative units by a given
spatial scale including "year", "adm{0,1,2}guid, and "{ctry/prov/dist}"
as appropriate}

\item{start.date}{date: "YYYY-MM-DD"}

\item{end.date}{date: "YYYY-MM-DD"}

\item{spatial.scale}{chr: "prov" or "dist" or "ctry"}

\item{intervals.manual}{boolean: Should user input their own timeliness interval
requirements? Default is false}
}
\value{
tibble
}
\description{
creates a table for timeliness by geographic unit including the
number of timely stool samples for each interval and percent timely
}
