% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.timely.detection.01.R
\name{f.timely.detection.01}
\alias{f.timely.detection.01}
\title{Function to calculate timeliness of detection}
\usage{
f.timely.detection.01(
  afp.data,
  es.data,
  ctryseq.data,
  start.date,
  end.date,
  rolling = T
)
}
\arguments{
\item{afp.data}{tibble: AFP data which includes classification of AFP cases with onset date and date of notification to HQ,}

\item{es.data}{tibble: ES data which includes classification of samples with collection date and date of notification to HQ,}

\item{ctryseq.data}{tibble: countries with sequencing capacity within or outside of the country,
country (ADM0_NAME),
classification of AFP cases & ES samples
onset date of AFP cases and collection date of ES samples
date of notification to HQ (date.notification.to.hq),}

\item{start.date}{chr: "YYYY-MM-DD"}

\item{end.date}{chr: "YYYY-MM-DD"}

\item{rolling}{boolean: default FALSE}
}
\value{
list
}
\description{
Function to calculate the overall timeliness of detection in AFP & ES POLIS data
}
\examples{
afp.es.detect.ind <- f.timly.detection.01(
  afp.data = data$afp.data,
  es.data = data$es.data,
  ctryseq.data = data$ctryseq.data,
  start.date = "2019-06-01",
  end.date = "2022-01-31",
  rolling = T
)

overall.detection.01 <- afp.es.detect.ind[[1]]
overall.detection.02 <- afp.es.detect.ind[[2]]
}
