% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kppi.functions.R
\name{f.year.roll}
\alias{f.year.roll}
\title{Calculates the year a week number falls into.}
\usage{
f.year.roll(wk)
}
\arguments{
\item{wk}{\code{numeric} Week number.}
}
\value{
\code{str} The year the week falls into.
}
\description{
Calculates the year a week number falls into.
}
\examples{
f.year.roll(32)
}
\keyword{internal}
