% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{f.yrs.01}
\alias{f.yrs.01}
\title{Utility function to fix years}
\usage{
f.yrs.01(df, yrs)
}
\arguments{
\item{df}{\code{tibble} Dataframe to be used for fixing years.}

\item{yrs}{\code{array} Numeric array of years to fix data.}
}
\value{
\code{tibble} Long format table to fix years.
}
\description{
Utility function to fix years
}
