% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desk.review.functions.R
\name{fetch_dr_data}
\alias{fetch_dr_data}
\title{Fetch the Rds file to be used for a desk review}
\usage{
fetch_dr_data(country, year, local_dr_repo)
}
\arguments{
\item{country}{country name as a string}

\item{year}{year as an integer}

\item{local_dr_repo}{the local desk review repository}
}
\value{
A status message
}
\description{
Fetch the Rds file to be used for a desk review
}
