% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{fix_ctry_data_missing_guids}
\alias{fix_ctry_data_missing_guids}
\title{Function to fix unknown GUIDs in the AFP linelist by obtaining GUIDs found in the pop files}
\usage{
fix_ctry_data_missing_guids(afp.data, pop.data, guid_list, spatial_scale)
}
\arguments{
\item{afp.data}{AFP linelist (afp.all.2)}

\item{pop.data}{population file (prov.pop or dist.pop)}

\item{guid_list}{list of unknown GUIDs from the AFP linelist.
This is the output of check_afp_guid_ctry_data().}

\item{spatial_scale}{"prov" or "dist"}
}
\value{
AFP data with corrected GUIDs based on the population files.
}
\description{
Function to fix unknown GUIDs in the AFP linelist by obtaining GUIDs found in the pop files
}
