% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cg_expansion_map.R
\name{flag_cg_positives}
\alias{flag_cg_positives}
\title{Flag positives data from identified consequential geographies}
\usage{
flag_cg_positives(cg_super_regions, pos, start_year = 2016)
}
\arguments{
\item{cg_super_regions}{\code{sf} Spatial object of all CG super regions and a
flag for all their specific GUIDs. This is the output of \code{\link[=create_cg_super_regions]{create_cg_super_regions()}}.}

\item{pos}{\code{tibble} The positives file from the output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}.}

\item{start_year}{\code{int} The the earliest year for analysis. Defaults to 2016.}
}
\value{
\code{sf} All CG related positives flagged and ready for mapping with a point geography.
}
\description{
Obtains positive detections and determines whether they are part of the
consequential geographies or not. If they are, then they are flagged in the column \code{in_cg} as
\code{TRUE}.
}
\examples{
\dontrun{
polio_data <- get_all_polio_data()
cg <- sirfunctions_io("read", file_loc = "Data/misc/consequential_geographies.rds")
super_regions <- create_cg_super_regions(cg,
ctry = polio_data$global.ctry, prov = polio_data$global.prov, dist = polio_data$global.dist)
cg_positives <- flag_cg_positives(cg_super_regions, polio_data$pos)
}
}
