% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{force_load_polio_data_cache}
\alias{force_load_polio_data_cache}
\title{Force load data from the local cache}
\usage{
force_load_polio_data_cache(attach.spatial.data, output_format = ".rds")
}
\arguments{
\item{attach.spatial.data}{\code{logical} Should spatial data be attached?}

\item{output_format}{\code{str} Output format of the file to load to R.}
}
\value{
\code{list} List containing information pertinent to the global polio dataset.
}
\description{
In certain instances, it may be desirable to load the global polio data directly from the local
cache. For example, if connection from EDAV fails or if there are internet connection issues.
}
\examples{
\dontrun{
force_load_polio_data_cache(attach.spatial.data = TRUE)
}
}
