% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{freeze_dr_data}
\alias{freeze_dr_data}
\title{Freeze desk review data to the desk review folder in EDAV}
\usage{
freeze_dr_data(
  rds_obj,
  file_name,
  country = Sys.getenv("DR_COUNTRY"),
  year = as.numeric(format(Sys.Date(), "\%Y"))
)
}
\arguments{
\item{rds_obj}{\code{Robj} Object loaded in R. This would be \code{ctry.data}, for example.}

\item{file_name}{\code{str} Name given to the Rds object, do not append \code{.rds}. This is
what gets stored in EDAV.}

\item{country}{\code{str} Country as a string.}

\item{year}{\code{int} It is recommended to set this to the year
when the desk review was ran.}
}
\value{
A status message
}
\description{
Data from the desk review can be stored in EDAV so there's an exact copy of the
dataset used in the desk review. This ensures that even after years,
the desk reviews can be ran exactly as it was.
}
\examples{
\dontrun{
raw.data <- get_all_polio_data()
ctry.data <- init_dr("algeria")
freeze_dr_data(ctry.data, "algeria_ctry_data")
}

}
\seealso{
\code{\link[=init_dr]{init_dr()}}
}
