% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.parquet.R
\name{from_wkb_to_sf}
\alias{from_wkb_to_sf}
\title{Convert WKB back to sf column}
\usage{
from_wkb_to_sf(sf_data)
}
\arguments{
\item{sf_data}{\verb{arrow connection} Geodata arrow connection.}
}
\value{
\code{tibble} Geodata with \code{sf}.
}
\description{
Convert WKB back to sf column
}
\examples{
\dontrun{
raw_data <- build_parquet_raw_data()
kenya_ctry_sf <- raw_data$global.ctry |> 
    dplyr::filter(ctry == "KENYA") |> 
    dplyr::collect() |> 
    from_wkb_to_sf()
}
}
