% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_60_day_tab}
\alias{generate_60_day_tab}
\title{Table used for those requiring 60 day follow up}
\usage{
generate_60_day_tab(cases.need60day)
}
\arguments{
\item{cases.need60day}{summary table containing those that need 60 day follow-up}
}
\value{
a flextable with 60 day follow up data
}
\description{
Table used for those requiring 60 day follow up
}
