% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_60_day_table_data}
\alias{generate_60_day_table_data}
\title{Generate summary table for those requiring 60 day follow up}
\usage{
generate_60_day_table_data(stool.data, start_date, end_date)
}
\arguments{
\item{stool.data}{AFP data with stool adequacy columns}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}
}
\value{
a summary table for those requiring 60 day follow up
}
\description{
Generate summary table for those requiring 60 day follow up
}
