% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{generate_adhoc_map}
\alias{generate_adhoc_map}
\title{Create adhoc maps for emergences}
\usage{
generate_adhoc_map(
  raw.data,
  country,
  virus_type = "cVDPV 2",
  vdpv = T,
  new_detect = T,
  surv = c("AFP", "ES", "OTHER"),
  labels = "YES",
  owner = "CDC-GID-PEB",
  new_detect_expand = F,
  start_date = NULL,
  end_date = NULL,
  output = NULL,
  image_size = NULL,
  height = 6.2,
  width = 4.5,
  scale = 1.25,
  dpi = 300
)
}
\arguments{
\item{raw.data}{global polio data}

\item{country}{\code{str} or \code{list} country name or a list of country names}

\item{virus_type}{\code{str} virus type to include. Valid entries are "cVDPV 1", "cVDPV 2", "cVDPV 3", "WILD 1". Can pass a list.}

\item{vdpv}{\code{bool} whether to include VPDV in maps. Default TRUE.}

\item{new_detect}{\code{bool} whether to highlight new detections based on WHO HQ report date. Default TRUE.}

\item{surv}{surveillance options. AFP, ES, OTHER (Includes Case Contact, Community, Healthy Children Sampling).}

\item{labels}{Include labels for regions with virus detections.
Options: "ALL": All regions, "YES": Recent Detections - <13 months}

\item{owner}{who produced the map. Defaults to "CDC-GID-PEB"}

\item{new_detect_expand}{whether to expand the reporting window. Defaults to FALSE.}

\item{start_date}{start date. If not specified, defaults to 13 months prior to the download date of raw.data}

\item{end_date}{end date. If not specified, defaults to the download date of raw.data}

\item{output}{either a path to a local folder to save the map to, "sharepoint", or none. Defaults to NULL.}

\item{image_size}{standard sizes of the map outputs. Options are: "full_slide", "soco_slide", "half_slide". Defaults to NULL.}

\item{height}{height of the map}

\item{width}{width of the map}

\item{scale}{scale of the map}

\item{dpi}{dpi of the map}
}
\description{
Create adhoc maps for emergences
}
