% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_afp_by_month_summary}
\alias{generate_afp_by_month_summary}
\title{Generate summary from the afp.by.month summary tibble}
\usage{
generate_afp_by_month_summary(
  afp.by.month,
  ctry.data,
  start_date,
  end_date,
  by
)
}
\arguments{
\item{afp.by.month}{summary table of AFP cases}

\item{ctry.data}{RDS file containing country data}

\item{start_date}{start date of analysis}

\item{end_date}{end date of analysis}

\item{by}{either prov, dist, or year}
}
\value{
tibble summary table of AFP cases by month
}
\description{
Generate summary from the afp.by.month summary tibble
}
