% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_afp_ctry_year}
\alias{generate_afp_ctry_year}
\title{AFP cases by ctry and year}
\usage{
generate_afp_ctry_year(
  afp.by.month.ctry,
  start_date,
  end_date = lubridate::today(),
  output_path = Sys.getenv("DR_FIGURE_PATH"),
  .height = 5
)
}
\arguments{
\item{afp.by.month.ctry}{\code{tibble} Table summarizing AFP cases by month and province. This is the output of
\code{\link[=generate_afp_by_month_summary]{generate_afp_by_month_summary()}}.}

\item{start_date}{\code{str} Start date of the analysis.}

\item{end_date}{\code{str} End date of the analysis. By default, it displays the most recent date.}

\item{output_path}{\code{str} Local path to output the figure.}

\item{.height}{\code{int} Change the height of the figure. Defaults to 5.}
}
\value{
\code{ggplot} A tile plot displaying the number of AFP cases by month and province.
}
\description{
Generates a tile plot for the number of AFP cases per month by country.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
afp.by.month <- generate_afp_by_month_summary(
  raw.data$afp, "2021-01-01", "2023-12-31", "ctry",
  raw.data$ctry.pop
generate_afp_prov_year(afp.by.month, start_date, end_date)
}

}
