% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_afp_epicurve}
\alias{generate_afp_epicurve}
\title{Generate epicurve of AFP cases by year}
\usage{
generate_afp_epicurve(ctry.data, start_date, end_date)
}
\arguments{
\item{ctry.data}{RDS data countaining polio data for a country}

\item{start_date}{start date of the desk review}

\item{end_date}{end date of the desk review}
}
\value{
ggplot object containing the epicurve data
}
\description{
Generate epicurve of AFP cases by year
}
