% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_afp_prov_year}
\alias{generate_afp_prov_year}
\title{AFP cases by province and year}
\usage{
generate_afp_prov_year(
  afp.by.month.prov,
  start_date,
  end_date = lubridate::today(),
  output_path = Sys.getenv("DR_FIGURE_PATH")
)
}
\arguments{
\item{afp.by.month.prov}{\code{tibble} Table summarizing AFP cases by month and province. This is the output of
\code{\link[=generate_afp_by_month_summary]{generate_afp_by_month_summary()}}.}

\item{start_date}{\code{str} Start date of the analysis.}

\item{end_date}{\code{str} End date of the analysis. By default, it displays the most recent date.}

\item{output_path}{\code{str} Local path to output the figure.}
}
\value{
\code{ggplot} A tile plot displaying the number of AFP cases by month and province.
}
\description{
Generates a tile plot for the number of AFP cases per month by province.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
afp.by.month <- generate_afp_by_month(ctry.data$afp.all.2, start_date, end_date)
afp.by.month.prov <- generate_afp_by_month_summary(afp.by.month, ctry.data,
                                                   start_date, end_date, "prov"
                                                   )
generate_afp_prov_year(afp.by.month.prov, start_date, end_date)
}

}
