% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.table.functions.R
\name{generate_c1_rollup}
\alias{generate_c1_rollup}
\title{Generate C1 rollup for high-priority countries}
\usage{
generate_c1_rollup(
  c1,
  priority_level = "HIGH",
  who_region = NULL,
  .group_by = "rolling_period"
)
}
\arguments{
\item{c1}{\code{tibble} The output of \code{\link[=generate_c1_table]{generate_c1_table()}}.}

\item{priority_level}{\verb{str or list} Priority level. Defaults to \code{"HIGH"}. Valid
values are \verb{"LOW", "LOW (WATCHLIST)", "MEDIUM", "HIGH"}}

\item{who_region}{\code{str} WHO region to summarize the data to.}

\item{.group_by}{\code{str} How the rollup should be grouped. Defaults to the column \code{"rolling_period"}.}
}
\value{
\code{tibble} A summary rollup
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Generates a summary of how many of the high priority countries have met their
AFP and ES indicators.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
c1 <- generate_c1_table(raw_data, "2022-01-01", "2024-12-31")
c1_rollup <- generate_c1_rollup(c1)
}
}
