% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kppi.afp.functions.R
\name{generate_c1_table}
\alias{generate_c1_table}
\title{GPEI Strategy surveillance KPIs}
\usage{
generate_c1_table(
  raw_data,
  start_date,
  end_date,
  lab_locs = NULL,
  risk_table = NULL,
  rolling = T
)
}
\arguments{
\item{raw_data}{\code{list} Global polio surveillance dataset.
Output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}.}

\item{start_date}{\code{str} Start date of the analysis in YYYY-MM-DD format.}

\item{end_date}{\code{str} End date of the analysis in YYYY-MM-DD format.}

\item{lab_locs}{\code{tibble} Summary of the sequencing capacities of labs.
Output of \code{\link[=get_lab_locs]{get_lab_locs()}}. Defaults to \code{NULL}, which will download the information
directly from EDAV.}

\item{risk_table}{\code{tibble} GPSAP risk categorization for each country.}

\item{rolling}{\code{bool} Should the data be summarized for 12-month rolling averages. Defaults to \code{TRUE}.}
}
\value{
\code{tibble} Summary table of GPSAP KPIs.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Monitoring surveillance KPIs for Certification of Poliomyelitis Eradication at country
regional and global levels.
}
\examples{
raw_data <- get_all_polio_data(attach.spatial.data = FALSE)
c1 <- generate_c1_table(raw_data, "2021-01-01", "2023-12-31")
}
