% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.afp.functions.R
\name{generate_c4_table}
\alias{generate_c4_table}
\title{Laboratory surveillance KPIs}
\usage{
generate_c4_table(lab_data, afp_data, start_date, end_date, .group_by = "year")
}
\arguments{
\item{lab_data}{\code{tibble} Lab data containing information of ES or AFP samples.}

\item{afp_data}{\code{tibble} AFP surveillance data.}

\item{start_date}{\code{str} Start date of the analysis in YYYY-MM-DD format.}

\item{end_date}{\code{str} End date of the analysis in YYYY-MM-DD format.}

\item{.group_by}{\code{str} or \code{list} What columns to group the results by.}
}
\value{
\code{tibble} A summary of timeliness KPIs for lab data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Summarizes the timeliness of samples as it arrives in the lab and to sequencing
results. Samples may come from both AFP and ES samples.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data(attach.spatial.data = FALSE)
lab_data <- readr::read_csv("C:/Users/ABC1/Desktop/lab_data.csv")
c4 <- generate_c4_table(lab_data, raw_data$afp, "2021-01-01", "2024-12-31")
}
}
