% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_case_num_dose_g}
\alias{generate_case_num_dose_g}
\title{Immunization case status per year}
\usage{
generate_case_num_dose_g(
  ctry.data,
  start_date,
  end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH")
)
}
\arguments{
\item{ctry.data}{\code{list} A large list containing polio data of country.
This is the output of \code{\link[=extract_country_data]{extract_country_data()}} or \code{\link[=init_dr]{init_dr()}}. Note that \code{ctry_data} needs to be cleaned
via \code{\link[=clean_ctry_data]{clean_ctry_data()}} prior to running the function.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}

\item{output_path}{\code{str} Local path of where to save the figure to.}
}
\value{
\code{ggplot} A percent bar plot displaying NPAFP cases per year by immunization status.
}
\description{
Generates 4 stacked percent bar plots displaying immunization status of NPAFP cases per year for the country.
Note that this function only graphs immunization rates for children aged 6-59 months.
The "Missing" category is cases that have no value (i.e. NA) for doses.total.
The "Unknown" category has "99" for doses.total.
4 plots: 1 by total number of doses (combined OPV and IPV), 1 by SIA OPV, 1 by Routine OPV, 1 by IPV
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
ctry.data <- clean_ctry_data(ctry.data)
generate_case_num_dose_g(ctry.data, "2021-01-01", "2023-12-31")
}

}
