% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{generate_data}
\alias{generate_data}
\title{Handles the logic of which files to load in the current R session}
\usage{
generate_data(
  data_path,
  data_size,
  country_name,
  dr_data_path,
  attach_spatial_data,
  .data_folder,
  .polis_folder,
  .use_edav
)
}
\arguments{
\item{data_path}{path of the data folder used in the desk review}

\item{data_size}{\code{str} "small", "medium", or "large"}

\item{country_name}{\code{str} name of the country}

\item{dr_data_path}{\code{str} path to the dataset}

\item{attach_spatial_data}{whether to attach spatial data}

\item{.data_folder}{\code{str} Path to the data folder.}

\item{.polis_folder}{\code{str} Path to the POLIS folder.}

\item{.use_edav}{\code{bool} Whether to use EDAV or not.}
}
\value{
\code{list} large list containing polio data
}
\description{
Handles the logic of which files to load in the current R session
}
\keyword{internal}
