% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_dist_pop_map}
\alias{generate_dist_pop_map}
\title{Generate a map of population by district}
\usage{
generate_dist_pop_map(ctry.data, prov.shape, dist.shape, end_date)
}
\arguments{
\item{ctry.data}{RDS file of polio country data}

\item{prov.shape}{recent shape file of province}

\item{dist.shape}{recent shape file of district}

\item{end_date}{end date of the desk review}
}
\value{
a ggplot object
}
\description{
Generate a map of population by district
}
