% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.powerpoint.functions.R
\name{generate_dr_ppt}
\alias{generate_dr_ppt}
\title{Generate the desk review slide deck}
\usage{
generate_dr_ppt(
  ppt_template_path,
  ctry.data,
  start_date,
  end_date,
  pop.map,
  pop.map.prov,
  afp.case.map,
  afp.epi.curve,
  surv.ind.tab,
  afp.dets.prov.year,
  pop.tab,
  npafp.map,
  npafp.map.dist,
  stool.ad.maps,
  stool.ad.maps.dist,
  inad.tab.flex,
  tab.60d,
  case.num.dose.g,
  timely_nation,
  timely_prov,
  mapt_all,
  es.site.det,
  es.det.map,
  es.timely,
  es.table,
  country = Sys.getenv("DR_COUNTRY"),
  ppt_output_path = Sys.getenv("DR_POWERPOINT_PATH")
)
}
\arguments{
\item{ppt_template_path}{\code{str} Sath to the PowerPoint template.}

\item{ctry.data}{\code{list} List containing polio data for a country. Either the output of
\code{\link[=extract_country_data]{extract_country_data()}} or \code{\link[=init_dr]{init_dr()}}.}

\item{start_date}{\code{str} Start date of desk review.}

\item{end_date}{\code{str} End date of desk review.}

\item{pop.map}{\code{ggplot} Country pop map.}

\item{pop.map.prov}{\code{ggplot} Prov pop map.}

\item{afp.case.map}{\code{ggplot} Map of afp cases.}

\item{afp.epi.curve}{\code{ggplot} AFP epicurve.}

\item{surv.ind.tab}{\code{flextable} Surveillance indicator table.}

\item{afp.dets.prov.year}{\code{ggplot} AFP detections for province.}

\item{pop.tab}{\code{flextable} Table of population.}

\item{npafp.map}{\code{ggplot} NPAFP map for country level.}

\item{npafp.map.dist}{\code{ggplot} NPAFP map for district level.}

\item{stool.ad.maps}{\code{ggplot} Stool adequacy maps at province.}

\item{stool.ad.maps.dist}{\code{ggplot} Stool adequacy maps at district.}

\item{inad.tab.flex}{\code{flextable} Inadequate table.}

\item{tab.60d}{\code{flextable} 60-day follow-up table.}

\item{case.num.dose.g}{\code{ggplot} Immunization rates per year.}

\item{timely_nation}{\code{ggplot} Timeliness at country level.}

\item{timely_prov}{\code{ggplot} Timeliness at province level.}

\item{mapt_all}{\code{ggplot} Map with all indicators.}

\item{es.site.det}{\code{ggplot} ES site viral detection.}

\item{es.det.map}{\code{ggplot} ES site detection maps.}

\item{es.timely}{\code{ggplot} ES timeliness.}

\item{es.table}{\code{flextable} ES table.}

\item{country}{\code{str} Name of the country.}

\item{ppt_output_path}{\code{str} Path where the PowerPoint should be outputted.}
}
\value{
None.
}
\description{
The original function to build the desk review PowerPoint. This function has been
superseded by \code{\link[=generate_dr_ppt2]{generate_dr_ppt2()}}. The function outputs images to the PowerPoint
directly from objects, unlike \code{generate_dr_ppt2()} which uses images saved in a folder.
}
\examples{
\dontrun{
# Assume all figures and tables are assigned to the appropriate variable.
template_path <- "C:/Users/ABC1/Desktop/deskreview_template.pptx"
generate_dr_ppt(
  template_path, ctry.data, start_date, end_date, pop.map,
  pop.map, pop.map.prov, afp.case.map, afp.epi.curve,
  surv.ind.tab, afp.dets.prov.year, pop.tab, npafp.map,
  npafp.map.dist, stool.ad.maps, stool.ad.maps.dist,
  inad.tab.flex, tab.60d, case.num.dose.g,
  timely_nation, timely_prov,
  mapt_all, es.site.det, es.det.map, es.timely,
  es.table
)
}

}
