% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.powerpoint.functions.R
\name{generate_dr_ppt2}
\alias{generate_dr_ppt2}
\title{Generate the desk review slide deck from the figures folder
Generating the PowerPoint from the figures folder is generally faster and allows
figures to remain consistent. Tables remain as PowerPoint tables.}
\usage{
generate_dr_ppt2(
  ppt_template_path,
  ctry.data,
  assump,
  start_date,
  end_date,
  fig.path,
  surv.ind.tab,
  inad.tab.flex,
  tab.60d,
  es.table,
  country,
  ctry.data.error.log,
  ppt_output_path
)
}
\arguments{
\item{ppt_template_path}{path to the PowerPoint template}

\item{ctry.data}{RDS file containing polio data for a country}

\item{assump}{list of assumptions}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}

\item{fig.path}{file path to the figures folder}

\item{surv.ind.tab}{surveillance indicator table}

\item{inad.tab.flex}{inadequacy table}

\item{tab.60d}{60 day follow up table}

\item{es.table}{ES table}

\item{country}{name of the country}

\item{ppt_output_path}{path where the powerpoint should be outputted}
}
\value{
does not return anything
}
\description{
Generate the desk review slide deck from the figures folder
Generating the PowerPoint from the figures folder is generally faster and allows
figures to remain consistent. Tables remain as PowerPoint tables.
}
