% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.es.functions.R
\name{generate_es_data_long}
\alias{generate_es_data_long}
\title{Generate ES data with viral detection columns}
\usage{
generate_es_data_long(es.data)
}
\arguments{
\item{es.data}{tibble containing ES data}
}
\value{
tibble of ES data with viral detection columns
}
\description{
Generate ES data with viral detection columns
}
