% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_es_site_det}
\alias{generate_es_site_det}
\title{Dot plot of virus detections in ES sites}
\usage{
generate_es_site_det(
  ctry.data,
  es.data.long,
  es_start_date = (end_date - lubridate::years(1)),
  es_end_date = end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH"),
  vaccine_types = NULL,
  detection_types = NULL
)
}
\arguments{
\item{ctry.data}{RDS file of polio data for a country}

\item{es.data.long}{AFP data with viral detection columns}

\item{es_start_date}{start date of ES data}

\item{es_end_date}{end date of ES data}

\item{output_path}{where to save the figure}

\item{vaccine_types}{named list with vaccine types with color associations}

\item{detection_types}{named list with detection types with color associations}
}
\value{
ggplot object dot plot
}
\description{
Dot plot of virus detections in ES sites
}
