% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_es_tab}
\alias{generate_es_tab}
\title{ES table with indicators}
\usage{
generate_es_tab(
  es.data,
  es_start_date = (end_date - lubridate::years(1)),
  es_end_date = end_date
)
}
\arguments{
\item{es.data}{ES data}

\item{es_start_date}{ES start date}

\item{es_end_date}{ES end date}
}
\value{
flextable with ES indicators
}
\description{
ES table with indicators
}
