% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_inad_tab_region.R
\name{generate_inad_tab_region}
\alias{generate_inad_tab_region}
\title{Generate a stool adequacy summary at the regional level}
\usage{
generate_inad_tab_region(afp_data, cstool, start_date, end_date)
}
\arguments{
\item{afp_data}{\code{tibble} \code{afp.all.2} of the output of \code{\link[=extract_country_data]{extract_country_data()}}.}

\item{cstool}{\code{tibble} Stool adequacy at the country level. This is the output of \code{\link[=f.stool.ad.01]{f.stool.ad.01()}}.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}
}
\value{
\code{flextable} A summary of yearly stool adequacy at the regional level.
}
\description{
Generates a summary table at the country level highlighting issues around stool adequacy.
}
\examples{
\dontrun{
inad_region <- generate_inad_tab_region(ctry_data$afp.all.2, cstool)
}
}
