% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_incidence_map.R
\name{generate_incidence_map}
\alias{generate_incidence_map}
\title{Generate incidence map timeline}
\usage{
generate_incidence_map(
  pos_data,
  ctry_sf,
  emergence_group = NULL,
  emergence_colors = NULL,
  sources = c("AFP", "ENV"),
  start_date = NULL,
  end_date = NULL,
  monthly_rolling_window = 12,
  output_dir = NULL,
  fps = 2,
  drop_legend = FALSE,
  drop_ctry_labels = FALSE,
  drop_description = FALSE,
  pt_size = 0.7
)
}
\arguments{
\item{pos_data}{\code{tibble} Positives dataset.}

\item{ctry_sf}{\code{ctry_sf} Global country shapefile.}

\item{emergence_group}{\code{str} An emergence group or a vector of emergence group names.}

\item{emergence_colors}{\code{list} A named list where each emergence group is mapped its own
color. Names correspond to the emergence group, while values are the colors. If no colors
are passed, then colors are randomly selected for each emergence group.}

\item{sources}{\code{str} Source of detection or a vector of source names. Valid values are:
"AFP", "Community", "Contact", "ENV", "Healthy", "iVDPV".}

\item{start_date}{\code{str} Start date of the map. By default, it will be the earliest date of
detection.}

\item{end_date}{\code{str} End date of the map. By default, it will be the latest detection.}

\item{monthly_rolling_window}{\code{int} Monthly rolling window to show. Defaults to 12.}

\item{output_dir}{\code{str} Local path to the directory to output the figure. Defaults to NULL,
which does not export the figure.}

\item{fps}{\code{int} Frames per second. To increase the speed of the GIF, increase the fps.
By default, it is set to 2.}

\item{drop_legend}{\code{logical} Drop legends of the figure or not.}

\item{drop_ctry_labels}{\code{logical} Drop country labels for the map.}

\item{drop_description}{\code{logical} Drop the case counts of the map caption.}

\item{pt_size}{\code{int} Size of the points in the map. Defaults to 5.}
}
\value{
\code{gif} A GIF showing the emergence over time.
}
\description{
Generates an incidence map and an epicurve showing positive detections for
emergence groups.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data
generate_incidence_map(raw_data$pos, raw_data$global.ctry, emergence_group = c("YEM-TAI-1", "SOM-BAN-1", "ETH-TIG-1"))
}
}
