% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_int_data}
\alias{generate_int_data}
\title{Generate a summary table for sample timeliness intervals}
\usage{
generate_int_data(
  ctry.data,
  start_date,
  end_date,
  spatial.scale,
  lab.data = NULL
)
}
\arguments{
\item{ctry.data}{RDS object containing polio country data}

\item{start_date}{start date of the desk review}

\item{end_date}{end date of the desk review}

\item{spatial.scale}{Scale to summarize to. Either "ctry" or "prov". "dist" not available currently.}

\item{lab.data}{tibble of lab data}
}
\value{
a tibble summarizing median days for different intervals
}
\description{
Generate a summary table for sample timeliness intervals
}
