% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_iss_map}
\alias{generate_iss_map}
\title{Generate ISS map}
\usage{
generate_iss_map(
  iss.data,
  prov.shape,
  start_date,
  end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH")
)
}
\arguments{
\item{iss.data}{tibble of ISS/eSurv data}

\item{prov.shape}{shapefile containing province shapes}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}

\item{output_path}{where to save the figure}
}
\value{
a ggplot map
}
\description{
Generate ISS map
}
