% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_kpi_barchart}
\alias{generate_kpi_barchart}
\title{Generate KPI indicator bar charts}
\usage{
generate_kpi_barchart(df, indicator, target, label, faceting, y.axis.title)
}
\arguments{
\item{df}{\code{tibble} A KPI table, namely C1-C4.}

\item{indicator}{\code{str} Name of the indicator within the KPI table.}

\item{target}{\code{num} A numeric target.}

\item{label}{\code{str} Name of the column containing labels.}

\item{faceting}{\code{ggplot2} A ggplot2 faceting object.
Either using \code{\link[ggplot2:facet_grid]{ggplot2::facet_grid()}} or \code{\link[ggplot2:facet_wrap]{ggplot2::facet_wrap()}}.}

\item{y.axis.title}{\code{Str} Title of the y axis.}
}
\value{
\code{ggplot2} A bar chart.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

A generalized function to create bar charts using the KPI tables.
}
