% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.table.functions.R
\name{generate_kpi_lab_timeliness}
\alias{generate_kpi_lab_timeliness}
\title{Generate KPI lab intervals}
\usage{
generate_kpi_lab_timeliness(lab_data, start_date, end_date, afp_data = NULL)
\arguments{
\item{lab_data}{\code{tibble} Lab data containing information of ES or AFP samples.}

\item{start_date}{\code{str} Start date of the analysis in YYYY-MM-DD format.}

\item{end_date}{\code{str} End date of the analysis in YYYY-MM-DD format.}

\item{afp_data}{\code{tibble} AFP surveillance data.}
}
\value{
\code{tibble} lab data with timeliness columns.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Generates lab intervals related to the KPI code.
}
\details{
This function is used in both \link{generate_c4_table} and in the lab timeliness
KPI violin plots.
}
\keyword{internal}
