% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_kpi_map}
\alias{generate_kpi_map}
\title{Generate KPI maps}
\usage{
generate_kpi_map(
  c2,
  ctry_sf,
  who_region,
  indicator,
  .year,
  color_scheme,
  legend_title
)
}
\arguments{
\item{c2}{\code{tibble} Output of \code{\link[=generate_c2_table]{generate_c2_table()}}.}

\item{ctry_sf}{\code{sf} Country shapefile in long format.}

\item{who_region}{\code{str} WHO region. Valid values are:
\itemize{
\item \code{"AFRO"}: African Region
\item \code{"AMRO"}: Region of the Americas
\item \code{"EMRO"}: Eastern Mediterranean Region
\item \code{"EURO"}: European Region
\item \code{"SEARO"}:South-East Asia Region
\item \code{"WPRO"}:Western Pacific Region
}}

\item{indicator}{\verb{quoted var} Indicator variable.}

\item{.year}{\code{num} Numeric year of interest.}

\item{color_scheme}{\code{list} Named list with color mappings.}

\item{legend_title}{\code{str} Title of the legend.}
}
\value{
\code{ggplot} A ggplot object.
}
\description{
Generalized function to build KPI maps.
}
\keyword{internal}
