% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_kpi_stool_map}
\alias{generate_kpi_stool_map}
\title{Generate district level stool adequacy maps}
\usage{
generate_kpi_stool_map(
  c2,
  ctry_sf,
  dist_sf,
  who_region,
  .year,
  output_path = Sys.getenv("KPI_FIGURES")
)
}
\arguments{
\item{c2}{\code{tibble} Output of \code{\link[=generate_c2_table]{generate_c2_table()}}.}

\item{ctry_sf}{\code{sf} Country shapefile in long format.}

\item{dist_sf}{\code{sf} District shapefile in long format.}

\item{who_region}{\code{str} WHO region. Valid values are:
\itemize{
\item \code{"AFRO"}: African Region
\item \code{"AMRO"}: Region of the Americas
\item \code{"EMRO"}: Eastern Mediterranean Region
\item \code{"EURO"}: European Region
\item \code{"SEARO"}:South-East Asia Region
\item \code{"WPRO"}:Western Pacific Region
}}

\item{.year}{\code{num} Numeric year of interest.}

\item{output_path}{\code{str} Where to output the figure to. Defaults to the path
initialized when \code{\link[=init_kpi]{init_kpi()}} was ran.}
}
\value{
\code{ggplot} A stool adequacy map.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Generates a map of district level stool adequacy maps
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
c2 <- generate_c2_table(raw_data$afp, raw_data$dist, "2021-01-01",
"2023-12-31", c("ctry", "dist", "adm2guid", "year"))
ctry_sf <- load_clean_ctry_sp(type="long")
dist_sf <- load_clean_dist_sp(type="long")
map <- generate_kpi_stool_map(c2, ctry_sf, dist_sf, "AFRO", 2022, getwd())
}
}
