% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_kpi_stoolad_bar}
\alias{generate_kpi_stoolad_bar}
\title{Generate KPI Stool adequacy bar chart summary}
\usage{
generate_kpi_stoolad_bar(c1, afp_data, output_path = Sys.getenv("KPI_FIGURES"))
}
\arguments{
\item{c1}{\code{tibble} Output of \code{\link[=generate_c1_table]{generate_c1_table()}}.}

\item{afp_data}{\code{tibble} AFP dataset. List item of the output of
\code{\link[=get_all_polio_data]{get_all_polio_data()}}.}

\item{output_path}{\code{str} Folder location to output the image to.}
}
\value{
\code{ggplot2} A barplot.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Generates a bar chart highlighting percentage of priority countries
achieving stool adequacy targets.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
c1 <- generate_c1_table(raw_data, "2021-01-01", "2023-12-31", rolling = TRUE)
plot <- generate_kpi_stoolad_bar(c1, raw_data$afp)
}
}
