% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_kpi_violin}
\alias{generate_kpi_violin}
\title{Generates KPI violin plot}
\usage{
generate_kpi_violin(
  df,
  country.label,
  interval,
  priority_level,
  faceting,
  target,
  y.min = 0,
  y.max
)
}
\arguments{
\item{df}{\code{tibble} Data to be used.}

\item{country.label}{\verb{quoted var} Country label.}

\item{interval}{\verb{quoted var} Interval to use.}

\item{priority_level}{\verb{quoted var} Priority level column.}

\item{faceting}{\code{ggplot::facet} A faceting object.}

\item{target}{\code{num} Numeric target.}

\item{y.min}{\code{num} Minimum used in the y-axis.}

\item{y.max}{\code{num} Maximum used in the y-axis.}
}
\value{
\code{ggplot} A plot object.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is generalized to produce violin plots used in the KPI code.
}
\keyword{internal}
