% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_lab_culture_violin}
\alias{generate_lab_culture_violin}
\title{Timeliness of lab samples from collection to virus isolation results}
\usage{
generate_lab_culture_violin(
  lab_data,
  afp_data,
  start_date,
  end_date,
  output_path = Sys.getenv("KPI_FIGURES"),
  y_max = 250
)
}
\arguments{
\item{lab_data}{\code{tibble} Global lab dataset.}

\item{afp_data}{\code{tibble} AFP dataset.}

\item{start_date}{\code{str} Start date of the analysis formatted as "YYYY-MM-DD".}

\item{end_date}{\code{str} End date of the analysis formatted as "YYYY-MM-DD".}

\item{output_path}{\code{str} Where to output the figure to.}

\item{y_max}{\code{num} Maximum value in the y-axis.}
}
\value{
\code{ggplot} A violin plot showing timeliness of lab culture.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Shows the timeliness of lab culture results.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
lab_data <- edav_io("read", file_loc = get_constant("CLEANED_LAB_DATA"))
generate_lab_culture_violin(lab_data, raw_data$afp,
                            "2021-01-01", "2023-12-31", getwd())
}
}
