% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_lab_itd_violin}
\alias{generate_lab_itd_violin}
\title{Timeliness of virus isolation to ITD results}
\usage{
generate_lab_itd_violin(
  lab_data,
  afp_data,
  start_date,
  end_date,
  priority_level = c("HIGH", "MEDIUM", "LOW (WATCHLIST)", "LOW"),
  who_region = NULL,
  rolling = TRUE,
  output_path = Sys.getenv("KPI_FIGURES"),
  y_max = 60
)
}
\arguments{
\item{lab_data}{\code{tibble} Global lab dataset.}

\item{afp_data}{\code{tibble} AFP dataset.}

\item{start_date}{\code{str} Start date of the analysis formatted as "YYYY-MM-DD".}

\item{end_date}{\code{str} End date of the analysis formatted as "YYYY-MM-DD".}

\item{priority_level}{\code{list} Priority levels to display. Defaults to
\code{c("HIGH", "MEDIUM", "LOW (WATCHLIST)", "LOW")}.}

\item{who_region}{\code{list} Regions to display. Defaults to \code{NULL}, which shows
all of the regions.}

\item{rolling}{\code{bool} Using rolling periods or year-to-year? Defaults to \code{TRUE}.}

\item{output_path}{\code{str} Where to output the figure to.}

\item{y_max}{\code{num} Maximum value in the y-axis.}
}
\value{
\code{ggplot} A violin plot showing timeliness of ITD results from lab
culture.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Shows the timeliness of ITD results of specimens that require ITD. The target
is 7 days or less.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
lab_data <- edav_io("read", file_loc = get_constant("CLEANED_LAB_DATA"))
generate_lab_itd_violin(lab_data, raw_data$afp,
                        "2021-01-01", "2023-12-31", getwd())
}
}
