% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{generate_lab_timeliness}
\alias{generate_lab_timeliness}
\title{Generate timeliness intervals with lab data}
\usage{
generate_lab_timeliness(lab.data, spatial.scale, start.date, end.date)
}
\arguments{
\item{lab.data}{lab data}

\item{spatial.scale}{spatial scale to analyze the data. Valid values are "ctry", "prov", "dist"}

\item{start.date}{start date of analysis}

\item{end.date}{end date of analysis}
}
\value{
a table with timeliness data summary
}
\description{
Generate timeliness intervals with lab data
}
