% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_npafp_maps}
\alias{generate_npafp_maps}
\title{Map of NPAFP rate by province}
\usage{
generate_npafp_maps(prov.extract, ctry.shape, prov.shape, start_date, end_date)
}
\arguments{
\item{prov.extract}{province NPAFP rate}

\item{ctry.shape}{recent country shape}

\item{prov.shape}{recent province shape}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}
}
\value{
a map of NPAFP rates by province
}
\description{
Map of NPAFP rate by province
}
