% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_npafp_maps_dist}
\alias{generate_npafp_maps_dist}
\title{Generate map of NPAFP rates by district}
\usage{
generate_npafp_maps_dist(
  dist.extract,
  ctry.shape,
  prov.shape,
  dist.shape,
  start_date,
  end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH")
)
}
\arguments{
\item{dist.extract}{NPAFP rates by district}

\item{ctry.shape}{recent country shapefile}

\item{prov.shape}{recent province shapefile}

\item{dist.shape}{recent district shapefile}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}

\item{output_path}{where to save the figure}
}
\value{
ggplot object
}
\description{
Generate map of NPAFP rates by district
}
