% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_pop_map}
\alias{generate_pop_map}
\title{Generate a map of population data}
\usage{
generate_pop_map(
  ctry.data,
  ctry.shape,
  prov.shape,
  end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH"),
  caption_size = 11
)
}
\arguments{
\item{ctry.data}{RDS object containing country polio data}

\item{ctry.shape}{country shape file in long format}

\item{prov.shape}{province shape file in long format}

\item{end_date}{end date of the desk review}

\item{output_path}{where to save the figure}

\item{caption_size}{size of the caption}
}
\value{
ggplot object
}
\description{
Generate a map of population data
}
