% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_pop_tab}
\alias{generate_pop_tab}
\title{Summary table of indicators at the province level}
\usage{
generate_pop_tab(
  pnpafp,
  pstool,
  start_date,
  end_date,
  prov.case.ind = lifecycle::deprecated()
)
}
\arguments{
\item{pnpafp}{\code{tibble} NPAFP table. Output of \code{\link[=f.npafp.rate.01]{f.npafp.rate.01()}} at the province level.}

\item{pstool}{\code{tibble} Stool adequacy at province level. Output of \code{\link[=f.stool.ad.01]{f.stool.ad.01()}} at the province level.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}

\item{prov.case.ind}{\code{tibble} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Deprecated in favor of the more informative pnpafp param name.}
}
\value{
\code{flextable} Summary table of province NPAFP and stool adequacy rates per year.
}
\description{
Generates a table summarizing both NPAFP and stool adequacy rates at the province level and by year.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
start_date <- "2021-01-01"
end_date <- "2023-12-31"
prov.extract <- f.npafp.rate.01(
  afp.data = ctry.data$afp.all.2,
  pop.data = ctry.data$prov.pop,
  start.date = start_date,
  end.date = end_date,
  spatial.scale = "prov",
  pending = T,
  rolling = F,
  sp_continuity_validation = F
)
pstool <- f.stool.ad.01(
  afp.data = ctry.data$afp.all.2,
  admin.data = ctry.data$prov.pop,
  start.date = start_date,
  end.date = end_date,
  spatial.scale = "prov",
  missing = "good",
  bad.data = "inadequate",
  rolling = F,
  sp_continuity_validation = F
)
generate_pop_tab(prov.extract, pstool, start_date, end_date)
}

}
