% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_potentially_compatibles_cluster}
\alias{generate_potentially_compatibles_cluster}
\title{Creating a table of compatible and potentially compatible cases.}
\usage{
generate_potentially_compatibles_cluster(cases.need60day, create_cluster = F)
}
\arguments{
\item{cases.need60day}{summary table of cases that need 60 day follow-up}

\item{create_cluster}{whether to use clustering algorithm. Default to F.}
}
\value{
summary table with clustering info
}
\description{
Creating a table of compatible and potentially compatible cases.
}
