% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.powerpoint.functions.R
\name{generate_pptx_assumptions}
\alias{generate_pptx_assumptions}
\title{Generate PowerPoint assumptions}
\usage{
generate_pptx_assumptions(start_date, end_date)
}
\arguments{
\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}
}
\value{
list of strings
}
\description{
Generate PowerPoint assumptions
}
