% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.prov.functions.R
\name{generate_prov_60_day_table}
\alias{generate_prov_60_day_table}
\title{PROV 60-day follow-up table}
\usage{
generate_prov_60_day_table(cases.need60day, prov_name)
}
\arguments{
\item{cases.need60day}{Output from \code{generate_60_day_table_data()}.}

\item{prov_name}{PROV name.}
}
\value{
A flextable.
}
\description{
PROV 60-day follow-up table
}
