% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.prov.functions.R
\name{generate_prov_npafp_age_histogram}
\alias{generate_prov_npafp_age_histogram}
\title{Histogram of NPAFP case age in one prov}
\usage{
generate_prov_npafp_age_histogram(
  ctry.data,
  prov_name,
  start_date,
  end_date,
  binwidth = 1,
  output_path = NULL
)
}
\arguments{
\item{ctry.data}{Country data returned by \code{init_dr()}.}

\item{prov_name}{One value from \code{list_desk_review_prov(ctry.data)}.}

\item{start_date, end_date}{Analysis dates.}

\item{binwidth}{Width of histogram bins in years.}

\item{output_path}{Optional directory in which to save the PNG. \code{NULL} does
not write a file.}
}
\value{
A ggplot object.
}
\description{
Histogram of NPAFP case age in one prov
}
