% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.prov.functions.R
\name{generate_prov_stool_adequacy_map}
\alias{generate_prov_stool_adequacy_map}
\title{PROV-only stool adequacy map by district}
\usage{
generate_prov_stool_adequacy_map(
  dstool,
  prov_name,
  prov.shape,
  dist.shape,
  start_date,
  end_date,
  output_path = NULL,
  caption_size = 3
)
}
\arguments{
\item{dstool}{District output from \code{f.stool.ad.01()}.}

\item{prov_name}{PROV name.}

\item{prov.shape, dist.shape}{Long-format PROV and DIST geometry.}

\item{start_date, end_date}{Analysis dates.}

\item{output_path}{Optional output directory.}

\item{caption_size}{Caption text size.}
}
\value{
A ggplot object.
}
\description{
PROV-only stool adequacy map by district
}
