% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.prov.functions.R
\name{generate_prov_stool_adequacy_table}
\alias{generate_prov_stool_adequacy_table}
\title{PROV stool adequacy issues table}
\usage{
generate_prov_stool_adequacy_table(
  ctry.data,
  pstool,
  prov_name,
  start_date,
  end_date
)
}
\arguments{
\item{ctry.data}{Country data returned by \code{init_dr()}.}

\item{pstool}{Province output from \code{f.stool.ad.01()}.}

\item{prov_name}{PROV name.}

\item{start_date, end_date}{Analysis dates.}
}
\value{
A flextable.
}
\description{
PROV stool adequacy issues table
}
