% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_prov_timeliness_graph}
\alias{generate_prov_timeliness_graph}
\title{Timeliness intervals of samples at the province level}
\usage{
generate_prov_timeliness_graph(
  int.data,
  afp.prov.year.lab,
  output_path = Sys.getenv("DR_FIGURE_PATH")
)
}
\arguments{
\item{int.data}{\code{tibble} Summary table with timeliness intervals at the province level.}

\item{afp.prov.year.lab}{\code{tibble} Summary table of samples sent by year and province.
Output of \code{\link[=generate_prov_year_lab]{generate_prov_year_lab()}}.}

\item{output_path}{\code{str} Path where to output the figure.}
}
\value{
\code{ggplot} Plot of timeliness intervals at the country level.
}
\description{
A stacked horizontal bar graph for timeliness intervals of samples at the province level.
To get the full intervals from field to lab, the lab data needs to be attached. Otherwise,
only the timeliness intervals from the field up to when it was sent to lab will be displayed.
}
\examples{
\dontrun{
# Attaching lab data
lab_path <- "C:/Users/ABC1/Desktop/algeria_lab_data.csv"
ctry.data <- init_dr("algeria", lab_data_path = lab_path)
lab.timeliness.prov <- generate_lab_timeliness(ctry.data$lab.data, "prov", start_date, end_date)
int.data.prov <- generate_int_data(ctry.data, start_date, end_date,
                                   spatial.scale = "prov",
                                   lab.timeliness.prov
                                   )
generate_ctry_timeliness_graph(int.data.prov, afp.prov.year.lab)
}
}
