% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_prov_year_lab}
\alias{generate_prov_year_lab}
\title{Generate a summary of samples sent to lab by year and province}
\usage{
generate_prov_year_lab(ctry.data, start_date, end_date)
}
\arguments{
\item{ctry.data}{Rds file containing country polio data}

\item{start_date}{start date of the desk review}

\item{end_date}{end date of the desk review}
}
\value{
a tibble containing summary of samples sent to lab by year and province
}
\description{
Generate a summary of samples sent to lab by year and province
}
