% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_stool_ad_maps}
\alias{generate_stool_ad_maps}
\title{Stool adequacy maps by province}
\usage{
generate_stool_ad_maps(
  ctry.data,
  pstool,
  ctry.shape,
  prov.shape,
  start_date,
  end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH"),
  caption_size = 3
)
}
\arguments{
\item{ctry.data}{\code{list} Large list containing polio data of a country. This is the output
of either \code{\link[=extract_country_data]{extract_country_data()}} or \code{\link[=init_dr]{init_dr()}}.}

\item{pstool}{\code{tibble} Stool adequacy table at province level. This is the output of
\code{\link[=f.stool.ad.01]{f.stool.ad.01()}} calculated at the province level.}

\item{ctry.shape}{\code{sf} Country shapefile in long format.}

\item{prov.shape}{\code{sf} Province shapefile in long format.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}

\item{output_path}{\code{str} Where to save the figure to.}

\item{caption_size}{\code{numeric} Size of the caption. Defaults to 3.}
}
\value{
\code{ggplot} A map of stool adequacy rates for each province by year.
}
\description{
Generates maps that contain the stool adequacy rate for each province per year.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
pstool <- f.stool.ad.01(afp.data = ctry.data$afp.all.2,
                        admin.data = ctry.data$prov.pop,
                        start.date = start_date,
                        end.date = end_date,
                        spatial.scale = "prov",
                        missing = "good",
                        bad.data = "inadequate",
                        rolling = F,
                        sp_continuity_validation = F
                       )
ctry.shape <- load_clean_ctry_sp(ctry_name = "ALGERIA", type = "long")
prov.shape <- load_clean_prov_sp(ctry_name = "ALGERIA", type = "long")
generate_stool_ad_maps(ctry.data, pstool, ctry.shape, prov.shape, "2021-01-01", "2023-12-31")
}

}
