% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_timeliness_maps}
\alias{generate_timeliness_maps}
\title{Map containing timeliness of samples}
\usage{
generate_timeliness_maps(
  ctry.data,
  ctry.shape,
  prov.shape,
  start_date,
  end_date
)
}
\arguments{
\item{ctry.data}{RDS file containing polio data for a country}

\item{ctry.shape}{recent country shapefile}

\item{prov.shape}{recent province shapefile}

\item{start_date}{start date of desk review}

\item{end_date}{end dtae of desk review}
}
\value{
a map of timeliness of samples
}
\description{
Map containing timeliness of samples
}
