% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_timely_det_violin}
\alias{generate_timely_det_violin}
\title{Timely detection of AFP/ES samples}
\usage{
generate_timely_det_violin(
  raw_data,
  start_date,
  end_date,
  priority_level = c("HIGH", "MEDIUM", "LOW (WATCHLIST)", "LOW"),
  who_region = NULL,
  rolling = TRUE,
  output_path = Sys.getenv("KPI_FIGURES"),
  y_max = 250
)
}
\arguments{
\item{raw_data}{\code{list} Global polio data. Output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}}

\item{start_date}{\code{str} Analysis start date formatted as "YYYY-MM-DD".}

\item{end_date}{\code{str} Analysis end date formatted as "YYYY-MM-DD".}

\item{priority_level}{\code{list} Priority levels to display. Defaults to
\code{c("HIGH", "MEDIUM", "LOW (WATCHLIST)", "LOW")}.}

\item{who_region}{\code{list} Regions to display. Defaults to \code{NULL}, which shows
all of the regions.}

\item{rolling}{\code{bool} Using rolling periods or year-to-year? Defaults to \code{TRUE}.}

\item{output_path}{\code{str} Where to output the figure to.}

\item{y_max}{\code{num} The maximum y-axis value.}
}
\value{
\code{ggplot} A violin plot showing timeliness of detection.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Generates a violin plot highlighting the median detection time of samples.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
generate_timely_det_violin(raw_data, "2021-01-01", "2023-12-31", getwd())
}
}
