% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kppi.afp.functions.R
\name{generate_wild_vdpv_summary}
\alias{generate_wild_vdpv_summary}
\title{Generates a summary table regarding wild and VDPV cases}
\usage{
generate_wild_vdpv_summary(
  raw_data,
  start_date,
  end_date,
  risk_table = NULL,
  lab_locs = NULL
)
}
\arguments{
\item{raw_data}{\code{list} Global polio surveillance dataset.
Output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}.}

\item{start_date}{\code{str} Start date of the analysis in YYYY-MM-DD format.}

\item{end_date}{\code{str} End date of the analysis in YYYY-MM-DD format.}

\item{risk_table}{\code{tibble} GPSAP risk categorization for each country.}

\item{lab_locs}{\code{tibble} Summary of the sequencing capacities of labs.
Output of \code{\link[=get_lab_locs]{get_lab_locs()}}. Defaults to \code{NULL}, which will download the information
directly from EDAV.}
}
\value{
\code{tibble} Summary of wild and VDPV cases
}
\description{
Generates a summary table regarding wild and VDPV cases
}
\keyword{internal}
