% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.stool.ad.01.R
\name{generate_year_data}
\alias{generate_year_data}
\title{Helper function to generate year.data}
\usage{
generate_year_data(start_date, end_date)
}
\arguments{
\item{start_date}{start date}

\item{end_date}{end date}
}
\value{
a tibble with year data
}
\description{
Helper function to generate year.data
}
\keyword{internal}
