% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_all_polio_data}
\alias{get_all_polio_data}
\title{Retrieve all pre-processed Polio Data}
\usage{
get_all_polio_data(
  folder = "GID/PEB/SIR/Data/",
  afp.trunc = F,
  force.new.run = F
)
}
\arguments{
\item{folder}{str: location of the CDC pre-processed endpoint}

\item{afp.trunc}{boolean: default F, if T returns the afp linelist since 2019}

\item{force.new.run}{boolean: default F, if T will always run fully and cache}
}
\value{
named list containing polio data that is relevant to CDC
}
\description{
Download all Polio data from CDC pre-processed endpoint
}
