% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{get_all_polio_data_2}
\alias{get_all_polio_data_2}
\title{Pull global polio dataset}
\usage{
get_all_polio_data_2(
  dataset = "all",
  data_folder = "GID/PEB/SIR/Data",
  polis_folder = "GID/PEB/SIR/POLIS",
  core_ready_folder = "Core_Ready_Files",
  recreate.static.files = FALSE,
  use_edav = TRUE,
  azcontainer = get_azure_storage_connection(),
  cache = TRUE
)
}
\arguments{
\item{dataset}{\code{str} Name of the dataset. Defaults to 'all'.}

\item{data_folder}{\code{str} Path to data folder.}

\item{polis_folder}{\code{str} Path to the POLIS folder.}

\item{core_ready_folder}{\code{str} Name of the core ready folder.}

\item{recreate.static.files}{\code{logical} Whether to reprocess global polio data.}

\item{use_edav}{\code{logical} Whether to use EDAV or not.}

\item{azcontainer}{\code{azcontainer} Azure container object.}

\item{cache}{\code{logical} Whether to cache the preprocessed datasets in the \code{data/polis} folder.}
}
\value{
\code{list} Global polio datasets.
}
\description{
Pull global polio dataset
}
\examples{
\dontrun{
raw_data <- get_all_polio_data_2()
}
}
