% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_archived_polis_data}
\alias{get_archived_polis_data}
\title{Gets the path to the archived version of the polis folder}
\usage{
get_archived_polis_data(data_folder_path, edav)
}
\arguments{
\item{data_folder_path}{\code{str} Path to the data folder}

\item{edav}{\code{bool} Whether to use EDAV or  not.}
}
\value{
\code{str} Path to the archived polis folder
}
\description{
Obtains the path to the archived version of a polis folder within the
data folder.
}
\keyword{internal}
