% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc_childvaxview_data.R
\name{get_cdc_childvaxview_data}
\alias{get_cdc_childvaxview_data}
\title{Pull CDC NCIRD childvaxview immunization coverage data}
\usage{
get_cdc_childvaxview_data(limit = 1000, geo_level = NULL)
}
\arguments{
\item{limit}{\code{int} Number of rows to download, defaults to max allowed (1000).}

\item{geo_level}{\code{str} Geographic categories of coverage data.
Choose from: 'national', 'regional', 'state', or 'substate'.}
}
\value{
\code{tibble} Dataframe of vaccine coverage estimates for all VPDs.
}
\description{
Pull coverage data from API and filter by desired geographic level
}
\examples{
cdc_data <- get_cdc_childvaxview_data(geo_level="substate")

}
