% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_constant}
\alias{get_constant}
\title{Obtain a constant variable used in sirfunctions}
\usage{
get_constant(constant_name)
}
\arguments{
\item{constant_name}{\code{str} Name of the constant. Valid values include:
\itemize{
\item \code{"DEFAULT_EDAV_FOLDER"}
\item \code{"CTRY_RISK_CAT"}
\item \code{"LAB_LOCATIONS"}
\item \code{"DR_TEMPLATE"}
\item \code{"SIRFUNCTIONS_GITHUB_TREE"}
\item \code{"AFRO_LAB_DATA"}
\item \code{"EMRO_LAB_DATA"}
}}
}
\value{
\code{str} A string, typically a file path or a URL.
}
\description{
Some links used in certain functions are hardcoded, but may change in the future.
For ease of link maintenance, constants will be added to this function.
}
\examples{
get_constant("DEFAULT_EDAV_FOLDER")
}
