% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.main.functions.R
\name{get_ctry_abbrev}
\alias{get_ctry_abbrev}
\title{Get country abbreviations}
\usage{
get_ctry_abbrev(afp_data)
}
\arguments{
\item{afp_data}{\code{tibble} AFP dataset}
}
\value{
\code{tibble} A tibble with the country, abbreviation, and WHO region
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Gets the country abbreviation from the AFP dataset.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry_abbrev <- get_ctry_abbrev(raw_data$afp)
}
}
