% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_diff_cols}
\alias{get_diff_cols}
\title{Get the columns where records differ in a group. Useful for identifying where duplicates differ after
performing a distinct() operation}
\usage{
get_diff_cols(df, id_col)
}
\arguments{
\item{df}{data frame or tibble}

\item{id_col}{\code{string} column used as a unique identifier for records}
}
\value{
tibble showing the columns where duplicates differ
}
\description{
Get the columns where records differ in a group. Useful for identifying where duplicates differ after
performing a distinct() operation
}
