% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_edav_data}
\alias{get_edav_data}
\title{Interactive loading of EDAV data}
\usage{
get_edav_data(path = get_constant("DEFAULT_EDAV_FOLDER"))
}
\arguments{
\item{path}{\code{str} Path to start at initially.}
}
\value{
\code{tibble} Data from the EDAV environment.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is a way to interactively work with files in the EDAV
environment, which is convenient as we don't have to search for files within
Azure Storage Explorer.
}
\examples{
\dontrun{
test <- get_edav_data()
}
}
