% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_edav_pins_board.R
\name{get_edav_pins_board}
\alias{get_edav_pins_board}
\title{Obtain a pins board from EDAV}
\usage{
get_edav_pins_board(
  board_loc = "GID/PEB/SIR/pins_board",
  azcontainer = get_azure_storage_connection()
)
}
\arguments{
\item{board_loc}{\code{str} Location of the pins board in Azure.}

\item{azcontainer}{\code{azcontainer} An Azure container.}
}
\value{
\verb{pins board} An Azure pins board.
}
\description{
This is a convenience function to obtain a pins board in EDAV. After the pins
board is loaded, several functions from the pins package are available to explore
the board. Please see the pins package \href{https://pins.rstudio.com/articles/pins.html}{get started} page.
}
\examples{
\dontrun{
edav_board <- get_edav_pins_board()
}
}
