% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.table.functions.R
\name{get_es_site_age}
\alias{get_es_site_age}
\title{Find established sites}
\usage{
get_es_site_age(es_data, end_date, sample_threshold = 10)
}
\arguments{
\item{es_data}{\code{tibble} ES data}

\item{end_date}{\code{str} End date to anchor analysis to}

\item{sample_threshold}{\code{int} Threshold of sample number to be established.}
}
\value{
\code{tibble} A summary of established ES sites
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Finds established ES sites defined as those that has been active for for
at least 12 months since the end date with at least 10 collections.
}
\keyword{internal}
