% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.stool.ad.01.R
\name{get_incomplete_adm}
\alias{get_incomplete_adm}
\title{Helper function to filter GUIDs that are not consistent across temporal scale}
\usage{
get_incomplete_adm(admin_data, spatial_scale, start_date, end_date)
}
\arguments{
\item{admin_data}{population dataset}

\item{spatial_scale}{"ctry", "prov", or "dist"}

\item{start_date}{start date}

\item{end_date}{end date}
}
\value{
a list of GUIDs not present in the time period
}
\description{
Helper function to filter GUIDs that are not consistent across temporal scale
}
\keyword{internal}
