% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.table.functions.R
\name{get_lab_date_col_missingness}
\alias{get_lab_date_col_missingness}
\title{Get missingness of date variables in the lab dataset}
\usage{
get_lab_date_col_missingness(lab_data, group_by = NULL)
}
\arguments{
\item{lab_data}{\code{tibble} Lab data.}

\item{group_by}{\code{str} A column or a vector of columns to group results by.}
}
\value{
\code{tibble} Summary of missingness of date variables
}
\description{
Obtains the percentage of missingness in date variables within the lab dataset
}
\examples{
\dontrun{
get_lab_date_col_missingness(lab_data)
}
}
