% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.parquet.R
\name{get_partition_cols}
\alias{get_partition_cols}
\title{Gets the column used to partition a column}
\usage{
get_partition_cols(name)
}
\arguments{
\item{name}{\code{str} Name of the column}
}
\value{
\code{chr} A character vector of columns to partition with.
}
\description{
Gets the column used to partition a column
}
\examples{
\dontrun{
get_partition_cols("afp")
}
}
\keyword{internal}
