% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.powerpoint.functions.R
\name{get_ppt_template}
\alias{get_ppt_template}
\title{Get path of the PowerPoint template}
\usage{
get_ppt_template(path = NULL)
}
\arguments{
\item{path}{\code{str} Path to the PowerPoint template. If \code{NULL},
will prompt user to download from the sg-desk-review GitHub repository}
}
\value{
\code{str} Local path of the PowerPoint template.
}
\description{
The desk review PowerPoint template is used to build the desk review slide deck.
The function will either download the template from the sg-desk-reviews GitHub page
or get it locally.
}
\examples{
\dontrun{
get_ppt_template()

# If present locally
template_path <- "C:/Users/ABC1/Desktop/deskreview_template.pptx"
ppt_template <- get_ppt_template(template_path)
}

}
