% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{get_region}
\alias{get_region}
\title{Determines whether lab data is EMRO or AFRO}
\usage{
get_region(country_name = Sys.getenv("DR_COUNTRY"))
}
\arguments{
\item{country_name}{name of the country}
}
\value{
string "EMRO" or "AFRO"
}
\description{
Determines whether lab data is EMRO or AFRO
}
