% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpd.dal.R
\name{get_vpd_data}
\alias{get_vpd_data}
\title{Get vaccine preventable diseases dataset}
\usage{
get_vpd_data(
  vpd_name = NULL,
  variable_name = NULL,
  years = NULL,
  ctry_name = NULL,
  iso3_codes = NULL,
  add_ctry_sf = TRUE,
  add_ctry_pop = TRUE
)
}
\arguments{
\item{vpd_name}{\code{str} A VPD or a list of VPDs. Defaults to \code{NULL}, which
returns the full VPD dataset.}

\item{variable_name}{\code{str} A variable or a list of variables. Defaults to \code{NULL},
which returns a dataset containing all the variable names.}

\item{years}{\code{int} A year or a list of years. Defaults to \code{NULL}, which
returns a dataset containing all the years available.}

\item{ctry_name}{\code{str} A country or a list country names. Defaults to \code{NULL},
which returns all the countries in the dataset.}

\item{iso3_codes}{\code{str} An ISO3 code or a list of ISO3 codes. Defaults to \code{NULL},
which returns all the ISO3 codes in the dataset.}

\item{add_ctry_sf}{\code{logical} Attach the country shapefile? Defaults to \code{TRUE}.}

\item{add_ctry_pop}{\code{logical} Attach the country population data? Defaults to \code{TRUE}.}
}
\value{
\code{list} A list containing the VPD data, and optionally the shapefile
and population data.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Gets the VPD dataset from EDAV.
}
\examples{
\dontrun{
vpd_data <- get_vpd_data()
}
}
