% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpd.dal.R
\name{get_vpd_missingness}
\alias{get_vpd_missingness}
\title{Get years with missing data based on variable name and VPD}
\usage{
get_vpd_missingness(
  vpd_name = NULL,
  variable_name = NULL,
  min_year = 1980,
  max_year = lubridate::year(Sys.Date())
)
}
\arguments{
\item{vpd_name}{\code{str} A VPD name or a list of names.}

\item{variable_name}{\code{str} A variable name or a list of names.}

\item{min_year}{\code{int} Minimum year to analyze. Defaults to \code{1980}.}

\item{max_year}{\code{int} Maximum year to analyze. Defaults to the current year.}
}
\value{
\code{tibble} A summary table of years with missing data for each country
for a particular VPD and variable name.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

For each pair of VPD and variable name, returns the years with missing data
for each country.
}
\examples{
\dontrun{
missing_years <- get_vpd_missingness("Cholera", "cases")
}
}
