% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vpd.dal.R
\name{get_vpd_vars}
\alias{get_vpd_vars}
\title{Obtain available variables for VPD data}
\usage{
get_vpd_vars()
}
\value{
\code{tibble} A tibble with the available variable choices.
}
\description{
\code{lifecycle::badge("experimental")}

Obtains the variable names available for each vaccine preventable disease (VPD)
in the \code{variable} column.
}
\examples{
\dontrun{
vpd_vars <- get_vpd_vars()
}
}
