% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{impute_dist_afp}
\alias{impute_dist_afp}
\title{Imputation algorithm to fill in missing district data from EPID}
\usage{
impute_dist_afp(afp.data)
}
\arguments{
\item{afp.data}{tibble of AFP data}
}
\value{
tibble of AFP data with filled district data
}
\description{
Imputation algorithm to fill in missing district data from EPID
}
