% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{init_dr}
\alias{init_dr}
\title{Set up the folders and load polio data}
\usage{
init_dr(
  country_name,
  start_date = NULL,
  end_date = NULL,
  local_dr_folder = getwd(),
  sg_dr_folder = NULL,
  lab_data_path = NULL,
  iss_data_path = NULL,
  attach_spatial_data = T,
  branch = "main",
  source = F
)
}
\arguments{
\item{country_name}{\code{str} name of the country}

\item{start_date}{\code{str} start date of the desk review}

\item{end_date}{\code{str} end date of the desk review}

\item{local_dr_folder}{folder where the desk review code is located}

\item{sg_dr_folder}{folder where the local git repository is located}

\item{lab_data_path}{location of the lab data}

\item{iss_data_path}{location of the ISS data}

\item{attach_spatial_data}{boolean whether to include spatial data}

\item{branch}{what branch to download the DR functions from GitHub}

\item{source}{whether to source local functions or use sirfunctions}
}
\value{
\code{list} large list containing all dataframe for all polio data
}
\description{
Set up the folders and load polio data
}
