% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.main.functions.R
\name{init_kpi}
\alias{init_kpi}
\title{Initialize the KPI analysis pipeline}
\usage{
init_kpi(path = getwd(), name = NULL, edav = TRUE)
}
\arguments{
\item{path}{\code{str} Path to the folder containing the KPI analysis folders.
Defaults to current working directory.}

\item{name}{\code{str} Name of the KPI analysis folder. If not given any names, the
folder will be named the date the function is ran.}

\item{edav}{\code{logical} Whether to use EDAV to load the raw_data and lab_data files.
Defaults to \code{TRUE}.}
}
\value{
Does not return anything
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Sets up folder structures and environmental variables, as well as download
global polio data.
}
\examples{
\dontrun{
init_kpi(name = "kpi_jan_2024")
}
}
