% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desk.review.functions.R
\name{initiate_research}
\alias{initiate_research}
\title{Set up the folders and load polio data}
\usage{
initiate_research(country_name, start_date, end_date)
}
\arguments{
\item{country_name}{\code{str} name of the country}

\item{start_date}{\code{str} start date of the desk review}

\item{end_date}{\code{str} end date of the desk review}
}
\value{
\code{list} large list containing all dataframe for all polio data
}
\description{
Set up the folders and load polio data
}
