% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desk.review.functions.R
\name{initiate_research_logic}
\alias{initiate_research_logic}
\title{Handles the logic of which files to load in the current R session}
\usage{
initiate_research_logic(data_dir_path, data_size, country_name, dr_data_path)
}
\arguments{
\item{data_dir_path}{\code{str} path to the dataset}

\item{data_size}{\code{str} "small", "medium", or "large"}

\item{country_name}{\code{str} name of the country}

\item{dr_data_path}{\code{str} path to save the dataset to}
}
\value{
\code{list} large list containing polio data
}
\description{
Handles the logic of which files to load in the current R session
}
