% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.iss.functions.R
\name{iss_data_errors}
\alias{iss_data_errors}
\title{Checks for errors in the iss.data.
Currently reports the number of missing priority levels.}
\usage{
iss_data_errors(ctry.data, error_path = Sys.getenv("DR_ERROR_PATH"))
}
\arguments{
\item{ctry.data}{ctry.data with ISS/eSurv data attached}
}
\value{
error message
}
\description{
Checks for errors in the iss.data.
Currently reports the number of missing priority levels.
}
