% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{lab_data_errors}
\alias{lab_data_errors}
\title{Generate a log of potential errors in the lab data}
\usage{
lab_data_errors(
  ctry.data,
  start.date = start_date,
  end.date = end_date,
  error_path = Sys.getenv("DR_ERROR_PATH")
)
}
\arguments{
\item{ctry.data}{RDS object containing country polio data}

\item{start.date}{start date of the desk review}

\item{end.date}{end date of the desk review}

\item{error_path}{file path to the error folder}
}
\description{
Generate a log of potential errors in the lab data
}
