% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{lab_data_errors_region}
\alias{lab_data_errors_region}
\title{Check common errors in the regional lab data.}
\usage{
lab_data_errors_region(
  lab.data,
  ctry_name = Sys.getenv("DR_COUNTRY"),
  start.date,
  end.date,
  error_path = Sys.getenv("DR_ERROR_PATH"),
  ctry.data = lifecycle::badge("deprecated")
)
}
\arguments{
\item{lab.data}{\code{tibble} Regional lab data.}

\item{ctry_name}{\code{str} Name of the country. Defaults to the desk review country.}

\item{start.date}{\code{str} Start date of analysis.}

\item{end.date}{\code{str} End date of analysis.}

\item{error_path}{\code{str} Path to folder to save the error log.}

\item{ctry.data}{\code{list} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Passing lab data
attached to country data is deprecated. Please pass lab data directly to the
lab.data parameter.}
}
\value{
None. It outputs an Excel file locally containing the error log.
}
\description{
Error checking for regional lab data. This is a helper function meant to be used inside
\code{\link[=lab_data_errors]{lab_data_errors()}}.
}
\examples{
\dontrun{
lab_path <- "C:/Users/XRG9/lab_data_region.csv"
ctry.data <- init_dr("algeria", lab_data_path = lab_path)
lab_data_errors_region(ctry.data$lab.data, "2021-01-01", "2023-12-31")
}
}
\keyword{internal}
