% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{lab_data_errors_who}
\alias{lab_data_errors_who}
\title{Checks for common data errors in lab data}
\usage{
lab_data_errors_who(
  ctry.data,
  start.date,
  end.date,
  error_path = Sys.getenv("DR_ERROR_PATH")
)
}
\arguments{
\item{ctry.data}{RDS object containing polio data from a country}

\item{start.date}{start date of the desk review}

\item{end.date}{end date of the desk review}

\item{error_path}{folder to save the log to}
}
\description{
Checks for common data errors in lab data
}
