% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{last_detections_prov}
\alias{last_detections_prov}
\title{Identify last detections at the province level}
\usage{
last_detections_prov(data_p, .start_date, .end_date)
}
\arguments{
\item{data_p}{dataset containing epids with cases}

\item{.start_date}{start date}

\item{.end_date}{end date}
}
\value{
summary table identifying last detections at the province level
}
\description{
Identify last detections at the province level
}
