% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{list_required_files_for_processing}
\alias{list_required_files_for_processing}
\title{Creates the "download table", with paths to files required for recreating static files}
\usage{
list_required_files_for_processing(data_folders_paths, polis_folder, use_edav)
}
\arguments{
\item{data_folders_paths}{\code{list} Output of \code{\link[=check_data_folder]{check_data_folder()}}.}

\item{polis_folder}{\code{str} POLIS folder containing preprocessed data. NOT the subfolder under the data folder.}

\item{use_edav}{\code{logical} Whether to use EDAV or not.}
}
\value{
\code{tibble} Dataset containing paths to required files.
}
\description{
Creates the "download table", with paths to files required for recreating static files
}
\keyword{internal}
