% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{load_clean_prov_sp}
\alias{load_clean_prov_sp}
\title{Download province geographic data}
\usage{
load_clean_prov_sp(
  azcontainer = suppressMessages(get_azure_storage_connection()),
  fp = "GID/PEB/SIR/Data/spatial/global.prov.rds",
  prov_guid = NULL,
  prov_name = NULL,
  ctry_name = NULL,
  end_year = lubridate::year(Sys.Date()),
  st_year = 2000,
  data_only = FALSE,
  type = NULL,
  version = "standard",
  edav = TRUE,
  end.year = lifecycle::deprecated(),
  st.year = lifecycle::deprecated(),
  data.only = lifecycle::deprecated()
)
}
\arguments{
\item{azcontainer}{Azure validated container object}

\item{fp}{\code{str} Location of geodatabase.}

\item{prov_guid}{\verb{str array} Array of all province GUIDS that you want to pull.}

\item{prov_name}{\verb{str array} Array of all province names that you want to pull.}

\item{ctry_name}{\verb{str array} Array of all country names that you want to pull.}

\item{end_year}{\code{int} Last year you want to pull information for. Default is current year.}

\item{st_year}{\code{int} Earlier year of spatial data you want to pull. Default is 2000.}

\item{data_only}{\code{logical} Whether to return a tibble with shapefiles or not. Defaults to \code{FALSE}.}

\item{type}{\code{str} Whether to return a spatial object for every year group. Defaults to \code{NULL}.
\itemize{
\item \code{"long"} Return a dataset for every year group.
\item \code{NULL} Return a dataset only with unique GUIDs and when they were active.
}}

\item{version}{\code{str} Specify whether to return standard shapefiles or new shapefiles
still under evaluation/development. Default is \code{"standard"}.
\itemize{
\item \code{"standard"} Standard shapefiles.
\item \code{"dev"}  New shapefiles still under evaluation/development.
}}

\item{edav}{\code{logical} Load from EDAV? Defaults to \code{TRUE}.}

\item{end.year}{\code{int} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Renamed in favor of
\code{end_year}.}

\item{st.year}{\code{int} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Renamed in favor of
\code{st_year}.}

\item{data.only}{\code{logical} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} Renamed in favor of
\code{data_only}.}
}
\value{
\code{tibble} or \code{sf} Dataframe containing spatial data.
}
\description{
Pulls province shapefiles directly from the geodatabase
}
\examples{
\dontrun{
prov <- load_clean_prov_sp(ctry_name = c("ALGERIA", "NIGERIA"), st_year = 2019)
prov.long <- load_clean_prov_sp(ctry_name = "ALGERIA", st_year = 2019, type = "long")
}
}
