% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desk.review.functions.R
\name{load_data}
\alias{load_data}
\title{Loads .Rds file}
\usage{
load_data(data_dir_path)
}
\arguments{
\item{data_dir_path}{\code{str} data path}
}
\value{
\code{list} large list containing the polio data
}
\description{
Loads .Rds file
}
