% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{load_sharepoint_env}
\alias{load_sharepoint_env}
\title{Load the sharepoint environment}
\usage{
load_sharepoint_env(raw.data)
}
\arguments{
\item{raw.data}{global polio data output by sirfunctions::get_all_polio_data()}
}
\value{
sharepoint path
}
\description{
Load the sharepoint environment
}
\keyword{internal}
