% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.npafp.rate.01.R
\name{npafp_rolling}
\alias{npafp_rolling}
\title{Calculate the NPAFP rate on a rolling basis}
\usage{
npafp_rolling(afp.data, year.pop.data, start_date, end_date, spatial_scale)
}
\arguments{
\item{afp.data}{AFP linelist, either from \code{raw.data} or \code{ctry.data}.}

\item{year.pop.data}{Summary table containing year and pop data.}

\item{start_date}{Start date to calculate the rolling interval for.}

\item{end_date}{End date to calculate the rolling interval for.}

\item{spatial_scale}{Spatial scale. Valid arguments are: \verb{"ctry", "prov", "dist}}
}
\value{
A summary table including NPAFP rates and population data.
}
\description{
Calculates the NPAFP rate on a rolling basis, based on the start and end dates
specified.
}
