% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.npafp.rate.01.R
\name{npafp_year}
\alias{npafp_year}
\title{Calculate NPAFP by year and geographic level}
\usage{
npafp_year(afp.data, pop.data, year.data, spatial_scale)
}
\arguments{
\item{afp.data}{AFP linelist, either from \code{raw.data} or \code{ctry.data}.}

\item{pop.data}{Population data}

\item{year.data}{Summary table. Created from \code{generate_year_data()}.}

\item{spatial_scale}{Spatial scale. Valid arguments are: \verb{"ctry", "prov", "dist}}
}
\value{
A summary table including NPAFP rates and population data.
}
\description{
Helper function for calculating the NPAFP rate based on the geographic level.
This function is used inside \code{f.npafp.rate.01()}.
}
