% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{output_empty_image}
\alias{output_empty_image}
\title{Output an empty image}
\usage{
output_empty_image(output_path, img_title)
}
\arguments{
\item{output_path}{\code{str} Output path.}

\item{img_title}{\code{str} Title of the image.}
}
\value{
\code{NULL} silently upon success.
}
\description{
Output an empty image
}
\keyword{internal}
