% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{prep_npafp_table}
\alias{prep_npafp_table}
\title{Create the *.case.ind tables in desk reviews}
\usage{
prep_npafp_table(npafp.output, afp.all.2, start_date, end_date, spatial.scale)
}
\arguments{
\item{npafp.output}{output of running f.npafp.rate.01}

\item{afp.all.2}{AFP linelist}

\item{spatial.scale}{spatial scale to analyze. Valid values are "ctry", "prov", "dist"}
}
\value{
table to be used in summary tables for NPAFP rates
}
\description{
Create the *.case.ind tables in desk reviews
}
