% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{process_afp_epi_raw_data}
\alias{process_afp_epi_raw_data}
\title{Creates afp.epi dataset}
\usage{
process_afp_epi_raw_data(afp)
}
\arguments{
\item{afp}{\code{tibble} Output of \code{\link[=process_afp_raw_data]{process_afp_raw_data()}}.}
}
\value{
\code{tibble} Summary of AFP cases by year/epi-week per country.
}
\description{
Creates afp.epi dataset
}
\keyword{internal}
