% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{process_afp_raw_data}
\alias{process_afp_raw_data}
\title{Creates the AFP dataset of raw_data}
\usage{
process_afp_raw_data(dl_table, use_edav)
}
\arguments{
\item{dl_table}{\code{tibble} Output of \code{\link[=list_required_files_for_processing]{list_required_files_for_processing()}}.}

\item{use_edav}{\code{logical} Whether to use EDAV or not.}
}
\value{
\code{tibble} AFP dataset.
}
\description{
Creates the AFP dataset of raw_data
}
\keyword{internal}
