% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{process_metadata_raw_data}
\alias{process_metadata_raw_data}
\title{Creates metadata tag}
\usage{
process_metadata_raw_data(
  dl_table,
  raw_data,
  polis_folder,
  core_ready_folder,
  use_edav
)
}
\arguments{
\item{dl_table}{\code{tibble} Output of \code{\link[=list_required_files_for_processing]{list_required_files_for_processing()}}.}

\item{raw_data}{\code{list} Processed data combining all polio data.}

\item{polis_folder}{\code{str} Path to POLIS folder.}

\item{core_ready_folder}{\code{str} Name of the core ready folder.}

\item{use_edav}{\code{logical} Whether to use EDAV or not.}
}
\value{
\code{tibble} Metadata tibble.
}
\description{
Creates metadata tag
}
\keyword{internal}
