% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{process_paralytic_raw_data}
\alias{process_paralytic_raw_data}
\title{Creates paralytics cases dataset}
\usage{
process_paralytic_raw_data(afp)
}
\arguments{
\item{afp}{\code{tibble} Output of \code{\link[=process_afp_raw_data]{process_afp_raw_data()}}.}
}
\value{
\code{tibble} Dataset with paralytic cases only.
}
\description{
Creates paralytics cases dataset
}
\keyword{internal}
