% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{pull_data_from_dl_table}
\alias{pull_data_from_dl_table}
\title{Pull data listed in the download table}
\usage{
pull_data_from_dl_table(dl_table, grepl_pattern, use_edav)
}
\arguments{
\item{dl_table}{\code{tibble} Output of \code{\link[=list_required_files_for_processing]{list_required_files_for_processing()}}.}

\item{grepl_pattern}{\code{str} Pattern to use to filter the \code{dl_table}.}

\item{use_edav}{\code{logical} Whether to use EDAV or not.}
}
\value{
\code{tibble} One of the datasets listed in \code{dl_table}.
}
\description{
Pull data listed in the download table
}
\keyword{internal}
