% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{pull_map_data}
\alias{pull_map_data}
\title{Pull the main map data to build the map}
\usage{
pull_map_data(
  raw.data,
  .vdpv,
  country,
  surv,
  virus_type,
  .start_date,
  .end_date
)
}
\arguments{
\item{raw.data}{global polio data output from sirfunctions::get_all_polio_data()}

\item{.vdpv}{whether to include VDPV in maps}

\item{country}{name of the country or countries}

\item{surv}{surveillance type}

\item{virus_type}{virus type to map}

\item{.start_date}{start date}

\item{.end_date}{end date}
}
\value{
dataset used to build the map
}
\description{
Pull the main map data to build the map
}
