% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{read_excel_from_edav}
\alias{read_excel_from_edav}
\title{Reads an Excel file from EDAV to the R environment}
\usage{
read_excel_from_edav(src, ...)
}
\arguments{
\item{src}{\code{str} Path to the Excel file.}

\item{...}{Additional parameters of \code{\link[readxl:read_excel]{readxl::read_excel()}}.}
}
\value{
\code{tibble} or \code{list} A tibble or a list of tibbles containing data from
the Excel file.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

This function is an extension of the readxl() that adapts to files with
multiple tabs. If there are multiple tabs, each sheet are downloaded into a
named list with the corresponding tab name.
}
\details{
Actually, this function doesn't need to be used on EDAV files. It can work
with local files as well.
}
\keyword{internal}
