% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{recache_raw_data}
\alias{recache_raw_data}
\title{Should the local global polio dataset be recached?}
\usage{
recache_raw_data(analytic_folder, edav, output_format)
}
\arguments{
\item{analytic_folder}{\code{str} Path to the analytics folder.}

\item{edav}{\code{logical} Should we use EDAV? Defaults to TRUE.}

\item{output_format}{\code{str} Output format to load.}
}
\value{
\code{logical} Whether the global polio dataset should be cached.
}
\description{
The function checks the analytics folder for changes in the global polio dataset.
If there is a change between versions in EDAV and locally, returns whether the
it should be recached locally.
}
\details{
If \code{get_all_polio_data()} is used locally, then calling the function does not
really make sense given there is already a local copy of the global polio dataset.
}
\keyword{internal}
