% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{recache_spatial_data}
\alias{recache_spatial_data}
\title{Should the local spatial cache be updated?}
\usage{
recache_spatial_data(analytic_folder, spatial_folder, edav, output_format)
}
\arguments{
\item{analytic_folder}{\code{str} Path to the analytics folder.}

\item{spatial_folder}{\code{str} Path to the spatial folder.}

\item{edav}{\code{logical} Use EDAV?}

\item{output_format}{\code{str} Output format the spatial files are in.}
}
\value{
\code{logical} Should we recache the spatial data or not?
}
\description{
Should the local spatial cache be updated?
}
\keyword{internal}
