% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_polio_data_2.R
\name{reprocess_polio_data}
\alias{reprocess_polio_data}
\title{Reprocess the global polio dataset}
\usage{
reprocess_polio_data(
  data_folder,
  polis_folder,
  core_ready_folder,
  use_edav,
  cache
)
}
\arguments{
\item{data_folder}{\code{str} Path to the data folder.}

\item{polis_folder}{\code{str} Path to the POLIS folder.}

\item{core_ready_folder}{\code{str} Name of the core ready folder.}

\item{use_edav}{\code{logical} Whether to use EDAV or not.}

\item{cache}{\code{logical} Whether to cache the preprocessed data to the data/polis subfolder.}
}
\value{
\code{list} Processed raw data.
}
\description{
Reprocess the global polio dataset
}
\keyword{internal}
