% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.R
\name{send_outlook_email}
\alias{send_outlook_email}
\title{Send email through Outlook}
\usage{
send_outlook_email(title, body, recipient, attachment = NULL)
}
\arguments{
\item{title}{\code{str} Subject of message to be sent.}

\item{body}{\code{str} Long string of body of message to be sent.}

\item{recipient}{\code{str} A semicolon separated list of recipients.}

\item{attachment}{\code{str} Path to local document to be attached to email. Defaults to \code{NULL}.}
}
\value{
Status message whether the operation was a success or an error message.
}
\description{
Function to send an email through Outlook from R.
}
\examples{
\dontrun{
title_message <- "Test"
body_message <- "this is a test"
recipient_list <- c("ab123@email.com")
send_outlook_email(title_message, body_message, recipient_list)
}

}
