% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.R
\name{send_teams_message}
\alias{send_teams_message}
\title{Send a message on Microsoft Teams}
\usage{
send_teams_message(
  msg,
  team_id = "CGH-GID-PEB-SIR",
  channel = "CORE 2.0",
  attach = NULL,
  type = "text"
)
}
\arguments{
\item{msg}{\code{str} Message to be sent.}

\item{team_id}{\code{str} Teams ID. Defaults to \code{"CGH-GID-PEB-SIR"}.}

\item{channel}{\code{str} Channel where message should be sent.}

\item{attach}{\code{str} Local path of files to be attached in message.}

\item{type}{\code{str} Type of message to be sent. Either \code{"text"} or \code{"html"}.}
}
\value{
Status message whether the operation was a success or an error message.
}
\description{
Helper function to send message to validated MS Teams interface.
}
\examples{
\dontrun{
message <- "this is a test"
send_teams_message(message)
}

}
