% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desk.review.functions.R
\name{set_data_size}
\alias{set_data_size}
\title{Determine the size of data to pull}
\usage{
set_data_size(year)
}
\arguments{
\item{year}{\code{int} start year of the desk review}
}
\value{
\code{str} size of the data
}
\description{
Determine the size of data to pull
}
