% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desk.review.functions.R
\name{set_dr_local_folders}
\alias{set_dr_local_folders}
\title{Sets up local folders in the current working directory for desk review
data_path is expected to be ~/country/year}
\usage{
set_dr_local_folders(path)
}
\arguments{
\item{data_path}{\code{str} path where folders are to be created}
}
\description{
Sets up local folders in the current working directory for desk review
data_path is expected to be ~/country/year
}
