% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{set_emergence_colors}
\alias{set_emergence_colors}
\title{Set the emergence colors}
\usage{
set_emergence_colors(
  raw.data,
  country,
  start_date = NULL,
  end_date = NULL,
  get_unassigned = FALSE
)
}
\arguments{
\item{raw.data}{\code{list} Global polio data output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}.}

\item{country}{\code{str} or \code{list} Countries of interest.}

\item{start_date}{\code{str} Start date of the time span to look for emergences. Defaults to 13 months from the end date.}

\item{end_date}{\code{str} End date of the time span to look for emergences Defaults to download date of \code{raw.data}.}

\item{get_unassigned}{\code{bool} Get a list of emergence without a color mapped. This parameter is
useful for ensuring that emergences are all accounted for when making a map.}
}
\value{
\code{list} A named list containing the mapping of emergence and corresponding colors.
}
\description{
Used in conjunction to \code{\link[=generate_adhoc_map]{generate_adhoc_map()}}. The function returns a named
list with emergence names mapped to a color.
}
\examples{
\dontrun{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
emg.cols <- set_emergence_colors(raw.data, "algeria")
}

}
