% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.main.functions.R
\name{set_kpi_global_env}
\alias{set_kpi_global_env}
\title{Sets path to KPI folders}
\usage{
set_kpi_global_env(analysis_path)
}
\arguments{
\item{analysis_path}{\verb{Absolute path to the KPI run folder}}
}
\description{
Sets path to KPI folders
}
\keyword{internal}
