% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{set_zoom_level}
\alias{set_zoom_level}
\title{Adjust the zoom level of the map}
\usage{
set_zoom_level(g1, map_ref, country, m_base_ctry)
}
\arguments{
\item{g1}{\code{ggplot} Original ggplot map.}

\item{map_ref}{\code{ggplot} Reference map.}

\item{country}{\code{str} Country or countries to zoom into.}

\item{m_base_ctry}{\code{sf} Shapefile containing base map of country.}
}
\value{
\code{ggplot} Map with adjusted zoom level, if specified.
}
\description{
This is an internal function changes the zoom level of the map depending on the
user specification.
}
\keyword{internal}
