% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desk.review.functions.R
\name{setup_local_folders}
\alias{setup_local_folders}
\title{Sets up local folders in the current working directory for storing data and metadata}
\usage{
setup_local_folders(data_path, country_name)
}
\arguments{
\item{data_path}{\code{str} path where data is/to be stored}

\item{country_name}{\code{str} name of the country}
}
\value{
\code{str} returns the path containing containing the data set
}
\description{
Sets up local folders in the current working directory for storing data and metadata
}
