% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{sg_priority_map}
\alias{sg_priority_map}
\title{SG Prioritization Map}
\usage{
sg_priority_map(
  ctry_risk_cat = NULL,
  year = lubridate::year(Sys.Date()),
  output_path = Sys.getenv("KPI_FIGURES")
)
}
\arguments{
\item{ctry_risk_cat}{\code{tibble} Risk category for each country.
Defaults to \code{NULL}, which downloads the SG risk category data set from EDAV.}

\item{year}{\code{tibble} Active year for the country shape files.}

\item{output_path}{\code{str} Where to output the figure to. Defaults to the
path to the figures folder set when running \code{\link[=init_kpi]{init_kpi()}}.}
}
\value{
\code{ggplot} A map.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Creates a map showing the SG prioritization map.
}
\examples{
\dontrun{
sg_priority_map(output_path = getwd())
}
}
