% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.pop.functions.R
\name{spatial_validation}
\alias{spatial_validation}
\title{Check for GUIDs that has changed due to redistricting}
\usage{
spatial_validation(pop.data, spatial.scale)
}
\arguments{
\item{pop.data}{population data}

\item{spatial.scale}{spatial scale of the data}
}
\value{
list containing invalid GUIDs
}
\description{
Check for GUIDs that has changed due to redistricting
}
\keyword{internal}
