% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{split_concat_raw_data}
\alias{split_concat_raw_data}
\title{Split or concatenate raw.data by year}
\usage{
split_concat_raw_data(
  action,
  split.years = NULL,
  raw.data.all = NULL,
  raw.data.small.pull = NULL,
  raw.data.medium.pull = NULL,
  raw.data.large.pull = NULL
)
}
\arguments{
\item{action}{\code{str} Can either be to \code{"concat"} or \code{"split"}.}

\item{split.years}{\code{array} A numeric array of years by which data should be split.}

\item{raw.data.all}{\code{list} A list of data objects to be split.}

\item{raw.data.small.pull}{\code{list} A list of data objects to be concatenated. This is
the 'small' dataset, which consists of data from the past 6 years.}

\item{raw.data.medium.pull}{\code{list} A list of data objects to be concatenated. This is
the 'small' dataset, which consists of data from the past 9 years.}

\item{raw.data.large.pull}{\code{list} A list of data objects to be concatenated. This is
the 'small' dataset, which consists of data since 2000.}
}
\value{
\code{list} A list of lists or a single concatenated list.
}
\description{
Split or concatenate raw.data by year
}
\keyword{internal}
