% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.stool.ad.01.R
\name{stool_ad_rolling}
\alias{stool_ad_rolling}
\title{Helper function to generate stool adequacy on a rolling basis}
\usage{
stool_ad_rolling(stool.data, pop.data, start_date, end_date, spatial_scale)
}
\arguments{
\item{stool.data}{AFP dataset containing the adequacy final column}

\item{pop.data}{population data}

\item{start_date}{start date}

\item{end_date}{end date}

\item{spatial_scale}{"ctry", "prov", or "dist".}
}
\value{
summary table containing stool adequacy on a rolling basis
}
\description{
Helper function to generate stool adequacy on a rolling basis
}
\keyword{internal}
