% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{test_EDAV_connection}
\alias{test_EDAV_connection}
\title{Test network connection to the S drive}
\usage{
test_EDAV_connection(
  azcontainer = get_azure_storage_connection(),
  folder = "GID/PEB/SIR/Data",
  return_list = F,
  test_size = 1e+07
)
}
\arguments{
\item{azcontainer}{azure storage container}

\item{folder}{str: Location of download folder in S drive}

\item{return_list}{boolean: return a list of download time estimates, default F}

\item{test_size}{int: byte size of a theoretical folder}
}
\value{
System message with download and update time, potentially list output
}
\description{
Tests upload and download from S drive
}
