% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{test_S_drive_connection}
\alias{test_S_drive_connection}
\title{Test network connection to the S drive}
\usage{
test_S_drive_connection(
  folder = file.path("", "", "cdc.gov", "project", "CGH_GID_Active", "PEB", "SIR",
    "DATA", "Core 2.0", "datafiles_01"),
  return_list = F,
  test_size = 1e+07
)
}
\arguments{
\item{folder}{str: Location of download folder in S drive}

\item{return_list}{boolean: return a list of download time estimates, default F}

\item{test_size}{int: byte size of a theoretical folder}
}
\value{
System message with download and update time, potentially list output
}
\description{
Tests upload and download from S drive
}
