% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.parquet.R
\name{to_wkb_drop_sf}
\alias{to_wkb_drop_sf}
\title{Drop Shape column and convert to binary}
\usage{
to_wkb_drop_sf(sf_data)
}
\arguments{
\item{x}{\code{sf} or \code{data.frame} Geodata.}
}
\value{
\code{tibble} dData without any Shape column.
}
\description{
Drop Shape column and convert to binary
}
\details{
This function was written using the CDC EDAV Chatbot using the model GPT-5.2.
}
\keyword{internal}
