% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_spatial_data.R
\name{update_city_spatial_data}
\alias{update_city_spatial_data}
\title{Update City Spatial data}
\usage{
update_city_spatial_data(edav, city.loc = "Data/spatial/cities.new.rds")
}
\arguments{
\item{edav}{\code{logical} \code{TRUE} or \code{FALSE} depending on if final save location is in Azure.}

\item{city.loc}{\code{str} Location to save city spatial rds.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Function to download latest city data for mapping purposes.
}
