% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{update_polio_data}
\alias{update_polio_data}
\title{Update a local dataset with new data}
\usage{
update_polio_data(local_dataset, overwrite = T)
}
\arguments{
\item{local_dataset}{\code{str} File path to the global polio data RDS file.}

\item{overwrite}{\code{bool} Should the file be overwritten? Default \code{TRUE}.}
}
\value{
None.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Update a local global polio data (raw.data) with new data.
}
\examples{
\dontrun{
local_raw_data <- "C:/Users/ABC1/Desktop/raw.data.rds"
update_polio_data(local_raw_data, overwrite = FALSE)
}

}
