% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.parquet.R
\name{upload_parquet_to_edav}
\alias{upload_parquet_to_edav}
\title{Uploads a local parquet folder to EDAV}
\usage{
upload_parquet_to_edav(src, dest, container = NULL)
}
\arguments{
\item{src}{\code{str} Local path to the parquet folder.}

\item{dest}{\code{str} EDAV endpoint.}

\item{container}{\code{azcontainer} An instance of an Azure container.}
}
\value{
None.
}
\description{
Uploads a folder containing parquet files to EDAV
}
\examples{
\dontrun{
local_dir <- "C:/Users/ABC1/Desktop/parquet_folder"
edav_dir <- "ABC/parquet_folder"
upload_parquet_to_edav(local_dir, edav_dir)
}
}
