% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{add_new_detection}
\alias{add_new_detection}
\title{Identifying new detections}
\usage{
add_new_detection(data_p, country, date_3a, date_3a_a, date_3b)
}
\arguments{
\item{data_p}{map data}

\item{country}{country or a list of countries}

\item{date_3a}{???}

\item{date_3a_a}{???}

\item{date_3b}{???}
}
\value{
a dataset containing new detections
}
\description{
Identifying new detections
}
