% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.table.functions.R
\name{add_rolling_date_info}
\alias{add_rolling_date_info}
\title{Adds rolling date information}
\usage{
add_rolling_date_info(
  df,
  start_date,
  end_date,
  onset_col = "ParalysisOnsetDate"
)
}
\arguments{
\item{df}{\code{tibble} Dataframe to add rolling date info to.}

\item{start_date}{\code{str} Start date of the analysis.}

\item{end_date}{\code{str} End date of the analysis.}

\item{onset_col}{\code{str} Column used for the onset date. Mainly used to calculate
\code{weeks.from.end}.}
}
\value{
\code{tibble} Lab data with rolling date period information added.
}
\description{
Adds rolling date information
}
\keyword{internal}
