% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{build_detection_map}
\alias{build_detection_map}
\title{Build the detection map}
\usage{
build_detection_map(
  m_base_region,
  m_base_prov,
  data_p,
  m_data_prov,
  new_detect,
  virus_type,
  surv_options,
  start_date,
  date_3a,
  download_date,
  emg_cols,
  country,
  labels,
  clean_maps,
  data_r
)
}
\arguments{
\item{m_base_region}{map base for region}

\item{m_base_prov}{map base for provinces}

\item{data_p}{data for map creation}

\item{m_data_prov}{map data for provinces}

\item{new_detect}{whether to include new detections}

\item{virus_type}{virus type to display}

\item{surv_options}{surveillance options}

\item{start_date}{start date}

\item{date_3a}{???}

\item{download_date}{date global polio data was downloaded}

\item{emg_cols}{emergence colors}

\item{country}{country or countries of interest}

\item{labels}{whether to label provinces}

\item{clean_maps}{clean maps}

\item{data_r}{???}
}
\value{
a ggplot map
}
\description{
Build the detection map
}
