% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{check_afp_guid_ctry_data}
\alias{check_afp_guid_ctry_data}
\title{Check GUIDs present in afp.all.2 but not in the pop files}
\usage{
check_afp_guid_ctry_data(ctry.data)
}
\arguments{
\item{ctry.data}{object containing polio data for a country}
}
\value{
list containing errors in province and district GUIDs
}
\description{
Check GUIDs present in afp.all.2 but not in the pop files
}
