% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{check_afp_guid_ctry_data}
\alias{check_afp_guid_ctry_data}
\title{Check GUIDs present in the AFP linelist but not in the pop files}
\usage{
check_afp_guid_ctry_data(ctry.data)
}
\arguments{
\item{ctry.data}{Named \code{list} output of \code{extract_country_data()}, with spatial data attached.}
}
\value{
A \code{list} containing errors in province and district GUIDs.
}
\description{
The function will run a check in the AFP linelist for GUIDs that are not part
of the spatial files. In these instances, typically, unknown GUIDs are part of the
new geodatabase from WHO that get released in the next updated geodatabase. Therefore,
this function should be used only if necessary. For example, in instances where mapping an AFP case into a
a district is critical and the shapefile from \code{extract_country_data()} is not yet updated.
}
\examples{
raw.data <- get_all_polio_data() # must contain spatial data to run the function
ctry.data <- extract_country_data("algeria", raw.data)
error.list <- check_afp_guid_ctry_data(ctry.data)

}
