% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{clean_ctry_data}
\alias{clean_ctry_data}
\title{Cleans and adds additional columns used in the desk reviews}
\usage{
clean_ctry_data(ctry.data, start_date, end_date, es_start_date, es_end_date)
}
\arguments{
\item{ctry.data}{country data RDS object}

\item{start_date}{start date for desk review}

\item{end_date}{end date for desk review}

\item{es_start_date}{start date for environmental surveillance data}

\item{es_end_date}{end date for environmental surveillance data}
}
\value{
cleaned country data RDS object
}
\description{
Cleans and adds additional columns used in the desk reviews
}
