% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{clean_ctry_data}
\alias{clean_ctry_data}
\title{Cleans and adds additional columns used in the desk reviews}
\usage{
clean_ctry_data(ctry.data)
}
\arguments{
\item{ctry.data}{country data RDS object}
}
\value{
cleaned country data RDS object
}
\description{
Cleans and adds additional columns used in the desk reviews
}
