% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.es.functions.R
\name{clean_es_data}
\alias{clean_es_data}
\title{Clean environmental surveillance data}
\usage{
clean_es_data(es.data, dist.shape, ctry.data = lifecycle::deprecated())
}
\arguments{
\item{es.data}{\code{tibble} Environmental surveillance data.}

\item{dist.shape}{\code{sf} District shapefile.}

\item{ctry.data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
\code{list} This parameter has been deprecated in favor of explicitly passing dataframes
into the function. This allows for greater flexibility in the function.}
}
\value{
\code{tibble} Cleaned environmental surveillance data.
}
\description{
The cleaning step will attempt to impute missing site coordinates and create
standardized columns used in the desk review.
}
\examples{
\dontrun{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry.data <- extract_country_data("algeria", raw.data)
ctry.data$es <- clean_es_data(ctry.data$es, ctry.data$dist)
}

}
