% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.es.functions.R
\name{clean_es_data}
\alias{clean_es_data}
\title{Transform ES data cleaning with additional columns}
\usage{
clean_es_data(ctry.data)
}
\arguments{
\item{es.data}{tibble of ES data from ctry.data}
}
\value{
tibble of cleaned ES data
}
\description{
Transform ES data cleaning with additional columns
}
