% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.es.functions.R
\name{clean_es_data}
\alias{clean_es_data}
\title{ES data cleaning}
\usage{
clean_es_data(ctry.data)
}
\arguments{
\item{ctry.data}{\code{list} Polio data at the country level. This is either the output
of \code{\link[=init_dr]{init_dr()}} or \code{\link[=extract_country_data]{extract_country_data()}}.}
}
\value{
\code{tibble} Cleaned ES data.
}
\description{
The cleaning step will attempt to impute missing site coordinates and create
standardized columns used in the desk review.
}
\examples{
raw.data <- get_all_polio_data(attach.spatial.data = F)
ctry.data <- extract_country_data("algeria")
ctry.data$es <- clean_es_data(ctry.data)

}
