% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data}
\alias{clean_lab_data}
\title{Main lab data cleaning function. Automatically detects whether the dataset
came from WHO or the regional office}
\usage{
clean_lab_data(
  ctry.data,
  start.date = start_date,
  end.date = end_date,
  delim = "-",
  lab_locs_path = NULL
)
}
\arguments{
\item{ctry.data}{Rds file containing country polio data with lab.data attached}

\item{start.date}{start date of the desk review}

\item{end.date}{end date of the desk review}

\item{delim}{delimiter for EPIDs. Default is "-".}

\item{lab_locs_path}{location of testing lab locations. Default is NULL. Will download from EDAV, if necessary.}
}
\value{
a tibble containing the cleaned lab data
}
\description{
Main lab data cleaning function. Automatically detects whether the dataset
came from WHO or the regional office
}
