% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data}
\alias{clean_lab_data}
\title{Clean polio lab data}
\usage{
clean_lab_data(lab.data, ctry.data, start_date, end_date)
}
\arguments{
\item{lab.data}{raw lab data}

\item{ctry.data}{country data RDS object}
}
\value{
a tibble containing clean lab data
}
\description{
Clean polio lab data
}
