% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data}
\alias{clean_lab_data}
\title{Clean lab data}
\usage{
clean_lab_data(
  ctry.data,
  start.date = start_date,
  end.date = end_date,
  delim = "-",
  lab_locs_path = NULL
)
}
\arguments{
\item{ctry.data}{\code{list} Large list containing country polio data with lab data attached. Either from
\code{\link[=extract_country_data]{extract_country_data()}} or \code{\link[=init_dr]{init_dr()}}.}

\item{start.date}{\code{str} Start date of analysis.}

\item{end.date}{\code{str} End date of analysis.}

\item{delim}{\code{str} Delimiter for EPIDs. Default is \code{"-"}.}

\item{lab_locs_path}{\code{str} Location of testing lab locations. Default is \code{NULL}. Will download from EDAV, if necessary.}
}
\value{
\code{tibble} Cleaned lab data.
}
\description{
Main lab data cleaning function. Automatically detects whether the dataset
came from WHO or the regional office.
}
\examples{
\dontrun{
lab_path <- "C:/Users/XRG9/lab_data_who.csv"
ctry.data <- init_dr("algeria", lab_data_path = lab_path)
ctry.data$lab.data <- clean_lab_data(ctry.data, "2021-01-01", "2023-12-31")

# Not using the desk review pipeline
raw.data <- get_all_polio_data()
ctry.data <- extract_country_data("algeria")
ctry.data$lab.data <- read_csv(lab_path)
ctry.data$lab.data <- clean_lab_data(ctry.data, "2021-01-01", "2023-12-31")
}
}
