% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data_regional}
\alias{clean_lab_data_regional}
\title{Clean lab data from the regional offices.
Regional data have different columns compared to the WHO lab data. The cleaning
of the regional data is adapted from the lab data cleaning code from the GPSAP
indicator script.}
\usage{
clean_lab_data_regional(
  ctry.data,
  start.date,
  end.date,
  delim = "-",
  lab_locs_path = NULL
)
}
\arguments{
\item{ctry.data}{Rds polio data at the country level with lab data attached}

\item{start.date}{start date of the desk review}

\item{end.date}{end date of the desk review}

\item{delim}{delimiter used for EPIDs. Default is "-".}

\item{lab_locs_path}{path to CSV file containing lab location. Will pull from EDAV if not attached}
}
\value{
tibble of lab data
}
\description{
Clean lab data from the regional offices.
Regional data have different columns compared to the WHO lab data. The cleaning
of the regional data is adapted from the lab data cleaning code from the GPSAP
indicator script.
}
