% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data_regional}
\alias{clean_lab_data_regional}
\title{Clean lab data from the regional offices.}
\usage{
clean_lab_data_regional(
  ctry.data,
  ctry_name = Sys.getenv("DR_COUNTRY"),
  start.date,
  end.date,
  delim = "-",
  lab_locs_path = NULL
)
}
\arguments{
\item{ctry.data}{\code{list} Large list of country polio data with lab data attached. Either from
\code{\link[=extract_country_data]{extract_country_data()}} or \code{\link[=init_dr]{init_dr()}}.}

\item{ctry_name}{\code{str} Name of the country. Defaults to the desk review country.}

\item{start.date}{\code{str} Start date of analysis.}

\item{end.date}{\code{str} End date of analysis.}

\item{delim}{\code{str} Delimiter used for EPIDs. Default is \code{"-"}.}

\item{lab_locs_path}{\code{str} Path to CSV file containing lab location. Will pull from EDAV if not attached.}
}
\value{
\code{tibble} Cleaned lab data.
}
\description{
Regional data have different columns compared to the WHO lab data. The cleaning
of the regional data is adapted from the lab data cleaning code from the GPSAP
indicator script.
}
\examples{
\dontrun{
lab_path <- "C:/Users/XRG9/lab_data_region.csv"
ctry.data <- init_dr("algeria", lab_data_path = lab_path)
ctry.data$lab.data <- clean_lab_data_regional(ctry.data, "2021-01-01", "2023-12-31")

# Not using the desk review pipeline
raw.data <- get_all_polio_data()
ctry.data <- extract_country_data("algeria", raw.data)
ctry.data$lab.data <- read_csv(lab_path)
ctry.data$lab.data <- clean_lab_data_regional(ctry.data, "2021-01-01", "2023-12-31")
}
}
