% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data_regional}
\alias{clean_lab_data_regional}
\title{Clean lab data from the regional offices.
Regional data have different columns compared to the WHO lab data. The cleaning
of the regional data is adapted from the lab data cleaning code from the GPSAP
indicator script.}
\usage{
clean_lab_data_regional(
  lab.data,
  ctry.data,
  region,
  lab.locs,
  start_date,
  end_date,
  delim = "-"
)
}
\arguments{
\item{lab.data}{tibble lab data to be cleaned}

\item{ctry.data}{Rds polio data at the country level}

\item{region}{AFRO or EMRO}

\item{start_date}{start date of the desk review}

\item{end_date}{end date of the desk review}

\item{delim}{delimiter used for EPIDs. Default is "-".}
}
\value{
tibble of lab data
}
\description{
Clean lab data from the regional offices.
Regional data have different columns compared to the WHO lab data. The cleaning
of the regional data is adapted from the lab data cleaning code from the GPSAP
indicator script.
}
