% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data_who}
\alias{clean_lab_data_who}
\title{Clean polio lab data from WHO}
\usage{
clean_lab_data_who(ctry.data, start.date, end.date, delim = "-")
}
\arguments{
\item{ctry.data}{country data RDS object}

\item{start.date}{start date of the desk review}

\item{end.date}{end date of the desk review}

\item{delim}{delimiter used in the EPIDs. Defaults to "-".}
}
\value{
a tibble containing clean lab data
}
\description{
Clean polio lab data from WHO
}
