% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{clean_lab_data_who}
\alias{clean_lab_data_who}
\title{Clean polio lab data from WHO}
\usage{
clean_lab_data_who(ctry.data, start.date, end.date, delim = "-")
}
\arguments{
\item{ctry.data}{\code{list} Large list of country polio data. Output of either
\code{\link[=extract_country_data]{extract_country_data()}} or \code{\link[=init_dr]{init_dr()}}. Ensure that lab data is attached to \code{ctry.data}.}

\item{start.date}{\code{str} Start date of analysis.}

\item{end.date}{\code{str} End date of analysis.}

\item{delim}{\code{str} Delimiter used in the EPIDs. Defaults to \code{"-"}.}
}
\value{
\code{tibble} Cleaned lab data.
}
\description{
Cleans the lab data from WHO. This is used in \code{\link[=clean_lab_data]{clean_lab_data()}}, but can be used on its own.
}
\examples{
\dontrun{
lab_path <- "C:/Users/XRG9/lab_data_who.csv"
ctry.data <- init_dr("algeria", lab_data_path = lab_path)
ctry.data$lab.data <- clean_lab_data_who(ctry.data, "2021-01-01", "2023-12-31")

# Not using the desk review pipeline
raw.data <- get_all_polio_data()
ctry.data <- extract_country_data("algeria", raw.data)
ctry.data$lab.data <- read_csv(lab_path)
ctry.data$lab.data <- clean_lab_data_who(ctry.data, "2021-01-01", "2023-12-31")
}

}
