% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{compress_png}
\alias{compress_png}
\title{Compress PNG files using pngquant}
\usage{
compress_png(img, pngquant_path = NULL, suffix = "")
}
\arguments{
\item{img}{\code{str} File path to the png file.}

\item{pngquant_path}{\code{str} File path to pngquant executable file (pngquant.exe).}

\item{suffix}{\code{str} Optional parameter to add a suffix to the compressed image.}
}
\value{
None. Will output compressed image to the local folder.
}
\description{
Compress PNG files. The software pngquant is required to use this function.
It attempts to reduce the file size of images without major loss in image quality. Files sizes can be reduced
from 30-60\% using this function. The compressed file will be outputted to the same folder as the original image.
}
\examples{
\dontrun{
img_path <- "C:/Users/ABC1/Desktop/pic1.png"
pngquant_path <- "C:/Users/ABC1/Downloads/pngquant.exe"
compress_png(img_path, pngquant_path, "_compressed")
}

}
