% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_60_day_export}
\alias{create_60_day_export}
\title{Output for 60-day follow ups}
\usage{
create_60_day_export(
  cases.need60day,
  country = country_name,
  excel_output_path = Sys.getenv("DR_TABLE_PATH")
)
}
\arguments{
\item{cases.need60day}{table for 60 day follow up}

\item{country}{name of the country}

\item{excel_output_path}{output path}
}
\value{
does not return anything
}
\description{
Output for 60-day follow ups
}
