% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_afp_export}
\alias{create_afp_export}
\title{Exports AFP linelist with adequacy.final2 column}
\usage{
create_afp_export(stool.data, country, excel_output_path)
}
\arguments{
\item{stool.data}{afp data with final adequacy columns}

\item{excel_output_path}{output path of the Excel file}
}
\description{
Exports AFP linelist with adequacy.final2 column
}
