% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cg_expansion_map.R
\name{create_cg_expansion_map}
\alias{create_cg_expansion_map}
\title{Function to create the consequential geography expansion map}
\usage{
create_cg_expansion_map(polio.data, cg)
}
\arguments{
\item{polio.data}{\code{list} SIR polio data rds}

\item{cg}{\code{tibble} table containing data about existing CGs, the dataset
is expected to contain the following headers: \code{type}, \code{label}, \code{ctry}, \code{prov},
\code{dist}, \code{adm_level}. You can download an example dataset using:
sirfunctions_io("read", file_loc = "Data/misc/consequential_geographies.rds")
originating from this CG super region}
}
\value{
\code{ggplot} CG expansion map
\dontrun{

create_cg_expansion_map(
polio.data = get_all_polio_data(),
cg = sirfunctions_io("read", file_loc = "Data/misc/consequential_geographies.rds")
)

}
}
\description{
Function to create the consequential geography expansion map
}
