% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_emergence_group_gif.R
\name{create_emergence_group_gif}
\alias{create_emergence_group_gif}
\title{Generate Emergence Group Movement Gifs}
\usage{
create_emergence_group_gif(
  emergence_group,
  pos,
  dist,
  ctry,
  output_dir,
  include_env = TRUE,
  cumulative = TRUE,
  fps = 2
)
}
\arguments{
\item{emergence_group}{\code{str} Designation of the emergence group to review.}

\item{pos}{\code{tibble} Positives data set. This is \code{raw.data$pos}, which is part of the
output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}.}

\item{dist}{\code{sf} Shapefile of all districts.}

\item{ctry}{\code{sf} Shapefile of all countries.}

\item{output_dir}{\code{str} Directory where gif should be saved.}

\item{include_env}{\code{bool} To include environmental detections in analysis. Defaults to \code{TRUE}.}

\item{cumulative}{\code{bool} To display cases as cumulative. Defaults to \code{TRUE}.}

\item{fps}{\code{int} Frames per second. Defaults to 2.}
}
\value{
GIF written out to location of \code{output_dir}.
}
\description{
Generate the figures and stitch together a GIF to evaluate emergence group
movement over time, generally aggregated as cumulative per month
}
\examples{
\dontrun{

data <- get_all_polio_data(size = "medium")
pos <- data$pos
emergence_group <- "NIE-JIS-1"
dist <- data$global.dist
ctry <- data$global.ctry
include_env <- T
cumulative <- F
output_dir <- getwd()

create_emergence_group_gif(
  emergence_group, pos, dist, ctry, include_env,
  cumulative, output_dir
)
}

}
