% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_npafp_export}
\alias{create_npafp_export}
\title{Exports NPAFP indicator data summary tables}
\usage{
create_npafp_export(
  ctry.case.ind,
  prov.case.ind,
  dist.case.ind,
  excel_output_path = Sys.getenv("DR_TABLE_PATH")
)
}
\arguments{
\item{ctry.case.ind}{\code{tibble} Country NPAFP indicator summary table.}

\item{prov.case.ind}{\code{tibble} Province NPAFP indicator summary table.}

\item{dist.case.ind}{\code{tibble} District NPAFP indicator summary table.}

\item{excel_output_path}{`str Output path of the Excel file.}
}
\value{
None.
}
\description{
The function combines the NPAFP rate summary tables from \code{\link[=f.npafp.rate.01]{f.npafp.rate.01()}}
and exports to an Excel file, with each geographic level on its own tab.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
ctry.case.ind <- f.npafp.rate.01(ctry.data$afp.all.2, ctry.data$ctry.pop,
                                 "2021-01-01", "2023-01-01", "ctry"
                                 )
prov.case.ind <- f.npafp.rate.01(ctry.data$afp.all.2, ctry.data$prov.pop,
                                 "2021-01-01", "2023-01-01", "prov"
                                 )
dist.case.ind <- f.npafp.rate.01(ctry.data$afp.all.2, ctry.data$dist.pop,
                                 "2021-01-01", "2023-01-01", "dist"
                                 )
create_npafp_export(ctry.case.ind, prov.case.ind, dist.case.ind)
}

}
