% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_pop_check_export}
\alias{create_pop_check_export}
\title{Exports file for checking population roll-ups}
\usage{
create_pop_check_export(
  ctry.data,
  country = Sys.getenv("DR_COUNTRY"),
  excel_output_path = Sys.getenv("DR_TABLE_PATH")
)
}
\arguments{
\item{ctry.data}{\code{list} A large list containing polio data for a country. This is the output of either
\code{\link[=init_dr]{init_dr()}} or \code{\link[=extract_country_data]{extract_country_data()}}.}

\item{country}{\code{str} Name of the country.}

\item{excel_output_path}{\code{str} Output path of the Excel file.}
}
\value{
Nothing.
}
\description{
Export the population roll-ups and determine differences between each population counts.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
create_pop_check_export(ctry.data)
}

}
