% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_pop_check_export}
\alias{create_pop_check_export}
\title{Exports file for checking population roll-ups}
\usage{
create_pop_check_export(
  ctry.data,
  country = Sys.getenv("DR_COUNTRY"),
  excel_output_path = Sys.getenv("DR_TABLE_PATH")
)
}
\arguments{
\item{ctry.data}{RDS file containing polio data for a country}

\item{country}{name of the country}

\item{excel_output_path}{output path}
}
\description{
Exports file for checking population roll-ups
}
