% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_pot_comp_clust_export}
\alias{create_pot_comp_clust_export}
\title{Export potentially compatible and compatible summary table}
\usage{
create_pot_comp_clust_export(
  pot.c.clust,
  country = Sys.getenv("DR_COUNTRY"),
  excel_output_path = Sys.getenv("DR_TABLE_PATH")
)
}
\arguments{
\item{pot.c.clust}{\code{tibble} Potentially compatible cluster summary. The output of
\code{\link[=generate_potentially_compatibles_cluster]{generate_potentially_compatibles_cluster()}}.}

\item{country}{\code{str} Name of the country.}

\item{excel_output_path}{\code{str} Output path of where to store the Excel file.}
}
\value{
Nothing.
}
\description{
Exports the output of \code{\link[=generate_potentially_compatibles_cluster]{generate_potentially_compatibles_cluster()}} as an Excel file.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
stool.data <- generate_stool_data(ctry.data$afp.all.2, "good", "inadequate", "2021-01-01", "2023-12-31")
cases.need60day <- generate_60_day_table_data(stool.data, start_date, end_date)
pot.c.clust <- generate_potentially_compatibles_cluster(cases.need60day)
create_pot_comp_clust_export(pot.c.clust)
}

}
