% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.excel.export.functions.R
\name{create_stool_adequacy_export}
\alias{create_stool_adequacy_export}
\title{Export stool adequacy data}
\usage{
create_stool_adequacy_export(
  cstool,
  pstool,
  dstool,
  excel_output_path = Sys.getenv("DR_TABLE_PATH")
)
}
\arguments{
\item{cstool}{stool adequacy at country level}

\item{pstool}{stool adequacy at province level}

\item{dstool}{stool adequacy at district level}

\item{excel_output_path}{output path}
}
\value{
does not return anything
}
\description{
Export stool adequacy data
}
