% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{ctry_data_errors}
\alias{ctry_data_errors}
\title{Checks data quality errors from the country data}
\usage{
ctry_data_errors(ctry.data)
}
\arguments{
\item{ctry.data}{RDS object containing polio country data}
}
\value{
a list containing all the errors that are checked for
}
\description{
Checks data quality errors from the country data
}
