% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{ctry_data_errors}
\alias{ctry_data_errors}
\title{Checks data quality errors from the country data}
\usage{
ctry_data_errors(ctry.data, error_path = Sys.getenv("DR_ERROR_PATH"))
}
\arguments{
\item{ctry.data}{RDS object containing polio country data}

\item{error_path}{path where to store errors in ctry.data}
}
\description{
Checks data quality errors from the country data
}
