% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{duplicate_check}
\alias{duplicate_check}
\title{Assess duplicates in the get_all_polio_data() output}
\usage{
duplicate_check(.raw.data = raw.data)
}
\arguments{
\item{.raw.data}{Named \code{list} output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}}
}
\description{
Checks for duplicate records in AFP, other, SIA, and Virus datasets.
}
\examples{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
raw.data <- duplicate_check(raw.data)
}
