% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{edav_io}
\alias{edav_io}
\title{Helper function to read and write key data}
\usage{
edav_io(
  io,
  default_dir = "GID/PEB/SIR",
  file_loc = NULL,
  obj = NULL,
  azcontainer = suppressMessages(get_azure_storage_connection()),
  force_delete = F
)
}
\arguments{
\item{io}{str: "read", "write", "delete", "exists", "create" or "list"}

\item{default_dir}{str: "GID/PEB/SIR"}

\item{file_loc}{str: location to "read", "write", "exists", "create or "list"}

\item{obj}{default NULL object to be saved}

\item{azcontainer}{azure container object}

\item{force_delete}{boolean: use delete io without validation}
}
\value{
tibble
}
\description{
Helper function read and write key data to EDAV
}
