% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{edav_io}
\alias{edav_io}
\title{Helper function to read and write key data}
\usage{
edav_io(
  io,
  file_loc,
  obj = NULL,
  azcontainer = suppressMessages(get_azure_storage_connection())
)
}
\arguments{
\item{io}{str: "read", "write", "delete" or "list"}

\item{file_loc}{str: location to "read", "write" or "list"}

\item{obj}{default NULL object to be saved}

\item{azcontainer}{azure container object}
}
\value{
tibble
}
\description{
Helper function read and write key data to EDAV
}
