% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{edav_io}
\alias{edav_io}
\title{Helper function to read and write key data to the EDAV environment}
\usage{
edav_io(
  io,
  default_dir = "GID/PEB/SIR",
  file_loc = NULL,
  obj = NULL,
  azcontainer = suppressMessages(get_azure_storage_connection()),
  force_delete = F,
  local_path = NULL,
  ...
)
}
\arguments{
\item{io}{\code{str} The type of operation to perform in EDAV.
\itemize{
\item \code{"read"} Read a file from EDAV, must be an rds, csv, rda, or xls/xlsx file.
\item \code{"write"} Write a file to EDAV, must be an rds, csv, rda, or xls/xlsx file. To
write an Excel file with multiple sheets, pass a named list containing the tibbles
of interest. See examples.
\item \code{"exists.dir"} Returns a boolean after checking to see if a folder exists.
\item \code{"exists.file"}Returns a boolean after checking to see if a file exists.
\item \code{"create"} Creates a folder and all preceding folders.
\item \code{"list"} Returns a tibble with all objects in a folder.
\item \code{"upload"} Moves a file of any type to EDAV.
\item \code{"delete"} Deletes a file.
\item \code{"delete.dir"} Deletes a folder.
}}

\item{default_dir}{\code{str} The default directory in EDAV. \code{"GID/PEB/SIR"} is the default directory
for all SIR data in EDAV. Can be set to \code{NULL} if you provide the full directory path in \code{file_loc}.}

\item{file_loc}{\code{str} Location to "read", "write", "exists.dir", "exists.file", "create" or "list", can include
the information in \code{default_dir} if you set that parameter to \code{NULL}.}

\item{obj}{\code{robj} Object to be saved, needed for \code{"write"}. Defaults to \code{NULL}.}

\item{azcontainer}{Azure container object returned from \code{\link[=get_azure_storage_connection]{get_azure_storage_connection()}}.}

\item{force_delete}{\code{bool} Use delete io without verification in the command line.}

\item{local_path}{\code{str} Local file pathway to upload a file to EDAV. Default is \code{NULL}.
This parameter is only required when passing \code{"upload"} in the \code{io} parameter.}

\item{...}{Optional parameters that work with \code{\link[readr:read_delim]{readr::read_delim()}} or \code{\link[readxl:read_excel]{readxl::read_excel()}}.}
}
\value{
Output dependent on argument passed in the \code{io} parameter.
}
\description{
The function serves as the primary way to interact with the EDAV system from R. It can
read, write, create folders, check whether a file or a folder exists, upload files, and list
all files in a folder.
}
\examples{
\dontrun{
df <- edav_io("read", file_loc = "df1.csv") # read file from EDAV
# Passing parameters that work with read_csv or read_excel, like sheet or skip.
df2 <- edav_io("read", file_loc = "df2.xlsx", sheet = 1, skip = 2)
list_of_df <- list(df_1 = df, df_2 = df)
# Saves df to the test folder in EDAV
edav_io("write", file_loc = "Data/test/df.csv", obj = df)
# Saves list_of_df as an Excel file with multiple sheets.
edav_io("write", file_loc = "Data/test/df.xlsx", obj = list_of_df)
edav_io("exists.dir", "Data/nonexistentfolder") # returns FALSE
edav_io("exists.file", file_loc = "Data/test/df1.csv") # returns TRUE
edav_io("create", "Data/nonexistentfolder") # creates a folder called nonexistentfolder
edav_io("list") # list all files from the default directory
edav_io("upload", file_loc = "Data/test", local_path = "C:/Users/ABC1/Desktop/df2.csv")
}
}
