% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{extract_country_data}
\alias{extract_country_data}
\title{Extract country specific information from raw polio data}
\usage{
extract_country_data(.country, .raw.data = raw.data)
}
\arguments{
\item{.country}{str: a country name of interest}

\item{.raw.data}{list: list of raw data sources}
}
\description{
Extract country specific data from raw data
}
