% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{extract_country_data}
\alias{extract_country_data}
\title{Extract country specific information from raw polio data}
\usage{
extract_country_data(.country, .raw.data = raw.data)
}
\arguments{
\item{.country}{\code{str} Country name of interest. Case insensitive.}

\item{.raw.data}{\code{list} Output of \code{\link[=get_all_polio_data]{get_all_polio_data()}}.}
}
\value{
Named \code{list} with country specific datasets.
}
\description{
Filters country specific data from the CDC generated \code{raw.data} object from \code{\link[=get_all_polio_data]{get_all_polio_data()}}.
}
\examples{
\dontrun{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry.data <- extract_country_data("nigeria", raw.data)
}

}
