% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.color.schemes.R
\name{f.color.schemes}
\alias{f.color.schemes}
\title{Utility function for colors}
\usage{
f.color.schemes(type)
}
\arguments{
\item{type}{str: type of colors we can return}
}
\value{
list: color sets
}
\description{
Utility function to return SIR colors
}
