% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.color.schemes.R
\name{f.color.schemes}
\alias{f.color.schemes}
\title{Utility function for colors}
\usage{
f.color.schemes(type)
}
\arguments{
\item{type}{str: type of colors we can return. Accepted values include:
'epicurve', 'para.case', 'afp.prov', 'afp.dist', 'pot.comp', 'silence',
'silence.v2', 'cases', 'es'}
}
\value{
list: color sets
}
\description{
Utility function to return SIR colors
}
