% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.color.schemes.R
\name{f.color.schemes}
\alias{f.color.schemes}
\title{Utility function for colors}
\usage{
f.color.schemes(type)
}
\arguments{
\item{type}{\code{str} Type of colors we can return. Accepted values include:
\itemize{
\item \code{"epicurve"} Mapped to different \code{cdc.classification.all2} values.
\item \code{"para.case"} A subset of \code{"epicurve"} representing paralytic cases.
\item \code{"afp.prov"} Mapped for case counts at the province level.
\item \code{"afp.dist"} Mapped for case counts at the province level.
\item \code{"pot.comp"} Colors for categories of compatibles and potentially compatibles.
\item \code{"silence"} Colors to use to map silent populations.
\item \code{"silence.v2"} Colors to use to map silent populations.
\item \code{"cases"} Values to map case type.
\item \code{"es"} Values used in ES data.
\item \code{"mapval"} Values used for creating maps with percentages.
\item \code{"timeliness.col.vars"} Mapping intervals  used for lab timeliness intervals graphs.
\item \code{"emergence.groups"} Standard emergence group colors. Used primarily
with \code{\link[=generate_adhoc_map]{generate_adhoc_map()}}.
\item \code{"es.vaccine.types"} Default vaccine types. Used primarily with
\code{\link[=generate_es_site_det]{generate_es_site_det()}}.
\item \code{"es.detections"} Default detection types. Used primarily with
\code{\link[=generate_es_site_det]{generate_es_site_det()}}.
\item \code{"vpd.critical.ctry"}: Priority VPD countries
\item \code{"vpd.binary.ctry"}: Priority and non-priority VPD countries binary designation.
\item \code{"vpd.colors"}: VPD colors.
}}
}
\value{
Named \code{list} with color sets.
}
\description{
Utility function to return SIR color schemes used in various graphs and visualizations.
}
\examples{
color_list <- f.color.schemes("epicurve")

}
