% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.ev.rate.01.R
\name{f.ev.rate.01}
\alias{f.ev.rate.01}
\title{Calculate EV detection rate function}
\usage{
f.ev.rate.01(es.data, start.date, end.date)
}
\arguments{
\item{es.data}{\code{tibble} ES data which includes site name (site.name),
country (ADM0_NAME),
date of collection (collect.date), and a binary ev detection variable (ev.detect)
that indicates absence/presence (0, 1) of enterovius in an ES sample. This is
\code{ctry.data$es} of \code{\link[=extract_country_data]{extract_country_data()}} or \code{\link[=init_dr]{init_dr()}}, or \code{raw.data$es} of
\code{\link[=get_all_polio_data]{get_all_polio_data()}}.}

\item{start.date}{\code{str} Date in the format of \code{"YYYY-MM-DD"}.}

\item{end.date}{\code{str} Date in the format of \code{"YYYY-MM-DD"}.}
}
\value{
\code{tibble} Long format dataframe including site specific EV detection rates.
}
\description{
Function to calculate the EV detection rate in sites from POLIS.
}
\examples{
\dontrun{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry.data <- extract_country_data("algeria", raw.data)
ev_rates <- f.ev.rate.01(ctry.data$es, "2021-01-01", "2023-12-31")
}

}
