% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.expand.bbox.R
\name{f.expand.bbox}
\alias{f.expand.bbox}
\title{Expand bounding box}
\usage{
f.expand.bbox(bbox, X, Y, X2 = X, Y2 = Y, crs_out = 4326)
}
\arguments{
\item{bbox}{\verb{num vector} A bounding box generated by \code{sf::st_bbox()} or named numeric \code{vector} with xmin, ymin, xmax, ymax.}

\item{X}{\code{numeric} The distance in meters that we want to expand the bounding box by in the X direction}

\item{Y}{\code{numeric} The distance in meters that we want to expand the bounding box by in the Y direction}

\item{X2}{\code{numeric} If specified, the meters in the Easterly direction and meters X becomes meters in the Westerly direction.}

\item{Y2}{\code{numeric} If specified, the meters to the South. meters Y becomes meters to the North.}

\item{crs_out}{\code{int} EPSG coordinate system to return the bounding box in. Defaults to 4326 (lat/lng).}
}
\value{
A named numeric vector of bbox extents. This can be converted into an \code{sf} object using \code{sf::st_as_sfc()}.
}
\description{
(\emph{Sourced from} https://rdrr.io/github/Chrisjb/basemapR/src/R/expand_bbox.R)
A function to take a bounding box (generated using st_bbox) and expand it by x meters
in the X direction and y meters in the Y direction.
}
\examples{
ctry.shape <- load_clean_ctry_sp(ctry_name = "ALGERIA", st.year = 2019)
ctry.bbox <- sf::st_bbox(ctry.shape)
bbox_2 <- f.expand.bbox(ctry.bbox, 4, 4)

}
