% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.metadata.tag.R
\name{f.metadata.tag}
\alias{f.metadata.tag}
\title{Function to add metadata tags to figures and tables}
\usage{
f.metadata.tag(
  object,
  time_tag = ifelse(exists("raw.data"), raw.data$metadata$download_time, NA)
)
}
\arguments{
\item{object}{ggplot or flex table object: the figure or table to add metadata to}

\item{time_tag}{str: a date and time string, defaults to raw.data$metadata$download_time
and will throw an error if "raw.data" doesn't exist in your environment or is called something else}
}
\description{
add metadata tags to figures and tables
}
