% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.metadata.tag.R
\name{f.metadata.tag}
\alias{f.metadata.tag}
\title{Function to add metadata tags to figures and tables}
\usage{
f.metadata.tag(object, raw_data = NULL, time_tag = NULL)
}
\arguments{
\item{object}{ggplot or flextable object: the figure or table to add metadata to}

\item{raw_data}{Rds object: polio raw data or country data}

\item{time_tag}{str: a date and time string, defaults to raw.data$metadata$download_time}
}
\value{
a \code{ggplot} or \code{flextable} object with metadata added
}
\description{
add metadata tags to figures and tables
}
