% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.npafp.rate.01.R
\name{f.npafp.rate.01}
\alias{f.npafp.rate.01}
\title{Calculate Non-Polio AFP Rate Function}
\usage{
f.npafp.rate.01(
  afp.data,
  pop.data,
  start.date,
  end.date,
  spatial.scale,
  pending = T,
  rolling = F
)
}
\arguments{
\item{afp.data}{tibble: AFP data which includes GUID at a given spatial scale
formated as "adm{0,1,2}guid, onset date as "date" and cdc.classification.all2 which includes
c("NPAFP", "PENDING", "LAB PENDING")}

\item{pop.data}{tibble: Under 15 population data by a given spatial scale including
"year", "adm{0,1,2}guid, "u15pop", and "{ctry/prov/dist}" as appropriate}

\item{start.date}{chr: "YYYY-MM-DD"}

\item{end.date}{chr: "YYYY-MM-DD"}

\item{spatial.scale}{chr: "prov" or "dist" or "ctry"}

\item{pending}{boolean: default TRUE}

\item{rolling}{boolean: default FALSE}
}
\value{
tibble
}
\description{
Calculate the NPAFP rate from POLIS data
}
\examples{
f.npafp.rate.01(
afp.data = data$afp.data,
pop.data = data$pop.data.prov,
start.date = "2019-04-01",
end.date = "2020-09-01",
spatial.scale = "prov",
pending = T,
rolling = F)
}
