% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.plot.looks.R
\name{f.plot.looks}
\alias{f.plot.looks}
\title{Function to return specific plot looks}
\usage{
f.plot.looks(type)
}
\arguments{
\item{type}{String describing a value to be returned. Accepted values include:
'02', 'epicurve', 'geomtile', 'gpln_type1', 'gpln_type2'}
}
\value{
ggplot theme for addition into plot
}
\description{
Function to collate and return plot looks
}
