% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.plot.looks.R
\name{f.plot.looks}
\alias{f.plot.looks}
\title{Set plot looks}
\usage{
f.plot.looks(type)
}
\arguments{
\item{type}{\code{str} Type of graph format. Accepted values include:
\itemize{
\item \code{"02"}
\item \code{"epicurve"}
\item \code{"geomtile"}
\item \code{"gpln_type1"}
\item \code{"gpln_type2"}
}}
}
\value{
\verb{ggplot2 theme obj} A theme object that can be added into an existing plot.
}
\description{
The function serves to collate and return plot looks. Depending on the parameter,
specific values in a \code{ggplot2} theme object will be returned.
}
\examples{
epicurve_looks <- f.plot.looks("epicurve")
df <- datasets::iris
p1 <- ggplot2::ggplot() +
  ggplot2::geom_col(data = df, ggplot2::aes(x = Sepal.Length, y = Sepal.Width))
p2 <- p1 + epicurve_looks

}
