% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draft_stool_ad_function.R, R/f.stool.ad.01.R
\name{f.stool.ad.01}
\alias{f.stool.ad.01}
\title{Calculate Stool Adequacy Function}
\usage{
f.stool.ad.01(
  afp.data,
  admin.data,
  start.date,
  end.date,
  spatial.scale,
  missing = "good",
  bad.data = "inadequate",
  rolling = F
)

f.stool.ad.01(
  afp.data,
  admin.data,
  start.date,
  end.date,
  spatial.scale,
  missing = "good",
  bad.data = "inadequate",
  rolling = F
)
}
\arguments{
\item{afp.data}{tibble: AFP data which includes GUID at a given spatial scale
formated as "adm{0,1,2}guid, onset date as "date" and cdc.classification.all2
which includes "NOT-AFP"}

\item{start.date}{chr: "YYYY-MM-DD"}

\item{end.date}{chr: "YYYY-MM-DD"}

\item{spatial.scale}{chr: "prov" or "dist" or "ctry"}

\item{missing}{chr: "good" or "bad" or "missing"; default "good"}

\item{bad.data}{chr: "remove" or "inadequate" or "adequate"; default "inadequate"}

\item{ctry.data}{tibble: Full list of country administrative units by a given
spatial scale including "year", "adm{0,1,2}guid, and "{ctry/prov/dist}"
as appropriate}

\item{prov.dist}{chr: "prov" or "dist" or "ctry"}
}
\value{
tibble

tibble
}
\description{
creates an adequacy variable - 'missing' parameter defines how
missing data is treated - "Good" classifies missing data as good quality (POLIS method)
"Bad" classifies all missing as bad quality, "missing" excludes missing from
the calculations

creates an adequacy variable - 'missing' parameter defines how
missing data is treated - "Good" classifies missing data as good quality
(POLIS method)
"Bad" classifies all missing as bad quality, "missing" excludes missing from
the calculations
}
\examples{
f.stool.ad.01(
afp.data = data$afp.data,
admin.data = data$pop.data.prov,
start.date = "2019-04-01",
end.date = "2020-09-01",
spatial.scale = "prov",
missing = "good",
bad.data = "inadequate",
rolling = F
)
}
