% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.stool.ad.01.R
\name{f.stool.ad.01}
\alias{f.stool.ad.01}
\title{Calculate percent stool adequacy}
\usage{
f.stool.ad.01(
  afp.data,
  admin.data,
  start.date,
  end.date,
  spatial.scale,
  missing = "good",
  bad.data = "inadequate",
  rolling = F,
  sp_continuity_validation = T
)
}
\arguments{
\item{afp.data}{\code{tibble} AFP data which includes GUID at a given spatial scale
formatted as \verb{adm(0,1,2)guid}, onset date as \code{date} and \code{cdc.classification.all2}
which includes "NOT-AFP".}

\item{admin.data}{\code{tibble} Full list of country administrative units by a given
spatial scale including \code{year}, \verb{adm(0,1,2)guid}, and \code{ctry/prov/dist} (as appropriate).}

\item{start.date}{\code{str} Starting date for analysis formatted as "YYYY-MM-DD".}

\item{end.date}{\code{str} Ending date for analysis as "YYYY-MM-DD".}

\item{spatial.scale}{\code{str} Geographic level to group analysis on.
\itemize{
\item \code{"prov"} Province level.
\item \code{"dist"} District level.
\item \code{"ctry"} Country level.
}}

\item{missing}{\code{str} How to treat missing data. Valid values are: \verb{"good", "bad", "remove"}. Defaults to \code{"good"}.
When calculating the \code{adequacy.final} column:
\itemize{
\item \code{"good"} uses \code{adequacy.03}
\item \code{"bad"} uses \code{adequacy.01}
\item \code{"exclude"} uses \code{adequacy.02}
}}

\item{bad.data}{\code{str} How to  treat bad data. Valid values are:\verb{"remove", "inadequate"}. Defaults to \code{"inadequate"}.
\code{"inadequate"} treats samples with bad data as inadequate.}

\item{rolling}{\code{bool} Should data be analyzed on a rolling bases? Defaults to \code{FALSE}.}

\item{sp_continuity_validation}{\code{bool} Should GUIDs not present in all years of the dataset be excluded? Default \code{TRUE}.}
}
\value{
\code{tibble} Long format stool adequacy evaluations.
}
\description{
Creates a summary table of stool adequacy. The \code{missing} parameter defines how
missing data is treated. \code{"good"} classifies missing data as good quality
(POLIS method). \code{"bad"} classifies all missing as bad quality. \code{"missing"}
excludes missing from the calculations.
}
\examples{
raw.data <- get_all_polio_data()
stool.ads <- f.stool.ad.01(raw.data$afp, raw.data$ctry.pop,
                           "2021-01-01", "2023-12-31",
                           "ctry", sp_continuity_validation = FALSE)

}
