% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.timely.01.R
\name{f.timely.01}
\alias{f.timely.01}
\title{Calculate stool collection timeliness using only AFP data (DRAFT)}
\usage{
f.timely.01(
  afp.data,
  admin.data,
  start.date,
  end.date,
  spatial.scale,
  intervals.manual = F
)
}
\arguments{
\item{afp.data}{\code{tibble} AFP data which includes GUID at a given spatial scale
formated as \verb{adm(0,1,2)guid}, onset date as \code{date}.}

\item{admin.data}{\code{tibble} Full list of country administrative units by a given
spatial scale including \code{year}, \verb{adm(0,1,2)guid}, and \code{ctry/prov/dist}
as appropriate.}

\item{start.date}{\code{str} Start date of the analysis formatted as \code{"YYYY-MM-DD"}.}

\item{end.date}{\code{str} End date of the analysis formatted as \code{"YYYY-MM-DD"}.}

\item{spatial.scale}{\code{str} Spatial scale to group analysis by. Valid values are:
-\code{"prov"} Province level.
\itemize{
\item \code{"dist"} District level.
\item \code{"ctry"} Country level.
}}

\item{intervals.manual}{\code{logical} Should user input their own timeliness interval
requirements? Default is \code{FALSE}. This is only required if timeliness column such as
\code{noti.7d.on}, \code{inv.2d.noti} are not already calculated. This draft function will currently fail if
this parameter is set to \code{TRUE}.}
}
\value{
\code{tibble} A summary table of timeliness of stool collection.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Creates a table for timeliness by geographic unit including the
number of timely stool samples for each interval and percent timeliness. Currently,
the function will only work on \code{ctry.data} (output of \code{\link[=extract_country_data]{extract_country_data()}}).
}
\details{
This function is no longer maintained as it is not used in any of the analytic pipelines.
An equivalent function is \code{\link[=generate_int_data]{generate_int_data()}}, which expands and simplifies this
function by also being able to take lab data to calculate lab timeliness intervals.
}
\examples{
\dontrun{
raw.data <- get_all_polio_data()
ctry.data <- extract_country_data("algeria", raw.data)
stool.summary <- f.timely.01(
  ctry.data$afp.all.2, ctry.data$ctry.pop,
  lubridate::as_date("2021-01-01"),
  lubridate::as_date("2023-12-31"),
  "ctry"
)
}

}
\keyword{internal}
