% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.timely.detection.01.R
\name{f.timely.detection.01}
\alias{f.timely.detection.01}
\title{Function to calculate timeliness of detection}
\usage{
f.timely.detection.01(
  afp.data,
  es.data,
  ctryseq.data,
  start.date,
  end.date,
  rolling = F
)
}
\arguments{
\item{afp.data}{\code{tibble} AFP data which includes classification of AFP cases with onset date and date of notification to HQ.}

\item{es.data}{\code{tibble} ES data which includes classification of samples with collection date and date of notification to HQ.}

\item{ctryseq.data}{\code{tibble} A table consisting of the following columns for each country:
\itemize{
\item With sequencing capacity within or outside of the country
\item Country (\code{ADM0_NAME})
\item Classification of AFP cases & ES samples
\item Onset date of AFP cases and collection date of ES samples
\item Date of notification to HQ (\code{date.notification.to.hq})
}

This table is the output of \code{\link[=get_lab_locs]{get_lab_locs()}}.}

\item{start.date}{\code{str} Start date for evaluation with format \code{"YYYY-MM-DD"}.}

\item{end.date}{\code{str} End date for evaluation with format \code{"YYYY-MM-DD"}.}

\item{rolling}{\code{bool} Should timeliness be calculated in a rolling basis? Default \code{FALSE}.}
}
\value{
\code{list} A list with two \code{tibble}s with global and sub-global
AFP / ES detection timeliness evaluation.
}
\description{
Calculates the overall timeliness of detection in AFP & ES POLIS data.
}
\examples{
raw.data <- get_all_polio_data()
ctry.data <- extract_country_data("algeria", raw.data)
ctry.seq <- get_lab_locs()
global.summary <- f.timely.detection.01(raw.data$afp, raw.data$es, ctry.seq,
                                        "2021-01-01", "2023-12-31")
ctry.summary <- f.timely.detection.01(ctry.data$afp.all.2, ctry.data$es, ctry.seq,
                                      "2021-01-01", "2023-12-31")

}
