% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.timely.detection.01.R
\name{f.timely.detection.01}
\alias{f.timely.detection.01}
\title{Function to calculate timeliness of detection}
\usage{
f.timely.detection.01(
  afp.data,
  es.data,
  ctryseq.data,
  start.date,
  end.date,
  rolling = F
)
}
\arguments{
\item{afp.data}{tibble: AFP data which includes classification of AFP cases with onset date and date of notification to HQ,}

\item{es.data}{tibble: ES data which includes classification of samples with collection date and date of notification to HQ,}

\item{ctryseq.data}{tibble: countries with sequencing capacity within or outside of the country,
country (ADM0_NAME),
classification of AFP cases & ES samples
onset date of AFP cases and collection date of ES samples
date of notification to HQ (date.notification.to.hq),}

\item{start.date}{chr: "YYYY-MM-DD" - start date for evaluation}

\item{end.date}{chr: "YYYY-MM-DD" - end date for evaluation}

\item{rolling}{boolean: default FALSE}
}
\value{
list of two tibbles with global and sub-global AFP / ES detection timeliness evaluation
}
\description{
Function to calculate the overall timeliness of detection in AFP & ES POLIS data
}
