% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.timely.detection.01.R
\name{f.timely.detection.01}
\alias{f.timely.detection.01}
\title{Function to calculate timeliness of detection}
\usage{
f.timely.detection.01(
  afp.data,
  es.data,
  ctryseq.data,
  start.date,
  end.date,
  rolling = F
)
}
\arguments{
\item{afp.data}{\code{tibble} AFP data which includes classification of AFP cases with onset date and date of notification to HQ.}

\item{es.data}{\code{tibble} ES data which includes classification of samples with collection date and date of notification to HQ.}

\item{ctryseq.data}{\code{tibble} A table consisting of the following columns for each country:
\itemize{
\item With sequencing capacity within or outside of the country
\item Country (\code{ADM0_NAME})
\item Classification of AFP cases & ES samples
\item Onset date of AFP cases and collection date of ES samples
\item Date of notification to HQ (\code{date.notification.to.hq})
}}

\item{start.date}{\code{str} Start date for evaluation with format "YYYY-MM-DD".}

\item{end.date}{\code{str} End date for evaluation with format "YYYY-MM-DD".}

\item{rolling}{\code{bool} Should timeliness be calculated in a rolling basis? Default \code{FALSE}.}
}
\value{
\code{list} A list with two \code{tibble} with global and sub-global AFP / ES detection timeliness evaluation.
}
\description{
Calculates the overall timeliness of detection in AFP & ES POLIS data.
}
