% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.table.functions.R
\name{f.year.roll}
\alias{f.year.roll}
\title{Add rolling year label for each week number}
\usage{
f.year.roll(wk)
}
\arguments{
\item{wk}{\code{numeric} Week number.}
}
\value{
\code{str} The year the week falls into.
}
\description{
Add rolling year label for each week number
}
\examples{
f.year.roll(32)
}
\keyword{internal}
