% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{f.yrs.01}
\alias{f.yrs.01}
\title{Utility function to fix years}
\usage{
f.yrs.01(df, yrs)
}
\arguments{
\item{df}{tibble: dataframe to be used for fixing years}

\item{yrs}{numeric array: array of years to fix data}
}
\value{
long tibble
}
\description{
Utility function to fix years
}
