% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_cg_expansion_map.R
\name{flag_cg_positives}
\alias{flag_cg_positives}
\title{Function to flag positives data from identified consequential geographies}
\usage{
flag_cg_positives(cg_super_regions, pos, start.year = 2016)
}
\arguments{
\item{cg_super_regions}{\code{sp} spatial object of all CG super regions and a
flag for all their specific GUIDs}

\item{pos}{\code{tibble} the positives file from \code{polio.data}}

\item{start.year}{\code{int} the the earliest year for analysis, defaults to 2016}
}
\value{
\code{sp} all cg related positives flagged and ready for mapping
}
\description{
Function to flag positives data from identified consequential geographies
}
\examples{
\dontrun{
flag_cg_positives(cg_super_regions, pos)
}
}
