% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desk.review.functions.R
\name{freeze_dr_data}
\alias{freeze_dr_data}
\title{Freeze desk review data to the desk review folder in EDAV}
\usage{
freeze_dr_data(rds_obj, file_name)
}
\arguments{
\item{rds_obj}{Rds object loaded in R}

\item{file_name}{name given to the Rds object}
}
\value{
A status message
}
\description{
Freeze desk review data to the desk review folder in EDAV
}
