% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_60_day_tab}
\alias{generate_60_day_tab}
\title{60-day follow up table}
\usage{
generate_60_day_tab(cases.need60day)
}
\arguments{
\item{cases.need60day}{\code{tibble} Summary table containing those that need 60 day follow-up.
Output of \code{\link[=generate_60_day_table_data]{generate_60_day_table_data()}}.}
}
\value{
\code{flextable} A summary of cases requiring 60-day followups per year.
}
\description{
Generates a table summarizing the number of inadequate cases that need follow up.
}
\examples{
\dontrun{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry.data <- extract_country_data("algeria", raw.data)
stool.data <- generate_stool_data(ctry.data$afp.all.2, "good", "inadequate",
                                  "2021-01-01", "2023-12-31"
                                  )
cases.need60day <- generate_60_day_table_data(stool.data,
                                  "2021-01-01", "2023-12-31"
                                  )
generate_60_day_tab(cases.need60day)
}

}
