% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_60_day_table_data}
\alias{generate_60_day_table_data}
\title{Generate summary table for those requiring 60 day follow up}
\usage{
generate_60_day_table_data(stool.data, start_date, end_date)
}
\arguments{
\item{stool.data}{\code{tibble} AFP data with stool adequacy columns. This is the output of
\code{\link[=generate_stool_data]{generate_stool_data()}}.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}
}
\value{
\code{tibble} A summary table for those requiring 60 day follow up.
}
\description{
The 60-day table highlights the number of cases per year that need 60 day follow up. It
summarizes the number of cases due for follow up, those with recorded follow ups, number missing
follow ups,and compatible cases.
}
\examples{
raw.data <- get_all_polio_data(attach.spatial.data = F)
ctry.data <- extract_country_data("algeria", raw.data)
stool.data <- generate_stool_data(ctry.data$afp.all.2, "good", "inadequate", "2021-01-01", "2023-12-31")
table60.days <- generate_60_day_table_data(stool.data, "2021-01-01", "2023-12-31")

}
