% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.stool.ad.01.R
\name{generate_ad_final_col}
\alias{generate_ad_final_col}
\title{Helper function to add the adequacy.final column}
\usage{
generate_ad_final_col(afp.data)
}
\arguments{
\item{afp.data}{AFP dataset}
}
\value{
AFP dataset with the adequacy final column
}
\description{
Helper function to add the adequacy.final column
}
\examples{
\dontrun{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
stool.data <- generate_ad_final_col(raw.data$afp)
}

}
