% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_afp_by_month}
\alias{generate_afp_by_month}
\title{Generate AFP case counts by month}
\usage{
generate_afp_by_month(afp.data, start_date, end_date)
}
\arguments{
\item{afp.data}{tibble containing AFP data (afp.all.2)}
}
\value{
tibble containing summary of AFP data
}
\description{
Generate AFP case counts by month
}
