% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_afp_by_month_summary}
\alias{generate_afp_by_month_summary}
\title{Generate AFP case count summary}
\usage{
generate_afp_by_month_summary(
  afp_data,
  start_date,
  end_date,
  by,
  pop_data = NULL,
  ctry.data = lifecycle::deprecated()
)
}
\arguments{
\item{afp_data}{\code{tibble} AFP dataset.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}

\item{by}{\code{str} How to group the data by. Either \code{"prov"}, \code{"dist"}, or \code{"year"}.}

\item{pop_data}{\code{tibble} Population dataset.}

\item{ctry.data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}} \code{ctry.data} is no longer supported;
the function will explicitly ask for the AFP dataset instead of accessing it from a list.}
}
\value{
\code{tibble} Summary table of AFP cases by month and another grouping variable.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}

Summarize AFP case counts by month and another grouping variable.
}
\examples{
\dontrun{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry.data <- extract_country_data("algeria", raw.data)
afp.by.month <- generate_afp_by_month_summary(
  raw.data$afp, "2021-01-01", "2023-12-31", "ctry",
  raw.data$ctry.pop
)
}

}
