% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_afp_case_map}
\alias{generate_afp_case_map}
\title{Generate a map of AFP cases}
\usage{
generate_afp_case_map(
  ctry.data,
  ctry.shape,
  prov.shape,
  start_date,
  end_date = lubridate::today(),
  output_path = Sys.getenv("DR_FIGURE_PATH")
)
}
\arguments{
\item{ctry.data}{RDS object containing polio country data}

\item{ctry.shape}{recent country shapefile}

\item{prov.shape}{recent province shapefile}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}

\item{output_path}{where to save the figure}
}
\value{
ggplot object
}
\description{
Generate a map of AFP cases
}
