% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_afp_case_map}
\alias{generate_afp_case_map}
\title{Generate a map of AFP cases}
\usage{
generate_afp_case_map(
  afp.all,
  ctry.shape,
  prov.shape,
  start_date,
  end_date = lubridate::today(),
  output_path = Sys.getenv("DR_FIGURE_PATH")
)
}
\arguments{
\item{afp.all}{AFP linelist containing point geometry}

\item{ctry.shape}{province shapefile in long format}

\item{prov.shape}{province shapefile in long format}

\item{start_date}{start date}

\item{end_date}{end date}

\item{output_path}{where to save the figure}
}
\value{
ggplot object
}
\description{
Generate a map of AFP cases
}
