% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_afp_prov_year}
\alias{generate_afp_prov_year}
\title{AFP cases by province and year}
\usage{
generate_afp_prov_year(afp.by.month.prov, start_date, end_date)
}
\arguments{
\item{afp.by.month.prov}{table summarizing afp cases by month and province}
}
\value{
ggplot object
}
\description{
AFP cases by province and year
}
