% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.table.functions.R
\name{generate_c3_table}
\alias{generate_c3_table}
\title{Environmental surveillance KPIs by site}
\usage{
generate_c3_table(
  es_data,
  start_date,
  end_date,
  .group_by = c("ADM0_NAME", "ctry.guid", "reporting.year", "lat", "lng", "site.id",
    "site.name"),
  risk_category = NULL
)
}
\arguments{
\item{es_data}{\code{tibble} Environmental surveillance data.}

\item{start_date}{\code{str} Start date of the analysis in YYYY-MM-DD format.}

\item{end_date}{\code{str} End date of the analysis in YYYY-MM-DD format.}

\item{.group_by}{\code{list} How results should be grouped. By default, the summary
is grouped by country, site, and reporting year.}
}
\value{
\code{tibble} A summary table of environmental surveillance KPIs.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Environmental surveillance KPIs summarized by site.
}
\examples{
raw_data <- get_all_polio_data(attach.spatial.data = FALSE)
c3 <- generate_c3_table(raw_data$es, "2021-01-01", "2023-12-31")
}
