% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_case_num_dose_g}
\alias{generate_case_num_dose_g}
\title{Generate zero-dose children barcharts}
\usage{
generate_case_num_dose_g(ctry.data, start_date, end_date)
}
\arguments{
\item{ctry.data}{RDS file containing polio data of country}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}
}
\value{
ggplot barplot
}
\description{
Generate zero-dose children barcharts
}
