% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_ctry_timeliness_graph}
\alias{generate_ctry_timeliness_graph}
\title{Timeliness bar graph at the country level}
\usage{
generate_ctry_timeliness_graph(int.data)
}
\arguments{
\item{int.data}{summary table with timeliness intervals at the country level}
}
\value{
ggplot object bar graph with timeliness intervals
}
\description{
Timeliness bar graph at the country level
}
