% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/desk.review.functions.R
\name{generate_data}
\alias{generate_data}
\title{Handles the logic of which files to load in the current R session}
\usage{
generate_data(
  data_path,
  data_size,
  country_name,
  dr_data_path,
  attach_spatial_data
)
}
\arguments{
\item{data_path}{path of the data folder used in the desk review}

\item{data_size}{\code{str} "small", "medium", or "large"}

\item{country_name}{\code{str} name of the country}

\item{dr_data_path}{\code{str} path to the dataset}

\item{attach_spatial_data}{}
}
\value{
\code{list} large list containing polio data
}
\description{
Handles the logic of which files to load in the current R session
}
