% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.powerpoint.functions.R
\name{generate_dr_ppt}
\alias{generate_dr_ppt}
\title{Generate the desk review slide deck}
\usage{
generate_dr_ppt(
  ppt_template_path,
  ctry.data,
  assump,
  start_date,
  end_date,
  pop.map,
  pop.map.prov,
  afp.case.map,
  afp.epi.curve,
  surv.ind.tab,
  afp.dets.prov.year,
  npafp.map,
  npafp.map.dist,
  stool.ad.maps,
  stool.ad.maps.dist,
  inad.tab.flex,
  tab.60d,
  timely_nation,
  timely_prov,
  mapt_all,
  es.site.det,
  es.det.map,
  es.timely,
  es.table,
  country,
  ppt_output_path
)
}
\arguments{
\item{ppt_template_path}{path to the PowerPoint template}

\item{ctry.data}{RDS file containing polio data for a country}

\item{assump}{list of assumptions}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}

\item{pop.map}{country pop map}

\item{pop.map.prov}{prov pop map}

\item{afp.case.map}{map of afp cases}

\item{afp.epi.curve}{afp epicurv}

\item{surv.ind.tab}{surveillance indicator table}

\item{afp.dets.prov.year}{afp dets for province}

\item{npafp.map}{npafp map for country level}

\item{npafp.map.dist}{npafp map for district level}

\item{stool.ad.maps}{stool adequacy maps at province}

\item{stool.ad.maps.dist}{stool adequact maps at district}

\item{inad.tab.flex}{inadequacy table}

\item{tab.60d}{60 day follow up table}

\item{timely_nation}{timeliness at country level}

\item{timely_prov}{timeliness at province level}

\item{mapt_all}{map with all indicators}

\item{es.site.det}{ES site viral detection}

\item{es.det.map}{ES site detection maps}

\item{es.timely}{ES timeliness}

\item{es.table}{ES table}

\item{country}{name of the country}

\item{ppt_output_path}{path where the powerpoint should be outputted}
}
\value{
does not return anything
}
\description{
Generate the desk review slide deck
}
