% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.powerpoint.functions.R
\name{generate_dr_ppt2}
\alias{generate_dr_ppt2}
\title{Generate the desk review slide deck from the figures folder
Generating the PowerPoint from the figures folder is generally faster and allows
figures to remain consistent. Tables remain as PowerPoint tables.}
\usage{
generate_dr_ppt2(
  ctry.data,
  start_date,
  end_date,
  surv.ind.tab,
  inad.tab.flex,
  tab.60d,
  es.table,
  ppt_template_path = NULL,
  fig.path = Sys.getenv("DR_FIGURE_PATH"),
  country = Sys.getenv("DR_COUNTRY"),
  ppt_output_path = Sys.getenv("DR_POWERPOINT_PATH")
)
}
\arguments{
\item{ctry.data}{RDS file containing polio data for a country}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}

\item{surv.ind.tab}{surveillance indicator table}

\item{inad.tab.flex}{inadequacy table}

\item{tab.60d}{60 day follow up table}

\item{es.table}{ES table}

\item{ppt_template_path}{path to the PowerPoint template}

\item{fig.path}{file path to the figures folder}

\item{country}{name of the country}

\item{ppt_output_path}{path where the powerpoint should be outputted}
}
\value{
does not return anything
}
\description{
Generate the desk review slide deck from the figures folder
Generating the PowerPoint from the figures folder is generally faster and allows
figures to remain consistent. Tables remain as PowerPoint tables.
}
