% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.powerpoint.functions.R
\name{generate_dr_ppt2}
\alias{generate_dr_ppt2}
\title{Generate the desk review slide deck from the figures folder}
\usage{
generate_dr_ppt2(
  ctry.data,
  start_date,
  end_date,
  surv.ind.tab,
  inad.tab.flex,
  tab.60d,
  pop.tab,
  es.table,
  ppt_template_path = NULL,
  fig.path = Sys.getenv("DR_FIGURE_PATH"),
  country = Sys.getenv("DR_COUNTRY"),
  ppt_output_path = Sys.getenv("DR_POWERPOINT_PATH")
)
}
\arguments{
\item{ctry.data}{\code{list} Country polio data. Either the output of \code{\link[=extract_country_data]{extract_country_data()}} or
\code{\link[=init_dr]{init_dr()}}.}

\item{start_date}{\code{str} Start date of desk review.}

\item{end_date}{\code{str} End date of desk review.}

\item{surv.ind.tab}{\code{flextable} Surveillance indicator table}

\item{inad.tab.flex}{\code{flextable} Inadequates table.}

\item{tab.60d}{\code{flextable} 60-day follow-up table.}

\item{pop.tab}{\code{flextable} Population table.}

\item{es.table}{\code{flextable} ES table.}

\item{ppt_template_path}{\code{str} Path to the PowerPoint template.}

\item{fig.path}{\code{str} File path to the figures folder.}

\item{country}{\code{str} Name of the country.}

\item{ppt_output_path}{\code{str} Path where the PowerPoint should be outputted.}
}
\value{
None.
}
\description{
Generating the PowerPoint from the figures folder is generally faster and allows
figures to remain consistent. Tables remain as PowerPoint tables.
}
\examples{
\dontrun{
# Assume all figures and tables are assigned to the appropriate variable.
ppt_template <- "C:/Users/ABC1/Desktop/deskreview_template.pptx"
generate_dr_ppt2(ctry.data, start_date, end_date,
surv.ind.tab, inad.tab.flex, tab.60d, es.table,
ppt_template_path = ppt_template)
}

}
