% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.es.functions.R
\name{generate_es_data_long}
\alias{generate_es_data_long}
\title{Generate ES data with viral detection columns}
\usage{
generate_es_data_long(es.data)
}
\arguments{
\item{es.data}{\code{tibble} ES data.}
}
\value{
\code{tibble} ES data with viral detection columns.
}
\description{
The function importantly adds the \code{ed.detect}, \code{all_dets}, and \code{year} columns to the
environmental surveillance data. However, this function will probably be moved
as part of the \code{\link[=clean_es_data]{clean_es_data()}} function in the future.
}
\examples{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry.data <- extract_country_data("algeria")
ctry.data$es <- clean_es_data(ctry.data)
es.data.long <- generate_es_data_long(ctry.data$es)

}
