% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.es.functions.R
\name{generate_es_data_long}
\alias{generate_es_data_long}
\title{Generate ES data with viral detection columns}
\usage{
generate_es_data_long(es.data)
}
\arguments{
\item{es.data}{\code{tibble} ES data.}
}
\value{
\code{tibble} ES data with viral detection columns.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The function importantly adds the \code{ed.detect}, \code{all_dets}, and \code{year} columns to the
environmental surveillance data.
}
\details{
This function will is now part of the \code{\link[=clean_es_data]{clean_es_data()}}.
}
\examples{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry.data <- extract_country_data("algeria", raw.data)
ctry.data$es <- clean_es_data(ctry.data)
es.data.long <- generate_es_data_long(ctry.data$es)

}
\keyword{internal}
