% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_es_det_map}
\alias{generate_es_det_map}
\title{Generate ES detection map}
\usage{
generate_es_det_map(
  es.data,
  es.data.long,
  ctry.shape,
  prov.shape,
  es_start_date = (end_date - lubridate::years(1)),
  es_end_date = end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH")
)
}
\arguments{
\item{es.data}{ES data for a country}

\item{es.data.long}{ES data summary and in long format}

\item{ctry.shape}{recent country shapefile}

\item{prov.shape}{recent province shapefile}

\item{es_start_date}{ES start date}

\item{es_end_date}{ES end date}

\item{output_path}{where to save the figure}
}
\value{
ggplot map of ES detections
}
\description{
Generate ES detection map
}
