% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_es_det_map}
\alias{generate_es_det_map}
\title{Generate ES detection map}
\usage{
generate_es_det_map(
  es.data,
  ctry.shape,
  prov.shape,
  es_start_date,
  es_end_date
)
}
\arguments{
\item{es.data}{ES data for a country}

\item{ctry.shape}{recent country shapefile}

\item{prov.shape}{recent province shapefile}

\item{es_start_date}{ES start date}

\item{es_end_date}{ES end date}
}
\value{
ggplot map of ES detections
}
\description{
Generate ES detection map
}
