% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_es_site_det}
\alias{generate_es_site_det}
\title{Virus detection in ES sites}
\usage{
generate_es_site_det(
  sia.data,
  es.data,
  es_start_date = (lubridate::as_date(es_end_date) - lubridate::years(1)),
  es_end_date = end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH"),
  vaccine_types = NULL,
  detection_types = NULL,
  ctry.data = lifecycle::deprecated(),
  es.data.long = lifecycle::deprecated()
)
}
\arguments{
\item{sia.data}{\code{tibble} SIA surveillance data.}

\item{es.data}{Environmental surveillance data, cleaned using \code{\link[=clean_es_data]{clean_es_data()}} or a cleaned \code{ctry.data$es}.}

\item{es_start_date}{\code{str} Start date of analysis. By default, it is one year from the end date.}

\item{es_end_date}{\code{str} End date of analysis.}

\item{output_path}{\code{str} Local path to output the figure to.}

\item{vaccine_types}{\code{list} A named list with colors assigned names corresponding to vaccine types. By
default, it will use a prefilled list inside the function. However, the function will alert for missing
vaccine types and the user must pass another list appended by that vaccine type.}

\item{detection_types}{\code{list} A named list with colors assigned names corresponding to viral detection type.
By default, it will use a prefilled list inside the function. However, the function will alert for missing
detection types and the user must pass another list appended by that vaccine type.}

\item{ctry.data}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please pass the SIA data directly to sia.data instead of a list containing it.}

\item{es.data.long}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
Please pass cleaned ES data instead.}
}
\value{
\code{ggplot} A dot plot of viral detections per ES sites and SIA campaigns.
}
\description{
Generates a dot plot for viral detections across ES sites, with SIA dates
overlaid.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
es.data <- clean_es_data(ctry.data$es)
generate_es_site_det(ctry.data, es.data)
}
}
