% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_es_timely}
\alias{generate_es_timely}
\title{Generate ES timeliness scatterplot}
\usage{
generate_es_timely(
  es.data,
  es_start_date = (end_date - lubridate::years(1)),
  es_end_date = end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH")
)
}
\arguments{
\item{es.data}{ES data}

\item{es_start_date}{start date of ES}

\item{es_end_date}{end date of ES}

\item{output_path}{where to save the figure}
}
\value{
ggplot scatterplot for timeliness
}
\description{
Generate ES timeliness scatterplot
}
