% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_inad_tab}
\alias{generate_inad_tab}
\title{Issues with stool adequacy at the country level}
\usage{
generate_inad_tab(afp_data, cstool, start_date, end_date)
}
\arguments{
\item{afp_data}{\code{tibble} \code{afp.all.2} of the output of \code{\link[=extract_country_data]{extract_country_data()}}.}

\item{cstool}{\code{tibble} Stool adequacy at the country level. This is the output of \code{\link[=f.stool.ad.01]{f.stool.ad.01()}}.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}
}
\value{
\code{flextable} Summary table containing stool adequacy issues at the country level.
}
\description{
Generates a summary table at the country level highlighting issues around stool adequacy.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
start_date <- "2021-01-01"
end_date <- "2023-12-31"
cstool <- f.stool.ad.01(
  afp.data = ctry.data$afp.all.2,
  admin.data = ctry.data$ctry.pop,
  start.date = start_date,
  end.date = end_date,
  spatial.scale = "ctry",
  missing = "good",
  bad.data = "inadequate",
  rolling = F,
  sp_continuity_validation = F
)
generate_inad_tab(ctry.data, cstool, start_date, end_date)
}

}
