% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_iss_barplot}
\alias{generate_iss_barplot}
\title{Generate the ISS/eSURV barplot}
\usage{
generate_iss_barplot(iss.data, start_date, end_date)
}
\arguments{
\item{iss.data}{tibble of ISS data}

\item{start_date}{start date of the desk review}

\item{end_date}{end date of the desk review}
}
\value{
ggplot object of a barplot
}
\description{
Generate the ISS/eSURV barplot
}
