% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_kpi_ev_map}
\alias{generate_kpi_ev_map}
\title{EV detection rate map}
\usage{
generate_kpi_ev_map(
  c3,
  .year_label,
  who_region = NULL,
  output_path = Sys.getenv("KPI_FIGURES"),
  dot_size = 2.3,
  ctry_sf = NULL
)
}
\arguments{
\item{c3}{\code{tibble} Output of \code{\link[=generate_c3_table]{generate_c3_table()}}. This must be summarized
at the site level (i.e., use the default \code{.group_by}) param of the
\code{\link[=generate_c3_table]{generate_c3_table()}}.}

\item{.year_label}{\code{str} Roll up year (i.e., \verb{"Year 1", "Year 2", ...}).}

\item{who_region}{\code{str} Name of the region or a list of regions.}

\item{output_path}{\code{str} Where to output the figure to. Defaults to the
figure path assigned after running \code{\link[=init_kpi]{init_kpi()}}.}

\item{dot_size}{\code{num} Point size.}

\item{ctry_sf}{\code{sf} Country shapefile in long format. Output of
\code{\link[=load_clean_ctry_sp]{load_clean_ctry_sp()}} with \code{type = "long"}. Defaults to \code{NULL}, which will download
the required country shapefile when the function is ran.}
}
\value{
\code{ggplot} A map showing EV detection rate by site.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Generates a map of EV detection rates for each environmental surveillance site.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
c3 <- generate_c3_table(raw_data$es, "2021-01-01", "2023-12-31")
map <- generate_kpi_ev_map(c3, "Year 1", "AFRO", getwd())
}
}
