% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_kpi_npafp_map}
\alias{generate_kpi_npafp_map}
\title{Generate district level NPAFP maps by region}
\usage{
generate_kpi_npafp_map(
  c2,
  year,
  who_region = NULL,
  risk_category = NULL,
  output_path = Sys.getenv("KPI_FIGURES"),
  ctry_sf = NULL,
  dist_sf = NULL
)
}
\arguments{
\item{c2}{\code{tibble} Output of \code{\link[=generate_c2_table]{generate_c2_table()}}.}

\item{year}{\code{num} Numeric year of interest.}

\item{who_region}{\code{str} A WHO region or a list of regions. Valid values are:
\itemize{
\item \code{"AFRO"}: African Region
\item \code{"AMRO"}: Region of the Americas
\item \code{"EMRO"}: Eastern Mediterranean Region
\item \code{"EURO"}: European Region
\item \code{"SEARO"}:South-East Asia Region
\item \code{"WPRO"}:Western Pacific Region
}}

\item{output_path}{\code{str} Where to output the figure to. Defaults to the path
initialized when \code{\link[=init_kpi]{init_kpi()}} was ran.}

\item{ctry_sf}{\code{sf} Country shapefile in long format. Output of
\link{load_clean_ctry_sp(type = "long")}. Defaults to \code{NULL}, which will download
the required country shapefile when the function is ran.}

\item{dist_sf}{\code{sf} District shapefile in long format. Output of
\link{load_clean_dist_sp(type = "long")}. Defaults to \code{NULL}, which will download
the required district shapefile when the function is ran.}
}
\value{
\code{ggplot} A district NPAFP map.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Generates a map of district NPAFP rates.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
c2 <- generate_c2_table(raw_data$afp, raw_data$dist, "2021-01-01",
"2023-12-31", c("ctry", "dist", "adm2guid", "year"))
map <- generate_kpi_npafp_map(c2, 2023, "AFRO", output_path = getwd())
}
}
