% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kpi.figure.functions.R
\name{generate_kpi_npafp_map}
\alias{generate_kpi_npafp_map}
\title{Generate district level NPAFP maps by region}
\usage{
generate_kpi_npafp_map(
  c2,
  ctry_sf,
  dist_sf,
  who_region,
  .year,
  output_path = Sys.getenv("KPI_FIGURES")
)
}
\arguments{
\item{c2}{\code{tibble} Output of \code{\link[=generate_c2_table]{generate_c2_table()}}}

\item{ctry_sf}{\code{sf} Country shapefile in long format.}

\item{dist_sf}{\code{sf} District shapefile in long format.}

\item{who_region}{\code{str} WHO region.}

\item{.year}{\code{num} Numeric year of interest.}

\item{output_path}{\code{str} Where to output the figure to. Defaults to the path
initialized when \code{\link[=init_kpi]{init_kpi()}} was ran.}
}
\value{
\code{ggplot} A district NPAFP map.
}
\description{
Generate district level NPAFP maps by region
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
c2 <- generate_c2_table(raw_data$afp, raw_data$dist, "2021-01-01",
"2023-12-31", c("ctry", "dist", "adm2guid", "year"))
ctry_sf <- load_clean_ctry_sp(type="long")
dist_sf <- load_clean_dist_sp(type="long")
map <- generate_kpi_npafp_map(c2, ctry_sf, dist_sf, "AFRO", 2022, getwd())
}
}
