% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_npafp_maps}
\alias{generate_npafp_maps}
\title{Maps of NPAFP rates by province and year}
\usage{
generate_npafp_maps(
  prov.extract,
  ctry.shape,
  prov.shape,
  start_date,
  end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH"),
  caption_size = 2
)
}
\arguments{
\item{prov.extract}{\code{tibble} Province NPAFP rate table. This is the output of
\code{\link[=f.npafp.rate.01]{f.npafp.rate.01()}} calculated at the province level.}

\item{ctry.shape}{\code{sf} Country shape in long format.}

\item{prov.shape}{\code{sf} Province shape in long format.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}

\item{output_path}{\code{str} Local path where the figure is saved to.}

\item{caption_size}{\code{numeric} Size of the caption. Default is \code{2}.}
}
\value{
\code{ggplot} Map of NPAFP rates by province.
}
\description{
Generates a map of NPAFP rates for each province per year.
}
\examples{
\dontrun{
ctry.data <- init_dr("algeria")
prov.extract <- f.npafp.rate.01(afp.data = ctry.data$afp.all.2,
                                pop.data = ctry.data$prov.pop,
                                start.date = start_date,
                                end.date = end_date,
                                spatial.scale = "prov",
                                pending = T,
                                rolling = F,
                                sp_continuity_validation = F
                                )
ctry.shape <- load_clean_ctry_sp(ctry_name = "ALGERIA", type = "long")
prov.shape <- load_clean_prov_sp(ctry_name = "ALGERIA", type = "long")
generate_npafp_maps(prov.extract, ctry.shape, prov.shape, "2021-01-01", "2023-12-31")
}
}
\seealso{
\code{\link[=load_clean_ctry_sp]{load_clean_ctry_sp()}}, \code{\link[=load_clean_prov_sp]{load_clean_prov_sp()}}
}
