% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_npafp_maps}
\alias{generate_npafp_maps}
\title{Map of NPAFP rate by province}
\usage{
generate_npafp_maps(
  prov.extract,
  ctry.shape,
  prov.shape,
  start_date,
  end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH"),
  caption_size = 2
)
}
\arguments{
\item{prov.extract}{province NPAFP rate}

\item{ctry.shape}{recent country shape}

\item{prov.shape}{recent province shape}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}

\item{output_path}{where to save the figure}

\item{caption_size}{size of the caption}
}
\value{
ggplot object
}
\description{
Map of NPAFP rate by province
}
