% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_npafp_maps_dist}
\alias{generate_npafp_maps_dist}
\title{Generate map of NPAFP rates by district}
\usage{
generate_npafp_maps_dist(
  dist.extract,
  ctry.shape,
  prov.shape,
  dist.shape,
  start_date,
  end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH"),
  caption_size = 2
)
}
\arguments{
\item{dist.extract}{NPAFP rates by district table.}

\item{ctry.shape}{country shapefile in long format}

\item{prov.shape}{province shapefile in long format}

\item{dist.shape}{district shapefile in long format}

\item{start_date}{start date}

\item{end_date}{end date}

\item{output_path}{where to save the figure}

\item{caption_size}{size of the caption}
}
\value{
ggplot object
}
\description{
Generate map of NPAFP rates by district
}
