% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_pop_map}
\alias{generate_pop_map}
\title{Generate a map of population data}
\usage{
generate_pop_map(ctry.data, prov.shape, end_date)
}
\arguments{
\item{ctry.data}{RDS object}

\item{prov.shape}{province of most recent shape file}

\item{end_date}{end date of the desk review}
}
\value{
a ggplot object containing map of population data
}
\description{
Generate a map of population data
}
