% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_pop_tab}
\alias{generate_pop_tab}
\title{Generate table for population}
\usage{
generate_pop_tab(prov.case.ind, pstool, start_date, end_date)
}
\arguments{
\item{prov.case.ind}{case indicator at province level}

\item{pstool}{stool adequacy at porvince level}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}
}
\value{
a flextable containing population indicators
}
\description{
Generate table for population
}
