% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_potentially_compatibles_cluster}
\alias{generate_potentially_compatibles_cluster}
\title{Title}
\usage{
generate_potentially_compatibles_cluster(cases.need60day)
}
\arguments{
\item{cases.need60day}{summary table of cases that need 60 day follow-up}
}
\value{
summary table with clustering info
}
\description{
Title
}
