% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_potentially_compatibles_cluster}
\alias{generate_potentially_compatibles_cluster}
\title{Creating a table of compatible and potentially compatible cases}
\usage{
generate_potentially_compatibles_cluster(cases.need60day, create_cluster = F)
}
\arguments{
\item{cases.need60day}{\code{tibble} Summary table of cases that need 60-day follow-up.
This is the output of \code{\link[=generate_60_day_table_data]{generate_60_day_table_data()}}.}

\item{create_cluster}{\code{bool} Add column for clusters? Default to \code{FALSE}.}
}
\value{
\code{tibble} A summary table of cases.
}
\description{
Creates a table of compatible and potentially compatible cases, with an
optional parameter to run a clustering algorithm.
}
\examples{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry.data <- extract_country_data("algeria", raw.data)
stool.data <- generate_stool_data(ctry.data$afp.all.2, "2021-01-01", "2023-12-31",
                                  "good", "inadequate"
                                  )
table60.days <- generate_60_day_table_data(stool.data, "2021-01-01", "2023-12-31")
pot.c.clust <- generate_potentially_compatibles_cluster(table60.days,
                                                        create_cluster = TRUE
                                                        )

}
