% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_prov_timeliness_graph}
\alias{generate_prov_timeliness_graph}
\title{Timeliness interval bar graph at a province level}
\usage{
generate_prov_timeliness_graph(int.data, afp.prov.year)
}
\arguments{
\item{int.data}{summary table of median timeliness intervals at a province level}

\item{afp.prov.year}{labels for AFP dataset summarized by year}
}
\value{
a ggplot object of timeliness interval at the province level
}
\description{
Timeliness interval bar graph at a province level
}
