% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_stool_ad_maps_dist}
\alias{generate_stool_ad_maps_dist}
\title{Stool adequacy map by district}
\usage{
generate_stool_ad_maps_dist(
  ctry.data,
  dstool,
  ctry.shape,
  dist.shape,
  prov.shape,
  start_date,
  end_date
)
}
\arguments{
\item{ctry.data}{RDS file of polio data for a country}

\item{dstool}{district stool adequacy}

\item{ctry.shape}{recent country shapefile}

\item{dist.shape}{recent district shapefile}

\item{prov.shape}{recent province shapefile}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}
}
\value{
ggplot map
}
\description{
Stool adequacy map by district
}
