% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_stool_ad_maps_dist}
\alias{generate_stool_ad_maps_dist}
\title{Stool adequacy map by district}
\usage{
generate_stool_ad_maps_dist(
  ctry.data,
  dstool,
  ctry.shape,
  prov.shape,
  dist.shape,
  start_date,
  end_date,
  output_path = Sys.getenv("DR_FIGURE_PATH"),
  caption_size = 3
)
}
\arguments{
\item{ctry.data}{RDS file of polio data for a country}

\item{dstool}{district stool adequacy table}

\item{ctry.shape}{country shapefile in long format}

\item{prov.shape}{province shapefile in long format}

\item{dist.shape}{district shapefile in long format}

\item{start_date}{start date}

\item{end_date}{end date}

\item{output_path}{where to save the figure}

\item{caption_size}{size of the caption}
}
\value{
ggplot object
}
\description{
Stool adequacy map by district
}
