% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.summary.table.functions.R
\name{generate_stool_data}
\alias{generate_stool_data}
\title{Generate stool adequacy columns in the AFP dataset}
\usage{
generate_stool_data(afp.data, start_date, end_date)
}
\arguments{
\item{afp.data}{tibble of AFP data (afp.all.2)}

\item{start_date}{start date of the desk review}

\item{end_date}{end date of the desk review}
}
\value{
a tibble containing stool adequacy columns
}
\description{
Generate stool adequacy columns in the AFP dataset
}
