% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_surv_ind_tab}
\alias{generate_surv_ind_tab}
\title{Generate surveillance indicator table}
\usage{
generate_surv_ind_tab(
  ctry.data,
  dis.extract,
  cstool,
  dstool,
  afp.case,
  country_name = Sys.getenv("DR_COUNTRY")
)
}
\arguments{
\item{ctry.data}{RDS file countaining polio data of a country}

\item{dis.extract}{district NPAFP rate}

\item{cstool}{stool adequacy at country level}

\item{dstool}{stool adequact at a district level}

\item{afp.case}{AFP cases}
}
\value{
a flex table object for timeliness indicators
}
\description{
Generate surveillance indicator table
}
