% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.figure.functions.R
\name{generate_timeliness_maps}
\alias{generate_timeliness_maps}
\title{Map containing timeliness of samples}
\usage{
generate_timeliness_maps(
  ctry.data,
  ctry.shape,
  prov.shape,
  start_date,
  end_date,
  mark_x = T,
  pt_size = 4,
  output_path = Sys.getenv("DR_FIGURE_PATH")
)
}
\arguments{
\item{ctry.data}{RDS file containing polio data for a country}

\item{ctry.shape}{country shapefile in long format}

\item{prov.shape}{province shapefile in long format}

\item{start_date}{start date}

\item{end_date}{end date}

\item{mark_x}{whether to put a mark on where AFP cases are less than 5}

\item{pt_size}{size of the marks}

\item{output_path}{where to save the figure}
}
\value{
ggplot object
}
\description{
Map containing timeliness of samples
}
