% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{generate_year_lab}
\alias{generate_year_lab}
\title{Generate a summary of AFP samples by year}
\usage{
generate_year_lab(ctry.data, start_date, end_date)
}
\arguments{
\item{ctry.data}{\code{list} Large list containing country polio data. This is the output
of \code{\link[=init_dr]{init_dr()}} or \code{\link[=extract_country_data]{extract_country_data()}}.}

\item{start_date}{\code{str} Start date of analysis.}

\item{end_date}{\code{str} End date of analysis.}
}
\value{
\code{tibble} A table containing summary of AFP cases by year and country.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

Generates a summary table of the number of AFP cases per country and year.
This function is used primarily with \code{\link[=generate_ctry_timeliness_graph]{generate_ctry_timeliness_graph()}} as a
label of the y-axis. However, as of sirfunctions 1.3.0,
\code{\link[=generate_ctry_timeliness_graph]{generate_ctry_timeliness_graph()}} now creates its own labels.
}
\examples{
\dontrun{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry.data <- extract_country_data("algeria", raw.data)
ctry.labels <- generate_year_lab(ctry.data, "2021-01-01", "2023-12-31")
}

}
\keyword{internal}
