% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_all_polio_data}
\alias{get_all_polio_data}
\title{Retrieve all pre-processed Polio Data}
\usage{
get_all_polio_data(
  size = "small",
  folder = "GID/PEB/SIR/Data/",
  force.new.run = F,
  recreate.static.files = F,
  attach.spatial.data = T
)
}
\arguments{
\item{size}{str: "small", "medium", "large", defaults to "small"}

\item{folder}{str: location of the CDC pre-processed endpoint}

\item{force.new.run}{boolean: default F, if T will run recent data and cache}

\item{recreate.static.files}{boolean: default F, if T will run all data and cache}

\item{attach.spatial.data}{boolean: default T, adds spatial data to downloaded object}
}
\value{
named list containing polio data that is relevant to CDC
}
\description{
Download POLIS data from the CDC pre-processed endpoint. By default
this function will return a "small" or recent dataset. This is primarily for data
that is from 2019 onwards. You can specify a "medium" sized dataset for data
that is from 2016 onwards. Finally the "large" sized dataset will provide information
from 2001 onwards. Regular pulls form the data will recreate the "small" dataset
when new information is availble and the Data Management Team can force the
creation of the "medium" and "large" static datasets as necessary.
}
