% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_all_polio_data}
\alias{get_all_polio_data}
\title{Retrieve all pre-processed polio data}
\usage{
get_all_polio_data(
  size = "small",
  folder = "GID/PEB/SIR/Data/",
  force.new.run = F,
  recreate.static.files = F,
  attach.spatial.data = T
)
}
\arguments{
\item{size}{\code{str} Size of data to download. Defaults to \code{"small"}.
\itemize{
\item \code{"small"}: data from 2019-present
\item \code{"medium"}: data from 2016-present
\item \code{"large"}: data from 2001-present
}}

\item{folder}{\code{str} Location of the CDC pre-processed endpoint, defaults to "GID/PEB/SIR/Data".}

\item{force.new.run}{\code{bool} Default \code{FALSE}, if \code{TRUE} will run recent data and cache.}

\item{recreate.static.files}{\code{bool} Default \code{FALSE}, if \code{TRUE} will run all data and cache.}

\item{attach.spatial.data}{\code{bool} Default \code{TRUE}, adds spatial data to downloaded object.}
}
\value{
Named \code{list} containing polio data that is relevant to CDC.
}
\description{
Download POLIS data from the CDC pre-processed endpoint. By default
this function will return a "small" or recent dataset. This is primarily for data
that is from 2019 onwards. You can specify a "medium" sized dataset for data
that is from 2016 onwards. Finally the "large" sized dataset will provide information
from 2001 onwards. Regular pulls form the data will recreate the "small" dataset
when new information is availble and the Data Management Team can force the
creation of the "medium" and "large" static datasets as necessary.
}
\examples{
\dontrun{
raw.data <- get_all_polio_data() # downloads data since 2019, including spatial files
}
raw.data <- get_all_polio_data(size = "small", attach.spatial.data = FALSE) # exclude spatial data

}
