% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_all_polio_data}
\alias{get_all_polio_data}
\title{Retrieve all pre-processed Polio Data}
\usage{
get_all_polio_data(
  folder = file.path("", "", "cdc.gov", "project", "CGH_GID_Active", "PEB", "SIR",
    "DATA", "Core 2.0", "datafiles_01")
)
}
\arguments{
\item{folder}{str: location of the CDC pre-processed endpoint}
}
\value{
named list containing polio data that is relevant to CDC
}
\description{
Download all Polio data from CDC pre-processed endpoint
}
