% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_azure_storage_connection}
\alias{get_azure_storage_connection}
\title{Validate connection to EDAV}
\usage{
get_azure_storage_connection()
}
\value{
azure container verification
}
\description{
Validate connection to CDC EDAV
}
