% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{get_azure_storage_connection}
\alias{get_azure_storage_connection}
\title{Validate connection to EDAV}
\usage{
get_azure_storage_connection(
  app_id = "04b07795-8ddb-461a-bbee-02f9e1bf7b46",
  auth = "authorization_code"
)
}
\arguments{
\item{app_id}{\code{str} Application ID defaults to "04b07795-8ddb-461a-bbee-02f9e1bf7b46",
this can be changed if you have a service principal.}

\item{auth}{\code{str} Authorization type defaults to "authorization_code",
this can be changed if you have a service principal.}
}
\value{
Azure container verification
}
\description{
Generate token which connects to CDC EDAV resources and
validates that the individual still has access. The current tenant ID
is hard coded for CDC resources.
}
\examples{
azcontainer <- get_azure_storage_connection()

}
