% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_cdc_childvaxview_data.R
\name{get_cdc_childvaxview_data}
\alias{get_cdc_childvaxview_data}
\title{Pull CDC NCIRD childvaxview immunization coverage data}
\usage{
get_cdc_childvaxview_data(
  geo_level = NULL,
  vaccines = NULL,
  limit = 1000,
  base_url = "https://data.cdc.gov/resource/fhky-rtsk.json"
)
}
\arguments{
\item{geo_level}{\code{str} Geographic categories of coverage data.
Choose from: 'national', 'regional', 'state', or 'substate'.}

\item{vaccines}{\code{str} A string or vector of strings of vaccines for which to provide coverage data.
Choose from: 'DTaP', 'Polio', 'Hep B', 'PCV', 'Varicella', 'MMR', 'Hib', 'Hep A', 'Influenza', 'Rotavirus', 'Combined 7 series'.}

\item{limit}{\code{int} Number of rows to download, defaults to max allowed (1000).}

\item{base_url}{\code{str} URL to US CDC NCIRD API endpoint.
Defaults to "https://data.cdc.gov/resource/fhky-rtsk.json".}
}
\value{
\code{tibble} Dataframe of vaccine coverage estimates for all VPDs.
}
\description{
Pull coverage data from API and filter by desired geographic level and vaccines.
}
\examples{
\dontrun{
cdc_data <- get_cdc_childvaxview_data(geo_level="substate")
cdc_data <- get_cdc_childvaxview_data(geo_level="national", vaccines=c("Polio","MMR"))
}

}
