% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{get_lab_locs}
\alias{get_lab_locs}
\title{Gets information related to testing labs}
\usage{
get_lab_locs(path)
}
\arguments{
\item{path}{path the lab location file}
}
\value{
tibble containing the test lab locations
}
\description{
Gets information related to testing labs
}
