% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{get_lab_locs}
\alias{get_lab_locs}
\title{Table of information regarding testing labs in each country}
\usage{
get_lab_locs(path = NULL)
}
\arguments{
\item{path}{\code{str} Path to the lab location file. Defaults to \code{NULL}.}
}
\value{
\code{tibble} A table containing the test lab location information.
}
\description{
Imports information on testing labs for each country, either from a CSV file
or downloaded from EDAV. If no argument is passed, the function will download
the table from EDAV.
}
\examples{
\dontrun{
ctry.seq <- get_lab_locs()
}

}
