% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.powerpoint.functions.R
\name{get_ppt_template}
\alias{get_ppt_template}
\title{Get path of the PowerPoint template}
\usage{
get_ppt_template(path = NULL, filename = "desk_review_template.pptx")
}
\arguments{
\item{path}{path to the PowerPoint template. If NULL, will download from the sg-desk-review GitHub repository}

\item{filename}{name of the PowerPoint template}
}
\value{
a string containing the path of the PowerPoint template
}
\description{
Get path of the PowerPoint template
}
