% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{get_region}
\alias{get_region}
\title{Determines whether lab data is EMRO or AFRO}
\usage{
get_region(country_name = Sys.getenv("DR_COUNTRY"))
}
\arguments{
\item{country_name}{\code{str} Name of the country.}
}
\value{
\code{str} A string, either \code{"EMRO"} or \code{"AFRO"}.
}
\description{
Outputs the name of the region which a country belongs to.
}
\examples{
\dontrun{
get_region("algeria")
}

}
