% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.es.functions.R
\name{impute_site_coord}
\alias{impute_site_coord}
\title{Adds coordinates for ES surveillance sites missing them}
\usage{
impute_site_coord(ctry.data)
}
\arguments{
\item{ctry.data}{RDS object containing country polio data}
}
\value{
tibble of ES data with imputed coordinates for sites missing them
}
\description{
Adds coordinates for ES surveillance sites missing them
}
