% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{lab_data_errors}
\alias{lab_data_errors}
\title{Checks for common data errors in lab data}
\usage{
lab_data_errors(lab.data, ctry.data, start_date, end_date)
}
\arguments{
\item{lab.data}{lab.data from WHO}

\item{ctry.data}{RDS object containing polio data from a country}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}
}
\value{
a list containing the errors
}
\description{
Checks for common data errors in lab data
}
