% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{lab_data_errors}
\alias{lab_data_errors}
\title{Generate a log of potential errors in the lab data}
\usage{
lab_data_errors(
  lab.data,
  afp.data,
  start.date = start_date,
  end.date = end_date,
  ctry_name = Sys.getenv("DR_COUNTRY"),
  error_path = Sys.getenv("DR_ERROR_PATH"),
  ctry.data = lifecycle::deprecated()
)
}
\arguments{
\item{lab.data}{\code{tibble} Polio lab data.}

\item{afp.data}{\code{tibble} AFP linelist.}

\item{start.date}{\code{str} Start date of the analysis.}

\item{end.date}{\code{str} End date of the analysis.}

\item{ctry_name}{\code{list} or \code{str} A name of a country or a list of countries.
Please pass lab data directly into lab.data parameter instead.}

\item{error_path}{\code{str} File path to store the error log.}

\item{ctry.data}{\code{list} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\value{
None. It outputs locally an Excel file containing the error log.
}
\description{
Checks the loaded lab data for potential issues. The function will detect whether
the lab data loaded either came from the regional office or from global.
}
\examples{
\dontrun{
lab_path <- "C:/Users/ABC1/Desktop/lab_data.xlsx"
start_date <- "2021-01-01"
end_date <- "2023-12-31"
ctry.data <- init_dr("algeria", lab_data_path = lab_path)
lab_data_errors(ctry.data$lab.data, ctry.data$afp.data)
}
}
