% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{lab_data_errors_region}
\alias{lab_data_errors_region}
\title{Check common errors in the regional lab data}
\usage{
lab_data_errors_region(
  ctry.data,
  start.date,
  end.date,
  error_path = Sys.getenv("DR_ERROR_PATH")
)
}
\arguments{
\item{ctry.data}{Rds file containing polio country data}

\item{start.date}{start date of the desk review}

\item{end.date}{end date of the desk review}

\item{error_path}{path to folder to save the error log}
}
\description{
Check common errors in the regional lab data
}
