% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{lab_data_errors_region}
\alias{lab_data_errors_region}
\title{Check common errors in the regional lab data.}
\usage{
lab_data_errors_region(
  lab.data,
  afp.data,
  start.date,
  end.date,
  ctry_name = Sys.getenv("DR_COUNTRY"),
  error_path = Sys.getenv("DR_ERROR_PATH"),
  ctry.data = lifecycle::deprecated()
)
}
\arguments{
\item{lab.data}{\code{tibble} Polio lab data.}

\item{afp.data}{\code{tibble} AFP linelist.}

\item{start.date}{\code{str} Start date of analysis.}

\item{end.date}{\code{str} End date of analysis.}

\item{ctry_name}{\code{str} Name of the country. Defaults to the desk review country.}

\item{error_path}{\code{str} Path to folder to save the error log.}

\item{ctry.data}{\code{list} Large list containing polio country data. Output of either
\code{\link[=extract_country_data]{extract_country_data()}} or \code{\link[=init_dr]{init_dr()}}. Note that lab data must be attached to\code{ctry.data} as
\code{lab_data}.}
}
\value{
None. It outputs an Excel file locally containing the error log.
}
\description{
Error checking for regional lab data. This is a helper function meant to be used inside
\code{\link[=lab_data_errors]{lab_data_errors()}}.
}
\examples{
\dontrun{
lab_path <- "C:/Users/XRG9/lab_data_region.csv"
ctry.data <- init_dr("algeria", lab_data_path = lab_path)
lab_data_errors_region(
  ctry.data$lab.data, ctry.data$afp.all.2,
  "2021-01-01", "2023-12-31"
)
}
}
\keyword{internal}
