% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{lab_data_errors_who}
\alias{lab_data_errors_who}
\title{Checks for common data errors in lab data}
\usage{
lab_data_errors_who(
  ctry.data,
  start.date = start_date,
  end.date = end_date,
  error_path = Sys.getenv("DR_ERROR_PATH")
)
}
\arguments{
\item{ctry.data}{RDS object containing polio data from a country}

\item{lab.data}{lab.data from WHO}

\item{start_date}{start date of desk review}

\item{end_date}{end date of desk review}
}
\description{
Checks for common data errors in lab data
}
