% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.lab.functions.R
\name{lab_data_errors_who}
\alias{lab_data_errors_who}
\title{Checks for common data errors in WHO lab data}
\usage{
lab_data_errors_who(
  lab.data,
  afp.data,
  start.date,
  end.date,
  ctry_name = Sys.getenv("DR_COUNTRY"),
  error_path = Sys.getenv("DR_ERROR_PATH"),
  ctry.data = lifecycle::deprecated()
)
}
\arguments{
\item{lab.data}{\code{tibble} Polio lab data.}

\item{afp.data}{\code{tibble} AFP linelist.}

\item{start.date}{\code{str} Start date of the analysis.}

\item{end.date}{\code{str} End date of the analysis.}

\item{ctry_name}{\code{list} or \code{str} A name of a country or a list of countries.
Please pass lab data directly into lab.data parameter instead.}

\item{error_path}{\code{str} File path to store the error log.}

\item{ctry.data}{\code{list} \ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}}
}
\description{
This function is used internally within \code{\link[=lab_data_errors]{lab_data_errors()}}. This checks for potential errors in the
WHO lab data.
}
\examples{
\dontrun{
lab_path <- "C:/Users/XRG9/lab_data_who.csv"
ctry.data <- init_dr("algeria", lab_data_path = lab_path)
lab_data_errors_who(
  ctry.data$lab.data, ctry.data$afp.all.2,
  "2021-01-01", "2023-12-31"
)
}
}
\keyword{internal}
