% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{load_clean_dist_sp}
\alias{load_clean_dist_sp}
\title{Download district geographic data}
\usage{
load_clean_dist_sp(
  azcontainer = suppressMessages(get_azure_storage_connection()),
  fp = "GID/PEB/SIR/Data/spatial/global.dist.rds",
  dist_guid = NULL,
  ctry_name = NULL,
  end.year = lubridate::year(Sys.Date()),
  st.year = 2000,
  data.only = F,
  type = NULL,
  version = "standard"
)
}
\arguments{
\item{azcontainer}{Azure validated container object.}

\item{fp}{\code{str} Location of geodatabase.}

\item{dist_guid}{\verb{str array} Array of all district GUIDS that you want to pull.}

\item{ctry_name}{\verb{str array} Array of all country names that you want to pull.}

\item{end.year}{\code{int} Last year you want to pull information for. Default is current year.}

\item{st.year}{\code{int} Earlier year of spatial data you want to pull. Default is 2000.}

\item{data.only}{\code{bool} Whether to return a tibble with shapefiles or not. Defaults to \code{FALSE}.}

\item{type}{\code{str} Whether to return a spatial object for every year group. Defaults to \code{NULL}.
\itemize{
\item \code{"long"} Return a dataset for every year group.
\item \code{NULL} Return a dataset only with unique GUIDs and when they were active.
}}

\item{version}{\code{str} Specify whether to return standard shapefiles or new shapefiles
still under evaluation/development. Default is \code{"standard"}.
\itemize{
\item \code{"standard"} Standard shapefiles.
\item \code{"dev"}  New shapefiles still under evaluation/development.
}}
}
\value{
A \code{tibble} or \code{sf} dataframe containing spatial data.
}
\description{
Pulls district shapefiles directly from the geodatabase.
}
\examples{
dist <- load_clean_dist_sp(ctry_name = c("ALGERIA", "NIGERIA"), st.year = 2019)
dist.long <- load_clean_dist_sp(ctry_name = "ALGERIA", st.year = 2019, type = "long")

}
