% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{load_clean_dist_sp}
\alias{load_clean_dist_sp}
\title{Standard function to load District data}
\usage{
load_clean_dist_sp(
  azcontainer = get_azure_storage_connection(),
  fp = "GID/PEB/SIR/Data/spatial/global.dist.rds",
  dist_guid = NULL,
  ctry_name = NULL,
  end.year = year(Sys.Date()),
  st.year = 2000,
  data.only = F,
  type = NULL
)
}
\arguments{
\item{azcontainer}{azure validated container object}

\item{fp}{str: Location of geodatabase}

\item{dist_guid}{array/str: Array of all district GUIDS that you want to pull}

\item{ctry_name}{array/str: Array of all country names that you want to pull}

\item{end.year}{int: last year you want to pull information for - default is current year}

\item{st.year}{int: earlier year of spatial data you want to pull - default is 2000}

\item{data.only}{boolean: default F, if true, returns a rectangular tibble instead of a shape file}

\item{type}{str: "long" or NULL, default NULL, if "long" returns a spatial object for every year group}
}
\value{
tibble or sf dataframe
}
\description{
Pulls district shapefiles directly from the geodatabase
}
