% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.iss.functions.R, R/dr.main.functions.R
\name{load_iss_data}
\alias{load_iss_data}
\title{Read ISS/eSURV data}
\usage{
load_iss_data(iss_data_path)

load_iss_data(iss_data_path)
}
\arguments{
\item{iss_data_path}{}

\item{iss_path}{path to the .csv file}
}
\value{
a tibble containing ISS/eSURV data

a tibble
}
\description{
Read ISS/eSURV data

Function to load the raw ISS/eSurv data
}
