% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{load_lab_data}
\alias{load_lab_data}
\title{Function to load the raw lab data}
\usage{
load_lab_data(lab_data_path)
}
\arguments{
\item{lab_data_path}{file path as a string to the lab data}
}
\value{
a tibble containing lab data
}
\description{
Function to load the raw lab data
}
