% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_afp_guid_to_sf.R
\name{match_afp_guid_to_sf}
\alias{match_afp_guid_to_sf}
\title{Correct GUIDs in the AFP linelist based on the population dataset}
\usage{
match_afp_guid_to_sf(afp_data, prov_pop, dist_pop)
}
\arguments{
\item{afp_data}{\code{tibble} AFP linelist.}

\item{prov_pop}{\code{tibble} Province population file.}

\item{dist_pop}{\code{tibble} District population file.}
}
\value{
\code{tibble} AFP linelist with corrected GUIDs
}
\description{
The function corrects the GUIDs in the AFP linelist to match the GUID associated
with a particular district or province for each year. In certain instances, this is
required when the population/spatial data sets are outdated, and new GUIDs are present in the
AFP linelist.
}
\examples{
\dontrun{
raw_data <- get_all_polio_data()
afp_fixed <- match_afp_guid_to_sf(raw_data$afp, raw_data$prov.pop, raw_data$dist.pop)
}
}
