% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.npafp.rate.01.R
\name{npafp_rolling}
\alias{npafp_rolling}
\title{Calculate the NPAFP rate on a rolling basis}
\usage{
npafp_rolling(
  afp.data,
  year.pop.data,
  start_date,
  end_date,
  spatial_scale,
  pending
)
}
\arguments{
\item{afp.data}{\code{tibble} AFP linelist, Either from \code{raw.data$afp} of \code{\link[=get_all_polio_data]{get_all_polio_data()}}
or \code{ctry.data$afp.all.2} of \code{\link[=extract_country_data]{extract_country_data()}}.}

\item{year.pop.data}{\code{tibble} Summary table containing year and pop data. This is created
inside \code{\link[=f.npafp.rate.01]{f.npafp.rate.01()}}.}

\item{start_date}{\code{str} Start date to calculate the rolling interval for.}

\item{end_date}{\code{str} End date to calculate the rolling interval for.}

\item{spatial_scale}{Spatial scale. Valid arguments are: \verb{"ctry", "prov", "dist}.}

\item{pending}{\code{logical} Should cases classified as \code{PENDING} or \verb{LAB PENDING} be included in calculations? Default \code{TRUE}.}
}
\value{
A summary table including NPAFP rates and population data.
}
\description{
Calculates the NPAFP rate on a rolling basis, based on the start and end dates
specified.
}
\keyword{internal}
