% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.npafp.rate.01.R
\name{npafp_year}
\alias{npafp_year}
\title{Calculate NPAFP by year and geographic level}
\usage{
npafp_year(afp.data, pop.data, year.data, spatial_scale, pending)
}
\arguments{
\item{afp.data}{\code{tibble} AFP linelist. Either from \code{raw.data$afp} of \code{\link[=get_all_polio_data]{get_all_polio_data()}}
or \code{ctry.data$afp.all.2} of \code{\link[=extract_country_data]{extract_country_data()}}.}

\item{pop.data}{\code{tibble} Population data. Either from \verb{raw.data$\{ctry/prov/dist\}.pop} of \code{\link[=get_all_polio_data]{get_all_polio_data()}}
or \verb{ctry.data$\{ctry/prov/dist\}.pop} of \code{\link[=extract_country_data]{extract_country_data()}}.}

\item{year.data}{\code{tibble} Summary table. Created from \code{\link[=generate_year_data]{generate_year_data()}}.}

\item{spatial_scale}{Spatial scale. Valid arguments are: \verb{"ctry", "prov", "dist}.}

\item{pending}{\code{logical} Should cases classified as \code{PENDING} or \verb{LAB PENDING} be included in calculations? Default \code{TRUE}.}
}
\value{
\code{tibble} A summary table including NPAFP rates and population data.
}
\description{
Helper function for calculating the NPAFP rate based on the geographic level.
This function is used inside \code{\link[=f.npafp.rate.01]{f.npafp.rate.01()}}.
}
\keyword{internal}
