% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.afp.functions.R
\name{prep_npafp_table}
\alias{prep_npafp_table}
\title{NPAFP indicator tables with additional columns}
\usage{
prep_npafp_table(npafp.output, afp.data, start_date, end_date, spatial.scale)
}
\arguments{
\item{npafp.output}{\code{tibble} Output of running \code{\link[=f.npafp.rate.01]{f.npafp.rate.01()}}.}

\item{afp.data}{\code{tibble} AFP linelist. Either \code{raw.data$afp} or \code{ctry.data$afp.all.2}.}

\item{start_date}{\code{str} Start date of the desk review.}

\item{end_date}{\code{str} End date of the desk review.}

\item{spatial.scale}{\code{str} Spatial scale to analyze. Valid values are \verb{"ctry", "prov", "dist"}.}
}
\value{
\code{tibble} NPAFP rate table with additional columns related to case counts.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

The function adds additional information to the NPAFP table. In particular,
the number of AFP cases based on the geographic grouping selected. It also
adds a column for the number of WPV, VDPV1-3 cases.
}
\details{
This function has been deprecated as the columns are now added
from \code{\link[=f.npafp.rate.01]{f.npafp.rate.01()}} directly.
}
\examples{
raw.data <- get_all_polio_data(attach.spatial.data = FALSE)
ctry.ind <- f.npafp.rate.01(raw.data$afp, raw.data$ctry.pop,
  "2021-01-01", "2023-12-31", "ctry",
  sp_continuity_validation = FALSE
)
ctry.ind <- prep_npafp_table(
  ctry.ind, raw.data$afp,
  "2021-01-01", "2023-12-31", "ctry"
)
}
\keyword{internal}
