% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{pull_map_data}
\alias{pull_map_data}
\title{Pull the main map data to build the map}
\usage{
pull_map_data(
  raw.data,
  .vdpv,
  country,
  surv,
  virus_type,
  .start_date,
  .end_date
)
}
\arguments{
\item{raw.data}{\code{list} global polio data output from \code{\link[=get_all_polio_data]{get_all_polio_data()}}}

\item{.vdpv}{\code{logical} Whether to include VDPV in maps.}

\item{country}{\code{str} or \code{list} Name of the country or countries.}

\item{surv}{\code{str} or \code{list} Surveillance type.}

\item{virus_type}{\code{str} or \code{list} Virus type to map.}

\item{.start_date}{\code{date} Start date.}

\item{.end_date}{\code{date} End date.}
}
\value{
\code{tibble} Filtered dataset used to build the map.
}
\description{
This function primarily pulls the relevant information to build the map. Inherits
areguments from \code{\link[=generate_adhoc_map]{generate_adhoc_map()}}.
}
\keyword{internal}
