% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.R
\name{send_outlook_email}
\alias{send_outlook_email}
\title{Send email through Outlook}
\usage{
send_outlook_email(title, body, recipient, attachment = NULL)
}
\arguments{
\item{title}{str: Subject of message to be sent}

\item{body}{str: long string of body of message to be sent}

\item{recipient}{str: semicolon separated list of recipients}

\item{attachment}{str: path to local document to be attached to email, defaults to NULL}
}
\value{
Success or error message
}
\description{
Helper function to send an email through Outlook from R
}
