% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{set_emergence_colors}
\alias{set_emergence_colors}
\title{Set the emergence colors}
\usage{
set_emergence_colors(raw.data, country)
}
\arguments{
\item{raw.data}{Global polio data output of sirfunctions::get_all_polio_data()}

\item{country}{A string or a list of strings containing the countries of interest}
}
\value{
a named list containing the mapping of emergence and corresponding colors
}
\description{
Set the emergence colors
}
