% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.pop.functions.R
\name{set_shapefiles}
\alias{set_shapefiles}
\title{Get the most recent shapefile at a given spatial scale}
\usage{
set_shapefiles(ctry.data, spatial.scale)
}
\arguments{
\item{ctry.data}{RDS file containing country data}

\item{spatial.scale}{either "ctry", "prov", "dist"}
}
\value{
sf object of the most recent shapefile
}
\description{
Get the most recent shapefile at a given spatial scale
}
