% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_adhoc_map.R
\name{set_zoom_level}
\alias{set_zoom_level}
\title{Adjust the zoom level of the map}
\usage{
set_zoom_level(g1, map_ref, country, m_base_ctry)
}
\arguments{
\item{g1}{original ggplot map}

\item{map_ref}{reference map}

\item{country}{country or countries to zoom into}

\item{m_base_ctry}{shapefile containing base map of country}
}
\value{
adjusted ggplot map with zoom
}
\description{
Adjust the zoom level of the map
}
