% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sirfunctions-package.R
\docType{package}
\name{sirfunctions-package}
\alias{sirfunctions}
\alias{sirfunctions-package}
\title{sirfunctions: Key Functions to Analyze Global Polio Surveillance Data}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

The sirfunctions package contains key functions used by the Surveillance, Innovation, and Research (SIR) team within the Polio Eradication Branch (PEB) at the Centers for Disease Control and Prevention (CDC). It includes functions to download cleaned global polio data from the World Health Organization (WHO) Polio Information System (POLIS) and geographic files. In addition, the package contains functions to visualize important epidemiological trends and perform key performance indicators such as NPAFP rates, EV rates, and stool adequacy. Apart from getting data and calculating key indicators, there are several functions that also facilitate communication, visualize trends, and perform data quality checks.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/nish-kishore/sirfunctions}
}

}
\author{
\strong{Maintainer}: Nishant Kishore \email{ynm2@cdc.gov} (\href{https://orcid.org/0000-0003-0408-2747}{ORCID})

Authors:
\itemize{
  \item Mervin Keith Cuadera \email{xrg9@cdc.gov} (\href{https://orcid.org/0000-0003-4898-2659}{ORCID})
  \item Nicholas Heaghney \email{uic3@cdc.gov}
  \item Elizabeth Krow-Lucal \email{yxn9@cdc.gov}
  \item Smita Chavan \email{wsy2@cdc.gov}
}

}
\keyword{internal}
