% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{split_concat_raw_data}
\alias{split_concat_raw_data}
\title{Split or concatenate raw.data by year}
\usage{
split_concat_raw_data(
  action,
  split.years = NULL,
  raw.data.all = NULL,
  raw.data.post.2019 = NULL,
  raw.data.2016.2019 = NULL,
  raw.data.2001.2016 = NULL
)
}
\arguments{
\item{action}{Can either be to \code{concat} or \code{split}}

\item{split.years}{years to split to}

\item{raw.data.all}{raw data for all years}

\item{raw.data.post.2019}{raw data from 2019 to present}

\item{raw.data.2016.2019}{raw data from 2016 to 2019}

\item{raw.data.2001.2016}{raw data from 2001 to 2016}
}
\value{
list of lists or a single concatenated list
}
\description{
Split or concatenate raw.data by year
}
