% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{split_concat_raw_data}
\alias{split_concat_raw_data}
\title{Split or concatenate raw.data by year}
\usage{
split_concat_raw_data(
  action,
  split.years = NULL,
  raw.data.all = NULL,
  raw.data.post.2019 = NULL,
  raw.data.2016.2019 = NULL,
  raw.data.2001.2016 = NULL
)
}
\arguments{
\item{action}{\code{str}: can either be to \code{concat} or \code{split}}

\item{split.years}{\code{list}: a list containing years to split the data to}

\item{raw.data.all}{\code{Rds}: raw data to be split}

\item{raw.data.post.2019}{\code{Rds}: 2019-present raw data}

\item{raw.data.2016.2019}{\code{Rds}: 2016-2019 raw data}

\item{raw.data.2001.2016}{\code{Rds}: 2001-2016 raw data}
}
\value{
list of lists or a single concatenated list
}
\description{
Split or concatenate raw.data by year
}
