% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{split_concat_raw_data}
\alias{split_concat_raw_data}
\title{Split or concatenate raw.data by year}
\usage{
split_concat_raw_data(
  action,
  split.years = NULL,
  raw.data.all = NULL,
  raw.data.post.2019 = NULL,
  raw.data.2016.2019 = NULL,
  raw.data.2001.2016 = NULL
)
}
\arguments{
\item{action}{\code{str} Can either be to \code{"concat"} or \code{"split"}.}

\item{split.years}{\code{array} A numeric array of years by which data should be split.}

\item{raw.data.all}{\code{list} A list of data objects to be split.}

\item{raw.data.post.2019}{\code{list} A list of data objects to be concatenated.}

\item{raw.data.2016.2019}{\code{list} A list of data objects to be concatenated.}

\item{raw.data.2001.2016}{\code{list} A list of data objects to be concatenated.}
}
\value{
\code{list} A list of lists or a single concatenated list.
}
\description{
Split or concatenate raw.data by year
}
