% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{test_EDAV_connection}
\alias{test_EDAV_connection}
\title{Test network connection to the EDAV}
\usage{
test_EDAV_connection(
  azcontainer = suppressMessages(get_azure_storage_connection()),
  folder = "GID/PEB/SIR/Data",
  return_list = F,
  test_size = 1e+07
)
}
\arguments{
\item{azcontainer}{Azure storage container}

\item{folder}{str: Location of download folder in EDAV}

\item{return_list}{boolean: return a list of download time estimates, default F}

\item{test_size}{int: byte size of a theoretical folder}
}
\value{
System message with download and update time, potentially list output
}
\description{
Tests upload and download from EDAV
}
