% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{test_EDAV_connection}
\alias{test_EDAV_connection}
\title{Test network connection to the EDAV}
\usage{
test_EDAV_connection(
  azcontainer = suppressMessages(get_azure_storage_connection()),
  folder = "GID/PEB/SIR/Data",
  return_list = F,
  test_size = 1e+07
)
}
\arguments{
\item{azcontainer}{Azure storage container provided by \code{\link[=get_azure_storage_connection]{get_azure_storage_connection()}}.}

\item{folder}{\code{str} Location of folder in the EDAV environment that you want to download
and upload data from.}

\item{return_list}{\code{bool} return a list of download time estimates. Defaults to \code{FALSE}.}

\item{test_size}{\code{int} byte size of a theoretical file to be uploaded or downloaded.}
}
\value{
System message with download and update time, potentially a list.
}
\description{
Tests upload and download from EDAV by creating a temporary file
of a given size and testing the time it takes to upload and download the file.
}
\examples{
test_EDAV_connection()

}
