% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{update_data}
\alias{update_data}
\title{Pull new data and write to the specified file path}
\usage{
update_data(
  data_size,
  country_name,
  dr_data_path,
  attach_spatial_data,
  .data_folder,
  .polis_folder,
  .use_edav
)
}
\arguments{
\item{data_size}{\code{str} size of polio data to pull.
"small" (>= 2019), "medium" (>= 2016), large" (full)}

\item{country_name}{\code{str} name of the country to pull data from}

\item{dr_data_path}{\code{str} path to save the data set to.
Expected path is ./country/year/data}

\item{attach_spatial_data}{whether to attach spatial data}

\item{.data_folder}{\code{str} Path to the data folder.}

\item{.polis_folder}{\code{str} Path to the POLIS folder.}

\item{.use_edav}{\code{logical} Whether to use EDAV or not.}
}
\value{
\code{list} large list containing polio data
}
\description{
Pull new data and write to the specified file path
}
\keyword{internal}
