% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dal.R
\name{update_polio_data}
\alias{update_polio_data}
\title{Update a local dataset with new data
Update a local global polio data (raw.data) with new data}
\usage{
update_polio_data(local_dataset, overwrite = T)
}
\arguments{
\item{local_dataset}{file path to the RDS file}

\item{overwrite}{should the file be overwritten? Default TRUE.}
}
\description{
Update a local dataset with new data
Update a local global polio data (raw.data) with new data
}
