% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_spatial_data.R
\name{update_road_spatial_data}
\alias{update_road_spatial_data}
\title{Update Roads Spatial data}
\usage{
update_road_spatial_data(
  edav,
  output_path = "GID/PEB/SIR/Data/spatial/roads.new.rds",
  resolution = 10
)
}
\arguments{
\item{edav}{\code{logical} \code{TRUE} or \code{FALSE} depending on if final save location is in Azure.}

\item{output_path}{\code{str} Absolute file path location to save roads spatial data. Outputs in RDS by default,
but also supports \code{.qs2} format.}

\item{resolution}{\code{int} options are 110 (low resolution), 50 (medium resolution)
and 10 (high resolution); default is 10}
}
\value{
\code{NULL}, invisibly.
}
\description{
Function to download latest roads data for mapping purposes
}
\examples{
\dontrun{
update_road_spatial_data(edav = TRUE)
}

}
