% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_spatial_data.R
\name{update_road_spatial_data}
\alias{update_road_spatial_data}
\title{Update Roads Spatial data}
\usage{
update_road_spatial_data(edav, road_loc = "Data/spatial/roads.new.rds")
}
\arguments{
\item{edav}{\code{logical} \code{TRUE} or \code{FALSE} depending on if final save location is in Azure.}

\item{road_loc}{\code{str} Location to save roads spatial rds.}
}
\value{
\code{NULL}, invisibly.
}
\description{
Function to download latest roads data for mapping purposes
}
