% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_vacc_cov_data.R
\name{update_vacc_cov_data}
\alias{update_vacc_cov_data}
\title{Extract vaccine coverage information from IHME estimates}
\usage{
update_vacc_cov_data(tif_folder, edav, output_folder, output_format = ".rds")
}
\arguments{
\item{tif_folder}{\code{str} absolute path to the folder containing all tifs of interest}

\item{edav}{\code{logical} \code{TRUE} or \code{FALSE} depending on if final save location is in Azure.}

\item{output_folder}{\code{str} Absolute folder path location to save ctry, prov and dist coverage
data. Outputs in RDS by default, but also supports \code{.qs2} format.}

\item{output_format}{\code{str} '.rds' or '.qs2'}
}
\value{
\code{NULL}, invisibly.
}
\description{
#' @description
Function to extract IHME vaccine coverage data to existing
spatial files. Key spatial data can be found at \verb{https://www.healthdata.org/research-analysis/health-topics/vaccine-coverage-data}. Please keep in mind that these sources may chance and reach out to IHME directly for further information. Enter your email to gain access to the data folder. Inside the data folder go to \verb{Geospatial Vaccine Coverage} > \verb{04_Rasters}. Download the following geotifs and place them into a single folder in your local machine:
\itemize{
\item \code{bcg1_cov_mean_raked_2000_2023.tif}
\item \code{dpt1_cov_mean_raked_2000_2023.tif}
\item \code{dpt3_cov_mean_raked_2000_2023.tif}
\item \code{mcv1_cov_mean_raked_2000_2023.tif}
\item \code{polio3_cov_mean_raked_2000_2023.tif}
}
}
\examples{
\dontrun{
update_vacc_cov_data()
}

}
