% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dr.main.functions.R
\name{upload_dr_to_github}
\alias{upload_dr_to_github}
\title{Upload desk review script to the sg-desk-review Github repository}
\usage{
upload_dr_to_github(file_path, repo_path, message = "updating file")
}
\arguments{
\item{file_path}{location of the file to upload to the sg-desk-review repo}

\item{repo_path}{local path of the sg-desk-review repo}

\item{message}{message to include in the commit}
}
\value{
Status message
}
\description{
Upload desk review script to the sg-desk-review Github repository
}
