% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.R
\name{upload_to_sharepoint}
\alias{upload_to_sharepoint}
\title{Upload file to Sharepoint}
\usage{
upload_to_sharepoint(
  file_to_upload,
  sharepoint_file_loc,
  site = "https://cdc.sharepoint.com/teams/CGH-GID-PEB-SIR283",
  drive = "Documents"
)
}
\arguments{
\item{file_to_upload}{str: local path of files to be uploaded}

\item{sharepoint_file_loc}{str: location in SharePoint to upload file}

\item{site}{str: SharePoint site location, defaults to "CGH-GID-PEB" or the site URL: "https://cdc.sharepoint.com/teams/CGH-GID-PEB-SIR283"}

\item{drive}{str: SharePoint drive to upload data to}
}
\value{
Success or error message
}
\description{
Helper function to upload file to MS SharePoint
}
