% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/com.R
\name{upload_to_sharepoint}
\alias{upload_to_sharepoint}
\title{Upload file to Sharepoint}
\usage{
upload_to_sharepoint(
  file_to_upload,
  sharepoint_file_loc,
  site = "https://cdc.sharepoint.com/teams/CGH-GID-PEB-SIR283",
  drive = "Documents"
)
}
\arguments{
\item{file_to_upload}{\code{str} Local path of files to be uploaded.}

\item{sharepoint_file_loc}{\code{str} Location in SharePoint to upload file.}

\item{site}{\code{str} SharePoint site location. Defaults to \code{"CGH-GID-PEB"}
or the site URL: \code{"https://cdc.sharepoint.com/teams/CGH-GID-PEB-SIR283"}.}

\item{drive}{\code{str} SharePoint drive to upload data to.}
}
\value{
Status message whether the operation was a success or an error message.
}
\description{
Helper function to upload file to MS SharePoint
}
\examples{
\dontrun{
file_path <- "C:/Users/ABC1/df1.csv"
sp_path <- "test_folder"
upload_to_sharepoint(file_path, sp_path)
}

}
